% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_python_packages}
\alias{assert_python_packages}
\title{Assert Python Packages}
\usage{
assert_python_packages(packages)
}
\arguments{
\item{packages}{(\code{character()})\cr
Python packages to check.}
}
\value{
(\code{character()})\cr
}
\description{
Assert that the given Python packages are available.
}
