% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_install_catboost.R
\name{install_catboost}
\alias{install_catboost}
\title{Install catboost}
\usage{
install_catboost(
  version = NULL,
  os = NULL,
  install_required = TRUE,
  INSTALL_opts = c("--no-multiarch", "--no-test-load"),
  ...
)
}
\arguments{
\item{version}{\code{(character(1))}
Version to install, if \code{NULL} installs latest}

\item{os}{\code{(character(1))} \cr
Operating system to install on, if \code{NULL} automatically detected}

\item{install_required}{\code{(logical(1))} \cr
If \code{TRUE} (default) and version is \code{NULL} then installs {curl} and
{jsonlite} if not already installed.}

\item{INSTALL_opts}{\code{(character())} \cr
Passed to \link[devtools:remote-reexports]{devtools::install_url}}

\item{...}{\code{ANY} \cr
Other arguments passed to \link[devtools:remote-reexports]{devtools::install_url}}
}
\description{
Helper function to install catboost
}
