% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_proba.R
\name{install_keras}
\alias{install_keras}
\title{Install Keras and Tensorflow}
\usage{
install_keras(
  method = "auto",
  conda = "auto",
  pip = FALSE,
  install_tensorflow = FALSE
)
}
\arguments{
\item{method, conda, pip}{See \link[reticulate:py_install]{reticulate::py_install}.}

\item{install_tensorflow}{If \code{TRUE} installs the dependency \code{tensorflow} package as well.}
}
\description{
Stripped back version of \link[keras:install_keras]{keras::install_keras}.
}
