% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_install_learners.R
\name{install_learners}
\alias{install_learners}
\title{Install Learner Dependencies}
\usage{
install_learners(.keys, repos = "https://cloud.r-project.org", ...)
}
\arguments{
\item{.keys}{(\code{character()}) \cr Keys passed to \link[mlr3:mlr_learners]{mlr_learners} specifying
learners to install.}

\item{repos}{(\code{character(1)}) \cr Passed to \link[utils:install.packages]{utils::install.packages}.}

\item{...}{(\code{ANY}) \cr Additional options to pass to \code{utils::install.packages()},
\code{BiocManager::install()} or \code{remotes::install_github()}.}
}
\description{
Install required dependencies for specified learners. Works for packages on
CRAN, Bioconductor and GitHub, otherwise must be manually installed.
}
