% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_pycox.R
\name{install_pycox}
\alias{install_pycox}
\title{Install Pycox With Reticulate}
\usage{
install_pycox(
  method = "auto",
  conda = "auto",
  pip = FALSE,
  install_torch = FALSE
)
}
\arguments{
\item{method, conda, pip}{See \link[reticulate:py_install]{reticulate::py_install}.}

\item{install_torch}{If \code{TRUE} installs the dependency \code{torch} package as well.}
}
\description{
Installs the python 'pycox' package via reticulate.
}
