% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_rfsrc.R
\name{install_rfsrc}
\alias{install_rfsrc}
\title{Install randomForestSRC from GitHub}
\usage{
install_rfsrc()
}
\description{
randomForestSRC is (probably temporarily) removed from CRAN. This function installs the
latest version from GitHub.
}
