% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_pycox.R
\name{install_torch}
\alias{install_torch}
\title{Install Torch With Reticulate}
\usage{
install_torch(method = "auto", conda = "auto", pip = FALSE)
}
\arguments{
\item{method, conda, pip}{See \link[reticulate:py_install]{reticulate::py_install}}
}
\description{
Installs the python 'torch' package via reticulate.
}
