% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{learner_is_runnable}
\alias{learner_is_runnable}
\title{Check if an example is runnable}
\usage{
learner_is_runnable(id)
}
\arguments{
\item{id}{(\code{character(1)})\cr
The id of the learner.}
}
\description{
Checks if an example is runnable by loading the required packages.
}
\keyword{internal}
