% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{lrn}
\alias{lrn}
\alias{lrns}
\title{Syntactic Sugar for Learner Construction}
\usage{
lrn(.key, ...)

lrns(.keys, ...)
}
\arguments{
\item{.key}{\code{(character(1))} \cr Key passed to \link[mlr3:mlr_learners]{mlr_learners} to retrieve
the learner.}

\item{...}{\code{ANY} \cr Passed to \link[mlr3:mlr_sugar]{mlr3::lrns}}

\item{.keys}{\code{(character())} \cr Keys passed to \link[mlr3:mlr_learners]{mlr_learners} to retrieve
the learners.}
}
\description{
Overloads \link[mlr3:mlr_sugar]{mlr3::lrn} to automatically detect if required packages are installed.
}
