% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner_h2o_classif_deeplearning.R
\name{mlr_learners_classif.h2o.deeplearning}
\alias{mlr_learners_classif.h2o.deeplearning}
\alias{LearnerClassifH2ODeeplearning}
\title{Classification H2O Deep Learning Learner}
\description{
Classification feed-forward multilayer artificial neural network learner.
Class \code{\link[h2o:h2o.deeplearning]{h2o::h2o.deeplearning()}} from package \CRANpkg{h2o}.
}
\section{Dictionary}{

This \link[mlr3:Learner]{Learner} can be instantiated via \link[mlr3:mlr_sugar]{lrn()}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{lrn("classif.h2o.deeplearning")
}\if{html}{\out{</div>}}
}

\section{Meta Information}{

\itemize{
\item Task type: \dQuote{classif}
\item Predict Types: \dQuote{response}, \dQuote{prob}
\item Feature Types: \dQuote{integer}, \dQuote{numeric}, \dQuote{factor}
\item Required Packages: \CRANpkg{mlr3}, \CRANpkg{h2o}
}
}

\section{Parameters}{
\tabular{lllll}{
   Id \tab Type \tab Default \tab Levels \tab Range \cr
   ignore_const_cols \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
   score_each_iteration \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   balance_classes \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   class_sampling_factors \tab numeric \tab NULL \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   max_after_balance_size \tab numeric \tab 5 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   use_all_factor_levels \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
   standardize \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
   activation \tab character \tab Rectifier \tab Rectifier, Tanh, TanhWithDropout, RectifierWithDropout, Maxout, MaxoutWithDropout \tab - \cr
   hidden \tab untyped \tab c(200L, 200L) \tab  \tab - \cr
   epochs \tab numeric \tab 10 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   train_samples_per_iteration \tab integer \tab -2 \tab  \tab \eqn{[-2, \infty)}{[-2, Inf)} \cr
   target_ratio_comm_to_comp \tab numeric \tab 0.05 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   seed \tab integer \tab -1 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   adaptive_rate \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
   rho \tab numeric \tab 0.99 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   epsilon \tab numeric \tab 1e-08 \tab  \tab \eqn{[1e-10, 1e-04]}{[1e-10, 1e-04]} \cr
   rate \tab numeric \tab 0.005 \tab  \tab \eqn{[0, 1]}{[0, 1]} \cr
   rate_annealing \tab numeric \tab 1e-06 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   rate_decay \tab numeric \tab 1 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   momentum_start \tab numeric \tab 0 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   momentum_ramp \tab numeric \tab 1e+06 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   momentum_stable \tab numeric \tab 0 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   nesterov_accelerated_gradient \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
   input_dropout_ratio \tab numeric \tab 0 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   hidden_dropout_ratios \tab numeric \tab 0.5 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   l1 \tab numeric \tab 0 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   l2 \tab numeric \tab 0 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   max_w2 \tab numeric \tab Inf \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   initial_weight_distribution \tab character \tab UniformAdaptive \tab UniformAdaptive, Uniform, Normal \tab - \cr
   initial_weight_scale \tab numeric \tab 1 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   loss \tab character \tab Automatic \tab Automatic, CrossEntropy, Quadratic, Absolute, Huber \tab - \cr
   distribution \tab character \tab AUTO \tab AUTO, bernoulli, multinomial \tab - \cr
   score_interval \tab numeric \tab 5 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   score_training_samples \tab integer \tab 10000 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   score_validation_samples \tab integer \tab 0 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   score_duty_cycle \tab numeric \tab 0.1 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   classification_stop \tab numeric \tab 0 \tab  \tab \eqn{[-1, \infty)}{[-1, Inf)} \cr
   stopping_rounds \tab integer \tab 5 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   stopping_metric \tab character \tab AUTO \tab AUTO, logloss, AUC, lift_top_group, misclassification, AUCPR, mean_per_class_error, custom, custom_increasing \tab - \cr
   stopping_tolerance \tab numeric \tab 0 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   max_runtime_secs \tab numeric \tab 0 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   score_validation_sampling \tab character \tab Uniform \tab Uniform, Stratified \tab - \cr
   diagnostics \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
   fast_mode \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
   force_load_balance \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
   variable_importances \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
   replicate_training_data \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
   single_node_mode \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   shuffle_training_data \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   missing_values_handling \tab character \tab MeanImputation \tab MeanImputation, Skip \tab - \cr
   quiet_mode \tab logical \tab - \tab TRUE, FALSE \tab - \cr
   autoencoder \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   sparse \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   col_major \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   average_activation \tab numeric \tab 0 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   sparsity_beta \tab numeric \tab 0 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   reproducible \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   export_weights_and_biases \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   mini_batch_size \tab integer \tab 1 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   categorical_encoding \tab character \tab AUTO \tab AUTO, Enum, OneHotInternal, OneHotExplicit, Binary, Eigen, LabelEncoder, SortByResponse, EnumLimited \tab - \cr
   elastic_averaging \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   elastic_averaging_moving_rate \tab numeric \tab 0.9 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   elastic_averaging_regularization \tab numeric \tab 0.001 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   verbose \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
}
}

\examples{
\dontshow{if (learner_is_runnable("classif.h2o.deeplearning")) withAutoprint(\{ # examplesIf}
# Define the Learner
learner = lrn("classif.h2o.deeplearning")
print(learner)

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
print(learner$importance())

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
\dontshow{\}) # examplesIf}
}
\references{
Fryda T, LeDell E, Gill N, Aiello S, Fu A, Candel A, Click C, Kraljevic T, Nykodym T, Aboyoun P, Kurka M, Malohlava M, Poirier S, Wong W (2025).
\emph{h2o: R Interface for the 'H2O' Scalable Machine Learning Platform}.
R package version 3.46.0.9, \url{https://github.com/h2oai/h2o-3}.
}
\seealso{
\itemize{
\item \link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}.
\item \code{as.data.table(mlr_learners)} for a table of available \link[mlr3:Learner]{Learners} in the running session (depending on the loaded packages).
\item Chapter in the \href{https://mlr3book.mlr-org.com/}{mlr3book}: \url{https://mlr3book.mlr-org.com/basics.html#learners}
\item \CRANpkg{mlr3learners} for a selection of recommended learners.
\item \CRANpkg{mlr3cluster} for unsupervised clustering learners.
\item \CRANpkg{mlr3pipelines} to combine learners with pre- and postprocessing steps.
\item \CRANpkg{mlr3tuning} for tuning of hyperparameters, \CRANpkg{mlr3tuningspaces} for established default tuning spaces.
}
}
\author{
awinterstetter
}
\section{Super classes}{
\code{\link[mlr3:Learner]{mlr3::Learner}} -> \code{\link[mlr3:LearnerClassif]{mlr3::LearnerClassif}} -> \code{LearnerClassifH2ODeeplearning}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LearnerClassifH2ODeeplearning-new}{\code{LearnerClassifH2ODeeplearning$new()}}
\item \href{#method-LearnerClassifH2ODeeplearning-importance}{\code{LearnerClassifH2ODeeplearning$importance()}}
\item \href{#method-LearnerClassifH2ODeeplearning-clone}{\code{LearnerClassifH2ODeeplearning$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="base_learner"><a href='../../mlr3/html/Learner.html#method-Learner-base_learner'><code>mlr3::Learner$base_learner()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="configure"><a href='../../mlr3/html/Learner.html#method-Learner-configure'><code>mlr3::Learner$configure()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="encapsulate"><a href='../../mlr3/html/Learner.html#method-Learner-encapsulate'><code>mlr3::Learner$encapsulate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="format"><a href='../../mlr3/html/Learner.html#method-Learner-format'><code>mlr3::Learner$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="help"><a href='../../mlr3/html/Learner.html#method-Learner-help'><code>mlr3::Learner$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict"><a href='../../mlr3/html/Learner.html#method-Learner-predict'><code>mlr3::Learner$predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict_newdata"><a href='../../mlr3/html/Learner.html#method-Learner-predict_newdata'><code>mlr3::Learner$predict_newdata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="print"><a href='../../mlr3/html/Learner.html#method-Learner-print'><code>mlr3::Learner$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="reset"><a href='../../mlr3/html/Learner.html#method-Learner-reset'><code>mlr3::Learner$reset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="selected_features"><a href='../../mlr3/html/Learner.html#method-Learner-selected_features'><code>mlr3::Learner$selected_features()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="train"><a href='../../mlr3/html/Learner.html#method-Learner-train'><code>mlr3::Learner$train()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="LearnerClassif" data-id="predict_newdata_fast"><a href='../../mlr3/html/LearnerClassif.html#method-LearnerClassif-predict_newdata_fast'><code>mlr3::LearnerClassif$predict_newdata_fast()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerClassifH2ODeeplearning-new"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerClassifH2ODeeplearning-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerClassifH2ODeeplearning$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerClassifH2ODeeplearning-importance"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerClassifH2ODeeplearning-importance}{}}}
\subsection{Method \code{importance()}}{
Variable importance scores computed by H2O.
Requires \code{variable_importances = TRUE}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerClassifH2ODeeplearning$importance()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Named \code{numeric()}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerClassifH2ODeeplearning-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerClassifH2ODeeplearning-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerClassifH2ODeeplearning$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
