% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner_h2o_classif_glm.R
\name{mlr_learners_classif.h2o.glm}
\alias{mlr_learners_classif.h2o.glm}
\alias{LearnerClassifH2OGLM}
\title{Classification H2O GLM Learner}
\description{
Generalized linear classification model.
Calls \code{\link[h2o:h2o.glm]{h2o::h2o.glm()}} from package \CRANpkg{h2o} with \code{family}
always set to \code{"binomial"}.
}
\section{Initial parameter values}{

\itemize{
\item \code{family}
\itemize{
\item Fixed to \code{"binomial"} by this learner wrapper.
}
\item \code{missing_values_handling}
\itemize{
\item H2O default is \code{"MeanImputation"}.
}
}
}

\section{Dictionary}{

This \link[mlr3:Learner]{Learner} can be instantiated via \link[mlr3:mlr_sugar]{lrn()}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{lrn("classif.h2o.glm")
}\if{html}{\out{</div>}}
}

\section{Meta Information}{

\itemize{
\item Task type: \dQuote{classif}
\item Predict Types: \dQuote{response}, \dQuote{prob}
\item Feature Types: \dQuote{integer}, \dQuote{numeric}, \dQuote{factor}
\item Required Packages: \CRANpkg{mlr3}, \CRANpkg{h2o}
}
}

\section{Parameters}{
\tabular{lllll}{
   Id \tab Type \tab Default \tab Levels \tab Range \cr
   max_iterations \tab integer \tab 50 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   seed \tab integer \tab - \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   beta_epsilon \tab numeric \tab 0 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   objective_epsilon \tab numeric \tab - \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   gradient_epsilon \tab numeric \tab - \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   standardize \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
   solver \tab character \tab IRLSM \tab AUTO, IRLSM, L_BFGS, COORDINATE_DESCENT, COORDINATE_DESCENT_NAIVE \tab - \cr
   link \tab character \tab logit \tab family_default, logit, log \tab - \cr
   ignore_const_cols \tab logical \tab - \tab TRUE, FALSE \tab - \cr
   score_each_iteration \tab logical \tab - \tab TRUE, FALSE \tab - \cr
   score_iteration_interval \tab integer \tab - \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   alpha \tab numeric \tab 0.5 \tab  \tab \eqn{[0, 1]}{[0, 1]} \cr
   prior \tab numeric \tab - \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   lambda \tab numeric \tab 1e-05 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   lambda_search \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   nlambdas \tab integer \tab - \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   lambda_min_ratio \tab numeric \tab - \tab  \tab \eqn{[0, 1]}{[0, 1]} \cr
   early_stopping \tab logical \tab - \tab TRUE, FALSE \tab - \cr
   stopping_rounds \tab integer \tab - \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   stopping_metric \tab untyped \tab - \tab  \tab - \cr
   stopping_tolerance \tab numeric \tab - \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   missing_values_handling \tab character \tab - \tab MeanImputation, Skip, PlugValues \tab - \cr
   plug_values \tab untyped \tab - \tab  \tab - \cr
   compute_p_values \tab logical \tab - \tab TRUE, FALSE \tab - \cr
   remove_collinear_columns \tab logical \tab - \tab TRUE, FALSE \tab - \cr
   beta_constraints \tab untyped \tab - \tab  \tab - \cr
   non_negative \tab logical \tab - \tab TRUE, FALSE \tab - \cr
   startval \tab untyped \tab - \tab  \tab - \cr
   calc_like \tab logical \tab - \tab TRUE, FALSE \tab - \cr
   cold_start \tab logical \tab - \tab TRUE, FALSE \tab - \cr
   max_active_predictors \tab integer \tab - \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   interactions \tab untyped \tab - \tab  \tab - \cr
   interaction_pairs \tab untyped \tab - \tab  \tab - \cr
   obj_reg \tab numeric \tab - \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   balance_classes \tab logical \tab - \tab TRUE, FALSE \tab - \cr
   class_sampling_factors \tab untyped \tab - \tab  \tab - \cr
   max_after_balance_size \tab numeric \tab - \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   max_runtime_secs \tab numeric \tab - \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   generate_scoring_history \tab logical \tab - \tab TRUE, FALSE \tab - \cr
   build_null_model \tab logical \tab - \tab TRUE, FALSE \tab - \cr
   generate_variable_inflation_factors \tab logical \tab - \tab TRUE, FALSE \tab - \cr
   gainslift_bins \tab integer \tab - \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   init_optimal_glm \tab logical \tab - \tab TRUE, FALSE \tab - \cr
   intercept \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
}
}

\examples{
\dontshow{if (learner_is_runnable("classif.h2o.glm")) withAutoprint(\{ # examplesIf}
# Define the Learner
learner = lrn("classif.h2o.glm")
print(learner)

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)


# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
\dontshow{\}) # examplesIf}
}
\references{
Fryda T, LeDell E, Gill N, Aiello S, Fu A, Candel A, Click C, Kraljevic T, Nykodym T, Aboyoun P, Kurka M, Malohlava M, Poirier S, Wong W (2025).
\emph{h2o: R Interface for the 'H2O' Scalable Machine Learning Platform}.
R package version 3.46.0.9, \url{https://github.com/h2oai/h2o-3}.
}
\seealso{
\itemize{
\item \link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}.
\item \code{as.data.table(mlr_learners)} for a table of available \link[mlr3:Learner]{Learners} in the running session (depending on the loaded packages).
\item Chapter in the \href{https://mlr3book.mlr-org.com/}{mlr3book}: \url{https://mlr3book.mlr-org.com/basics.html#learners}
\item \CRANpkg{mlr3learners} for a selection of recommended learners.
\item \CRANpkg{mlr3cluster} for unsupervised clustering learners.
\item \CRANpkg{mlr3pipelines} to combine learners with pre- and postprocessing steps.
\item \CRANpkg{mlr3tuning} for tuning of hyperparameters, \CRANpkg{mlr3tuningspaces} for established default tuning spaces.
}
}
\author{
awinterstetter
}
\section{Super classes}{
\code{\link[mlr3:Learner]{mlr3::Learner}} -> \code{\link[mlr3:LearnerClassif]{mlr3::LearnerClassif}} -> \code{LearnerClassifH2OGLM}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LearnerClassifH2OGLM-new}{\code{LearnerClassifH2OGLM$new()}}
\item \href{#method-LearnerClassifH2OGLM-clone}{\code{LearnerClassifH2OGLM$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="base_learner"><a href='../../mlr3/html/Learner.html#method-Learner-base_learner'><code>mlr3::Learner$base_learner()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="configure"><a href='../../mlr3/html/Learner.html#method-Learner-configure'><code>mlr3::Learner$configure()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="encapsulate"><a href='../../mlr3/html/Learner.html#method-Learner-encapsulate'><code>mlr3::Learner$encapsulate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="format"><a href='../../mlr3/html/Learner.html#method-Learner-format'><code>mlr3::Learner$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="help"><a href='../../mlr3/html/Learner.html#method-Learner-help'><code>mlr3::Learner$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict"><a href='../../mlr3/html/Learner.html#method-Learner-predict'><code>mlr3::Learner$predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict_newdata"><a href='../../mlr3/html/Learner.html#method-Learner-predict_newdata'><code>mlr3::Learner$predict_newdata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="print"><a href='../../mlr3/html/Learner.html#method-Learner-print'><code>mlr3::Learner$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="reset"><a href='../../mlr3/html/Learner.html#method-Learner-reset'><code>mlr3::Learner$reset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="selected_features"><a href='../../mlr3/html/Learner.html#method-Learner-selected_features'><code>mlr3::Learner$selected_features()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="train"><a href='../../mlr3/html/Learner.html#method-Learner-train'><code>mlr3::Learner$train()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="LearnerClassif" data-id="predict_newdata_fast"><a href='../../mlr3/html/LearnerClassif.html#method-LearnerClassif-predict_newdata_fast'><code>mlr3::LearnerClassif$predict_newdata_fast()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerClassifH2OGLM-new"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerClassifH2OGLM-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerClassifH2OGLM$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerClassifH2OGLM-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerClassifH2OGLM-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerClassifH2OGLM$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
