% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner_LiblineaR_classif_liblinearl2l1svc.R
\name{mlr_learners_classif.liblinearl2l1svc}
\alias{mlr_learners_classif.liblinearl2l1svc}
\alias{LearnerClassifLiblineaRL2L1SVC}
\title{L2-Regularized L1-Loss Support Vector Classification Learner}
\description{
Calls \link[LiblineaR:LiblineaR]{LiblineaR::LiblineaR} from package \CRANpkg{LiblineaR}.
}
\details{
Calls \code{\link[LiblineaR:LiblineaR]{LiblineaR::LiblineaR()}} with \code{type = 3}.
}
\section{Dictionary}{
 This \link{Learner} can be instantiated via the
\link[mlr3misc:Dictionary]{dictionary} \link{mlr_learners} or with the associated
sugar function \code{\link[=lrn]{lrn()}}:\preformatted{mlr_learners$get("classif.liblinearl2l1svc")
lrn("classif.liblinearl2l1svc")
}
}

\section{Traits}{

\itemize{
\item Packages: LiblineaR
\item Predict Types: response
\item Feature Types: numeric
\item Properties: multiclass, twoclass
}
}

\section{Custom mlr3 defaults}{

\itemize{
\item \code{epsilon}:
\itemize{
\item Actual default: 0.01
\item Adjusted default: 0.1
\item Reason for change: Param depends on param "type" which is handled
internally by choosing the mlr3 learner. The default is set to the actual
default of the respective "type".
}
}
}

\examples{
if (mlr3misc::require_namespaces(c("LiblineaR"))) {
  learner = mlr3::lrn("classif.liblinearl2l1svc")
  print(learner)

  # available parameters:
  learner$param_set$ids()
}
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}:
\link[mlr3:mlr_learners]{mlr3::mlr_learners}
}
\author{
be-marc
}
\section{Super classes}{
\code{\link[mlr3:Learner]{mlr3::Learner}} -> \code{\link[mlr3:LearnerClassif]{mlr3::LearnerClassif}} -> \code{LearnerClassifLiblineaRL2L1SVC}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LearnerClassifLiblineaRL2L1SVC$new()}}
\item \href{#method-clone}{\code{LearnerClassifLiblineaRL2L1SVC$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="format">}\href{../../mlr3/html/Learner.html#method-format}{\code{mlr3::Learner$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="help">}\href{../../mlr3/html/Learner.html#method-help}{\code{mlr3::Learner$help()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict">}\href{../../mlr3/html/Learner.html#method-predict}{\code{mlr3::Learner$predict()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict_newdata">}\href{../../mlr3/html/Learner.html#method-predict_newdata}{\code{mlr3::Learner$predict_newdata()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="print">}\href{../../mlr3/html/Learner.html#method-print}{\code{mlr3::Learner$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="reset">}\href{../../mlr3/html/Learner.html#method-reset}{\code{mlr3::Learner$reset()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="train">}\href{../../mlr3/html/Learner.html#method-train}{\code{mlr3::Learner$train()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerClassifLiblineaRL2L1SVC$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerClassifLiblineaRL2L1SVC$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
