% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner_ropls_classif_opls.R
\name{mlr_learners_classif.opls}
\alias{mlr_learners_classif.opls}
\alias{LearnerClassifOPLS}
\title{OPLS-DA Classification Learner}
\description{
Binary classification with orthogonal partial least squares discriminant analysis (OPLS-DA).
Calls \code{ropls::opls()} from Bioconductor package \CRANpkg{ropls}.

\code{ropls} is distributed via Bioconductor and not via CRAN, so it has to be installed with
\code{BiocManager::install("ropls")}.
}
\section{Initial parameter values}{

\itemize{
\item \code{predI}
\itemize{
\item Actual default: \code{NA}, which lets \code{ropls::opls()} determine the number of predictive components by
cross-validation.
\item Adjusted default: \code{1}
\item Reason for change: OPLS-DA is defined with a single predictive component, and a fixed number of components
keeps the model comparable across resampling iterations.
}
\item \code{orthoI}
\itemize{
\item Actual default: \code{0}
\item Adjusted default: \code{1}
\item Reason for change: With \code{orthoI = 0} no orthogonal component is extracted and \code{ropls::opls()} fits plain
PLS-DA instead of OPLS-DA.
}
\item \code{permI}
\itemize{
\item Actual default: \code{20}
\item Adjusted default: \code{0}
\item Reason for change: The permutation test is a model diagnostic that is not used for prediction and costs
\code{permI} extra model fits.
}
}
}

\section{Suppressed output}{

\code{ropls::opls()} writes a diagnostic figure and a model summary by default.
Both are suppressed by fixing \code{fig.pdfC} and \code{info.txtC} to \code{"none"}, so neither is exposed as a
hyperparameter.
}

\section{Number of components}{

\code{ropls::opls()} requires \code{predI + orthoI} to not exceed the smaller of the number of observations and the
number of features, and errors otherwise.
With the initial parameter values this means that the learner needs at least two features.
}

\section{Dictionary}{

This \link[mlr3:Learner]{Learner} can be instantiated via \link[mlr3:lrn]{lrn()}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{lrn("classif.opls")
}\if{html}{\out{</div>}}
}

\section{Meta Information}{

\itemize{
\item Task type: \dQuote{classif}
\item Predict Types: \dQuote{response}
\item Feature Types: \dQuote{integer}, \dQuote{numeric}
\item Required Packages: \CRANpkg{mlr3}, \CRANpkg{mlr3extralearners}, \CRANpkg{ropls}
}
}

\section{Parameters}{
\tabular{lllll}{
   Id \tab Type \tab Default \tab Levels \tab Range \cr
   predI \tab integer \tab - \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   orthoI \tab integer \tab 0 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   algoC \tab character \tab default \tab default, nipals, svd \tab - \cr
   crossvalI \tab integer \tab 7 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   log10L \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   scaleC \tab character \tab standard \tab none, center, pareto, standard \tab - \cr
   permI \tab integer \tab 20 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
}
}

\examples{
\dontshow{if (learner_is_runnable("classif.opls")) withAutoprint(\{ # examplesIf}
# Define the Learner
learner = lrn("classif.opls")
print(learner)

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)


# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
\dontshow{\}) # examplesIf}
}
\references{
Thévenot, A. E, Roux, Aurélie, Xu, Ying, Ezan, Eric, Junot, Christophe (2015).
\dQuote{Analysis of the Human Adult Urinary Metabolome Variations with Age, Body Mass Index, and Gender by Implementing a Comprehensive Workflow for Univariate and OPLS Statistical Analyses.}
\emph{Journal of Proteome Research}, \bold{14}(8), 3322--3335.
\doi{10.1021/acs.jproteome.5b00354}.
}
\seealso{
\itemize{
\item \link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}.
\item \code{as.data.table(mlr_learners)} for a table of available \link[mlr3:Learner]{Learners} in the running session (depending on the loaded packages).
\item Chapter in the \href{https://mlr3book.mlr-org.com/}{mlr3book}: \url{https://mlr3book.mlr-org.com/chapters/chapter2/data_and_basic_modeling.html#sec-learners}
\item \CRANpkg{mlr3learners} for a selection of recommended learners.
\item \CRANpkg{mlr3cluster} for unsupervised clustering learners.
\item \CRANpkg{mlr3pipelines} to combine learners with pre- and postprocessing steps.
\item \CRANpkg{mlr3tuning} for tuning of hyperparameters, \CRANpkg{mlr3tuningspaces} for established default tuning spaces.
}
}
\author{
Takha65
}
\section{Super classes}{
\code{\link[mlr3:Learner]{mlr3::Learner}} -> \code{\link[mlr3:LearnerClassif]{mlr3::LearnerClassif}} -> \code{LearnerClassifOPLS}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-LearnerClassifOPLS-initialize}{\code{LearnerClassifOPLS$new()}}
    \item \href{#method-LearnerClassifOPLS-clone}{\code{LearnerClassifOPLS$clone()}}
  }
}
\if{html}{\out{<details><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="base_learner"><a href='../../mlr3/html/Learner.html#method-Learner-base_learner'><code>mlr3::Learner$base_learner()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="configure"><a href='../../mlr3/html/Learner.html#method-Learner-configure'><code>mlr3::Learner$configure()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="encapsulate"><a href='../../mlr3/html/Learner.html#method-Learner-encapsulate'><code>mlr3::Learner$encapsulate()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="format"><a href='../../mlr3/html/Learner.html#method-Learner-format'><code>mlr3::Learner$format()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="help"><a href='../../mlr3/html/Learner.html#method-Learner-help'><code>mlr3::Learner$help()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict"><a href='../../mlr3/html/Learner.html#method-Learner-predict'><code>mlr3::Learner$predict()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict_newdata"><a href='../../mlr3/html/Learner.html#method-Learner-predict_newdata'><code>mlr3::Learner$predict_newdata()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="print"><a href='../../mlr3/html/Learner.html#method-Learner-print'><code>mlr3::Learner$print()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="reset"><a href='../../mlr3/html/Learner.html#method-Learner-reset'><code>mlr3::Learner$reset()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="selected_features"><a href='../../mlr3/html/Learner.html#method-Learner-selected_features'><code>mlr3::Learner$selected_features()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="train"><a href='../../mlr3/html/Learner.html#method-Learner-train'><code>mlr3::Learner$train()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="LearnerClassif" data-id="predict_newdata_fast"><a href='../../mlr3/html/LearnerClassif.html#method-LearnerClassif-predict_newdata_fast'><code>mlr3::LearnerClassif$predict_newdata_fast()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerClassifOPLS-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerClassifOPLS-initialize}{}}}
\subsection{\code{LearnerClassifOPLS$new()}}{
  Creates a new instance of this \link[R6:R6Class]{R6} class.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LearnerClassifOPLS$new()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerClassifOPLS-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerClassifOPLS-clone}{}}}
\subsection{\code{LearnerClassifOPLS$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LearnerClassifOPLS$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
