% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py_base_class_fastai.R
\name{mlr_learners_classif.python.fastai}
\alias{mlr_learners_classif.python.fastai}
\alias{LearnerPythonClassifFastai}
\title{Classification Neural Network Learner}
\description{
Simple and fast neural nets for tabular data classification.
Calls \code{\link[fastai:tabular_learner]{fastai::tabular_learner()}} from package \CRANpkg{fastai}.
}
\section{Dictionary}{

This \link[mlr3:Learner]{Learner} can be instantiated via \link[mlr3:mlr_sugar]{lrn()}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{lrn("classif.python.fastai")
}\if{html}{\out{</div>}}
}

\section{Meta Information}{

\itemize{
\item Task type: \dQuote{classif}
\item Predict Types: \dQuote{response}, \dQuote{prob}
\item Feature Types: \dQuote{logical}, \dQuote{integer}, \dQuote{numeric}, \dQuote{factor}, \dQuote{ordered}
\item Required Packages: \CRANpkg{mlr3}, \CRANpkg{mlr3extralearners}, \CRANpkg{fastai}, \CRANpkg{reticulate}
}
}

\section{Parameters}{
\tabular{lllll}{
   Id \tab Type \tab Default \tab Levels \tab Range \cr
   act_cls \tab untyped \tab - \tab  \tab - \cr
   bn_cont \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
   bn_final \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   drop_last \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   embed_p \tab numeric \tab 0 \tab  \tab \eqn{[0, 1]}{[0, 1]} \cr
   emb_szs \tab untyped \tab NULL \tab  \tab - \cr
   n_epoch \tab integer \tab - \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   eval_metric \tab untyped \tab - \tab  \tab - \cr
   layers \tab untyped \tab - \tab  \tab - \cr
   loss_func \tab untyped \tab - \tab  \tab - \cr
   lr \tab numeric \tab 0.001 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   metrics \tab untyped \tab - \tab  \tab - \cr
   n_out \tab integer \tab - \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   num_workers \tab integer \tab 0 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   opt_func \tab untyped \tab - \tab  \tab - \cr
   patience \tab integer \tab 1 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   pin_memory \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
   procs \tab untyped \tab NULL \tab  \tab - \cr
   ps \tab untyped \tab NULL \tab  \tab - \cr
   shuffle \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   train_bn \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
   wd \tab integer \tab - \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   wd_bn_bias \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   use_bn \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
   y_range \tab untyped \tab NULL \tab  \tab - \cr
   bs \tab integer \tab 50 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
}
}

\references{
Howard, Jeremy, Gugger, Sylvain (2020).
\dQuote{Fastai: A Layered API for Deep Learning.}
\emph{Information}, \bold{11}(2), 108.
ISSN 2078-2489, \doi{10.3390/info11020108}.
}
\author{
annanzrv
}
\section{Super classes}{
\code{\link[mlr3:Learner]{mlr3::Learner}} -> \code{\link[mlr3:LearnerClassif]{mlr3::LearnerClassif}} -> \code{mlr3extralearners::LearnerPythonClassif} -> \code{LearnerPythonClassifFastai}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{internal_valid_scores}}{(named \code{list()} or \code{NULL})
The validation scores extracted from \code{eval_protocol} which itself is set by fitting the \code{fastai::tab_learner}.
If early stopping is activated, this contains the validation scores of the model for the optimal \code{n_epoch},
otherwise the \code{n_epoch} for the final model.}

\item{\code{internal_tuned_values}}{(named \code{list()} or \code{NULL})
If early stopping is activated, this returns a list with \code{n_epoch},
which is the last epoch that yielded improvement w.r.t. the \code{patience}, extracted by \code{max(eval_protocol$epoch)+1}}

\item{\code{validate}}{(\code{numeric(1)} or \code{character(1)} or \code{NULL})
How to construct the internal validation data. This parameter can be either \code{NULL},
a ratio, \code{"test"}, or \code{"predefined"}.}

\item{\code{marshaled}}{(\code{logical(1)})
Whether the learner has been marshaled.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LearnerPythonClassifFastai-new}{\code{LearnerPythonClassifFastai$new()}}
\item \href{#method-LearnerPythonClassifFastai-marshal}{\code{LearnerPythonClassifFastai$marshal()}}
\item \href{#method-LearnerPythonClassifFastai-unmarshal}{\code{LearnerPythonClassifFastai$unmarshal()}}
\item \href{#method-LearnerPythonClassifFastai-clone}{\code{LearnerPythonClassifFastai$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="base_learner"><a href='../../mlr3/html/Learner.html#method-Learner-base_learner'><code>mlr3::Learner$base_learner()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="configure"><a href='../../mlr3/html/Learner.html#method-Learner-configure'><code>mlr3::Learner$configure()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="encapsulate"><a href='../../mlr3/html/Learner.html#method-Learner-encapsulate'><code>mlr3::Learner$encapsulate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="format"><a href='../../mlr3/html/Learner.html#method-Learner-format'><code>mlr3::Learner$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="help"><a href='../../mlr3/html/Learner.html#method-Learner-help'><code>mlr3::Learner$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict"><a href='../../mlr3/html/Learner.html#method-Learner-predict'><code>mlr3::Learner$predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict_newdata"><a href='../../mlr3/html/Learner.html#method-Learner-predict_newdata'><code>mlr3::Learner$predict_newdata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="print"><a href='../../mlr3/html/Learner.html#method-Learner-print'><code>mlr3::Learner$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="reset"><a href='../../mlr3/html/Learner.html#method-Learner-reset'><code>mlr3::Learner$reset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="selected_features"><a href='../../mlr3/html/Learner.html#method-Learner-selected_features'><code>mlr3::Learner$selected_features()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="train"><a href='../../mlr3/html/Learner.html#method-Learner-train'><code>mlr3::Learner$train()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="LearnerClassif" data-id="predict_newdata_fast"><a href='../../mlr3/html/LearnerClassif.html#method-LearnerClassif-predict_newdata_fast'><code>mlr3::LearnerClassif$predict_newdata_fast()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3extralearners" data-topic="LearnerPythonClassif" data-id="py_requirements"><a href='../../mlr3extralearners/html/LearnerPythonClassif.html#method-LearnerPythonClassif-py_requirements'><code>mlr3extralearners::LearnerPythonClassif$py_requirements()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerPythonClassifFastai-new"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerPythonClassifFastai-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerPythonClassifFastai$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerPythonClassifFastai-marshal"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerPythonClassifFastai-marshal}{}}}
\subsection{Method \code{marshal()}}{
Marshal the learner's model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerPythonClassifFastai$marshal(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(any)\cr
Additional arguments passed to \code{\link[mlr3:marshaling]{mlr3::marshal_model()}}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerPythonClassifFastai-unmarshal"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerPythonClassifFastai-unmarshal}{}}}
\subsection{Method \code{unmarshal()}}{
Unmarshal the learner's model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerPythonClassifFastai$unmarshal(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(any)\cr
Additional arguments passed to \code{\link[mlr3:marshaling]{mlr3::marshal_model()}}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerPythonClassifFastai-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerPythonClassifFastai-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerPythonClassifFastai$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
