% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner_RWeka_classif_smo.R
\name{mlr_learners_classif.smo}
\alias{mlr_learners_classif.smo}
\alias{LearnerClassifSMO}
\title{Classification Support Vector Machine Learner}
\description{
Support Vector classifier trained with John Platt's sequential minimal optimization algorithm.
Calls \code{RWeka::SMO()} from \CRANpkg{RWeka}.
}
\section{Custom mlr3 parameters}{

\itemize{
\item \code{output_debug_info}:
\itemize{
\item original id: output-debug-info
}
\item \code{do_not_check_capabilities}:
\itemize{
\item original id: do-not-check-capabilities
}
\item \code{num_decimal_places}:
\itemize{
\item original id: num-decimal-places
}
\item \code{batch_size}:
\itemize{
\item original id: batch-size
}
\item \code{E_poly}:
\itemize{
\item original id: E
}
\item \code{L_poly}:
\itemize{
\item original id: L
}
\item \code{C_poly}:
\itemize{
\item original id: C
}
\item \code{C_logistic}:
\itemize{
\item original id: C
}
\item \code{R_logistic}:
\itemize{
\item original id: R
}
\item \code{M_logistic}:
\itemize{
\item original id: M
}
\item Reason for change: This learner contains changed ids of the following control arguments
since their ids contain irregular pattern
}
}

\section{Dictionary}{

This \link[mlr3:Learner]{Learner} can be instantiated via \link[mlr3:lrn]{lrn()}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{lrn("classif.smo")
}\if{html}{\out{</div>}}
}

\section{Meta Information}{

\itemize{
\item Task type: \dQuote{classif}
\item Predict Types: \dQuote{response}, \dQuote{prob}
\item Feature Types: \dQuote{logical}, \dQuote{integer}, \dQuote{numeric}, \dQuote{factor}, \dQuote{ordered}
\item Required Packages: \CRANpkg{mlr3}, \CRANpkg{RWeka}
}
}

\section{Parameters}{
\tabular{lllll}{
   Id \tab Type \tab Default \tab Levels \tab Range \cr
   subset \tab untyped \tab - \tab  \tab - \cr
   na.action \tab untyped \tab - \tab  \tab - \cr
   no_checks \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   C \tab numeric \tab 1 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   N \tab character \tab 0 \tab 0, 1, 2 \tab - \cr
   L \tab numeric \tab 0.001 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   P \tab numeric \tab 1e-12 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   M \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   V \tab integer \tab -1 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   W \tab integer \tab 1 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   K \tab character \tab PolyKernel \tab NormalizedPolyKernel, PolyKernel, Puk, RBFKernel, StringKernel \tab - \cr
   calibrator \tab untyped \tab "weka.classifiers.functions.Logistic" \tab  \tab - \cr
   E_poly \tab numeric \tab 1 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   L_poly \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   C_poly \tab integer \tab 25007 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   C_logistic \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   R_logistic \tab numeric \tab - \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   M_logistic \tab integer \tab -1 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   output_debug_info \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   do_not_check_capabilities \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   num_decimal_places \tab integer \tab 2 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   batch_size \tab integer \tab 100 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   options \tab untyped \tab NULL \tab  \tab - \cr
}
}

\examples{
\dontshow{if (learner_is_runnable("classif.smo")) withAutoprint(\{ # examplesIf}
# Define the Learner
learner = lrn("classif.smo")
print(learner)

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)


# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
\dontshow{\}) # examplesIf}
}
\references{
Platt J (1998).
\dQuote{Fast Training of Support Vector Machines using Sequential Minimal Optimization.}
In Schoelkopf B, Burges C, Smola A (eds.), \emph{Advances in Kernel Methods - Support Vector Learning}.
MIT Press.
\url{https://www.microsoft.com/en-us/research/publication/sequential-minimal-optimization-a-fast-algorithm-for-training-support-vector-machines/}.

Keerthi S, Shevade S, Bhattacharyya C, Murthy K (2001).
\dQuote{Improvements to Platt's SMO Algorithm for SVM Classifier Design.}
\emph{Neural Computation}, \bold{13}(3), 637-649.

Hastie T, Tibshirani R (1998).
\dQuote{Classification by Pairwise Coupling.}
In Jordan MI, Kearns MJ, Solla SA (eds.), \emph{Advances in Neural Information Processing Systems}, volume 10.
}
\seealso{
\itemize{
\item \link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}.
\item \code{as.data.table(mlr_learners)} for a table of available \link[mlr3:Learner]{Learners} in the running session (depending on the loaded packages).
\item Chapter in the \href{https://mlr3book.mlr-org.com/}{mlr3book}: \url{https://mlr3book.mlr-org.com/chapters/chapter2/data_and_basic_modeling.html#sec-learners}
\item \CRANpkg{mlr3learners} for a selection of recommended learners.
\item \CRANpkg{mlr3cluster} for unsupervised clustering learners.
\item \CRANpkg{mlr3pipelines} to combine learners with pre- and postprocessing steps.
\item \CRANpkg{mlr3tuning} for tuning of hyperparameters, \CRANpkg{mlr3tuningspaces} for established default tuning spaces.
}
}
\author{
damirpolat
}
\section{Super classes}{
\code{\link[mlr3:Learner]{mlr3::Learner}} -> \code{\link[mlr3:LearnerClassif]{mlr3::LearnerClassif}} -> \code{LearnerClassifSMO}
}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{marshaled}}{(\code{logical(1)})\cr
Whether the learner has been marshaled.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-LearnerClassifSMO-initialize}{\code{LearnerClassifSMO$new()}}
    \item \href{#method-LearnerClassifSMO-marshal}{\code{LearnerClassifSMO$marshal()}}
    \item \href{#method-LearnerClassifSMO-unmarshal}{\code{LearnerClassifSMO$unmarshal()}}
    \item \href{#method-LearnerClassifSMO-clone}{\code{LearnerClassifSMO$clone()}}
  }
}
\if{html}{\out{<details><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="base_learner"><a href='../../mlr3/html/Learner.html#method-Learner-base_learner'><code>mlr3::Learner$base_learner()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="configure"><a href='../../mlr3/html/Learner.html#method-Learner-configure'><code>mlr3::Learner$configure()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="encapsulate"><a href='../../mlr3/html/Learner.html#method-Learner-encapsulate'><code>mlr3::Learner$encapsulate()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="format"><a href='../../mlr3/html/Learner.html#method-Learner-format'><code>mlr3::Learner$format()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="help"><a href='../../mlr3/html/Learner.html#method-Learner-help'><code>mlr3::Learner$help()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict"><a href='../../mlr3/html/Learner.html#method-Learner-predict'><code>mlr3::Learner$predict()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict_newdata"><a href='../../mlr3/html/Learner.html#method-Learner-predict_newdata'><code>mlr3::Learner$predict_newdata()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="print"><a href='../../mlr3/html/Learner.html#method-Learner-print'><code>mlr3::Learner$print()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="reset"><a href='../../mlr3/html/Learner.html#method-Learner-reset'><code>mlr3::Learner$reset()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="selected_features"><a href='../../mlr3/html/Learner.html#method-Learner-selected_features'><code>mlr3::Learner$selected_features()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="train"><a href='../../mlr3/html/Learner.html#method-Learner-train'><code>mlr3::Learner$train()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="LearnerClassif" data-id="predict_newdata_fast"><a href='../../mlr3/html/LearnerClassif.html#method-LearnerClassif-predict_newdata_fast'><code>mlr3::LearnerClassif$predict_newdata_fast()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerClassifSMO-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerClassifSMO-initialize}{}}}
\subsection{\code{LearnerClassifSMO$new()}}{
  Creates a new instance of this \link[R6:R6Class]{R6} class.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LearnerClassifSMO$new()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerClassifSMO-marshal"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerClassifSMO-marshal}{}}}
\subsection{\code{LearnerClassifSMO$marshal()}}{
  Marshal the learner's model.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LearnerClassifSMO$marshal(...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{...}}{(any)\cr
Additional arguments passed to \code{\link[mlr3:marshal_model]{mlr3::marshal_model()}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerClassifSMO-unmarshal"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerClassifSMO-unmarshal}{}}}
\subsection{\code{LearnerClassifSMO$unmarshal()}}{
  Unmarshal the learner's model.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LearnerClassifSMO$unmarshal(...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{...}}{(any)\cr
Additional arguments passed to \code{\link[mlr3:unmarshal_model]{mlr3::unmarshal_model()}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerClassifSMO-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerClassifSMO-clone}{}}}
\subsection{\code{LearnerClassifSMO$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LearnerClassifSMO$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
