% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner_CoxBoost_cmprsk_coxboost.R
\name{mlr_learners_cmprsk.coxboost}
\alias{mlr_learners_cmprsk.coxboost}
\alias{LearnerCompRisksCoxboost}
\title{Competing Risks Cox Learner with Likelihood Based Boosting}
\description{
Fits a competing risks Cox model using likelihood based boosting.
Calls \code{CoxBoost::CoxBoost()} or \code{CoxBoost::cv.CoxBoost()} from package
\CRANpkg{CoxBoost}.
Predicts cumulative incidence functions (one CIF matrix per cause, rows = observations,
columns = times).
If \code{times} are not provided for prediction, the unique event times (across causes)
from the training set are used by default.
}
\section{Dictionary}{

This \link[mlr3:Learner]{Learner} can be instantiated via \link[mlr3:mlr_sugar]{lrn()}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{lrn("cmprsk.coxboost")
}\if{html}{\out{</div>}}
}

\section{Meta Information}{

\itemize{
\item Task type: \dQuote{cmprsk}
\item Predict Types: \dQuote{cif}
\item Feature Types: \dQuote{integer}, \dQuote{numeric}
\item Required Packages: \CRANpkg{mlr3}, \CRANpkg{mlr3cmprsk}, \CRANpkg{mlr3extralearners}, \CRANpkg{CoxBoost}
}
}

\section{Parameters}{
\tabular{lllll}{
   Id \tab Type \tab Default \tab Levels \tab Range \cr
   cv \tab logical \tab - \tab TRUE, FALSE \tab - \cr
   maxstepno \tab integer \tab 100 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   K \tab integer \tab 10 \tab  \tab \eqn{[2, \infty)}{[2, Inf)} \cr
   type \tab character \tab verweij \tab verweij, naive \tab - \cr
   folds \tab untyped \tab NULL \tab  \tab - \cr
   minstepno \tab integer \tab 50 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   start.penalty \tab numeric \tab - \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   iter.max \tab integer \tab 10 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   upper.margin \tab numeric \tab 0.05 \tab  \tab \eqn{[0, 1]}{[0, 1]} \cr
   unpen.index \tab untyped \tab - \tab  \tab - \cr
   standardize \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
   stepno \tab integer \tab 100 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   penalty \tab numeric \tab - \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   criterion \tab character \tab pscore \tab pscore, score, hpscore, hscore \tab - \cr
   stepsize.factor \tab numeric \tab 1 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   sf.scheme \tab character \tab sigmoid \tab sigmoid, linear \tab - \cr
   pendistmat \tab untyped \tab - \tab  \tab - \cr
   connected.index \tab untyped \tab - \tab  \tab - \cr
   x.is.01 \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   return.score \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
   trace \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   cmprsk \tab character \tab - \tab csh, ccsh \tab - \cr
   coupled.strata \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
   at.step \tab untyped \tab - \tab  \tab - \cr
   times \tab untyped \tab - \tab  \tab - \cr
}
}

\section{Custom mlr3 parameters}{

\itemize{
\item \code{cv}: If \code{TRUE}, uses cross-validation with \code{CoxBoost::cv.CoxBoost()} to
select the optimal number of boosting steps, then refits the final model using
\code{CoxBoost::CoxBoost()} with this value. Initialized to \code{FALSE} by default.
\item \code{penalty}: If set to \code{"optimCoxBoostPenalty"}, uses \code{CoxBoost::optimCoxBoostPenalty()}
to determine the optimal penalty value and number of steps (via
\code{CoxBoost::cv.CoxBoost()}) for the final model.
}
}

\section{Initial parameter values}{

\itemize{
\item \code{cmprsk}: Type of competing risks analysis. Initialized to \code{"csh"}
(\strong{cause-specific hazards}) which models each cause separately, treating other
causes as censored. \code{"ccsh"} (\strong{combined cause-specific hazards}) is another
option which combines information across causes for variable selection using
a joint criterion.
}
}

\examples{
\dontshow{if (learner_is_runnable("cmprsk.coxboost")) withAutoprint(\{ # examplesIf}
# Define the Learner
learner = lrn("cmprsk.coxboost", stepno = 10)
print(learner)

# Define a Task
task = tsk("pbc")
task$col_roles$feature = c("age", "chol", "copper", "platelet", "trig",
 "albumin", "alk.phos", "ast", "bili", "protime")

# Stratification based on event
task$set_col_roles(cols = "status", add_to = "stratum")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)
print(learner$selected_features(cause = 1)) # for cause = 1
print(learner$selected_features(cause = 2)) # for cause = 2

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
\dontshow{\}) # examplesIf}
}
\references{
Binder, Harald, Allignol, Arthur, Schumacher, Martin, Beyersmann, Jan (2009).
\dQuote{Boosting for high-dimensional time-to-event data with competing risks.}
\emph{Bioinformatics}, \bold{25}(7), 890--896.
}
\seealso{
\itemize{
\item \link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}.
\item \code{as.data.table(mlr_learners)} for a table of available \link[mlr3:Learner]{Learners} in the running session (depending on the loaded packages).
\item Chapter in the \href{https://mlr3book.mlr-org.com/}{mlr3book}: \url{https://mlr3book.mlr-org.com/basics.html#learners}
\item \CRANpkg{mlr3learners} for a selection of recommended learners.
\item \CRANpkg{mlr3cluster} for unsupervised clustering learners.
\item \CRANpkg{mlr3pipelines} to combine learners with pre- and postprocessing steps.
\item \CRANpkg{mlr3tuning} for tuning of hyperparameters, \CRANpkg{mlr3tuningspaces} for established default tuning spaces.
}
}
\author{
bblodfon
}
\section{Super classes}{
\code{\link[mlr3:Learner]{mlr3::Learner}} -> \code{\link[mlr3cmprsk:LearnerCompRisks]{mlr3cmprsk::LearnerCompRisks}} -> \code{LearnerCompRisksCoxboost}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LearnerCompRisksCoxboost-new}{\code{LearnerCompRisksCoxboost$new()}}
\item \href{#method-LearnerCompRisksCoxboost-selected_features}{\code{LearnerCompRisksCoxboost$selected_features()}}
\item \href{#method-LearnerCompRisksCoxboost-clone}{\code{LearnerCompRisksCoxboost$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="base_learner"><a href='../../mlr3/html/Learner.html#method-Learner-base_learner'><code>mlr3::Learner$base_learner()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="configure"><a href='../../mlr3/html/Learner.html#method-Learner-configure'><code>mlr3::Learner$configure()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="encapsulate"><a href='../../mlr3/html/Learner.html#method-Learner-encapsulate'><code>mlr3::Learner$encapsulate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="format"><a href='../../mlr3/html/Learner.html#method-Learner-format'><code>mlr3::Learner$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="help"><a href='../../mlr3/html/Learner.html#method-Learner-help'><code>mlr3::Learner$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict"><a href='../../mlr3/html/Learner.html#method-Learner-predict'><code>mlr3::Learner$predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict_newdata"><a href='../../mlr3/html/Learner.html#method-Learner-predict_newdata'><code>mlr3::Learner$predict_newdata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="print"><a href='../../mlr3/html/Learner.html#method-Learner-print'><code>mlr3::Learner$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="reset"><a href='../../mlr3/html/Learner.html#method-Learner-reset'><code>mlr3::Learner$reset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="train"><a href='../../mlr3/html/Learner.html#method-Learner-train'><code>mlr3::Learner$train()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerCompRisksCoxboost-new"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerCompRisksCoxboost-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerCompRisksCoxboost$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerCompRisksCoxboost-selected_features"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerCompRisksCoxboost-selected_features}{}}}
\subsection{Method \code{selected_features()}}{
Returns the set of cause-specific selected features which have non-zero coefficients.
Calls the internal \code{coef.CoxBoost()} function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerCompRisksCoxboost$selected_features(at_step = NULL, cause = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{at_step}}{(\code{integer(1)})\cr
Which boosting step to get the coefficients for. If no step is given
(default), the final boosting step is used.}

\item{\code{cause}}{(\code{integer(1)})\cr
Integer value indicating the event of interest.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{character()}) vector of feature names.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerCompRisksCoxboost-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerCompRisksCoxboost-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerCompRisksCoxboost$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
