% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner_botorch_regr_fullybayesian.R
\name{mlr_learners_regr.botorch_fullybayesian}
\alias{mlr_learners_regr.botorch_fullybayesian}
\alias{LearnerRegrBotorchFullyBayesian}
\title{BoTorch Fully Bayesian SingleTaskGP Regression Learner}
\description{
Fully Bayesian Gaussian Process via \href{https://botorch.org/}{botorch}, using the \code{SaasFullyBayesianSingleTaskGP} model.
Unlike standard GP models that use MAP estimation for hyperparameters, this model uses MCMC (NUTS) to learn
full posterior distributions over kernel hyperparameters.
See \href{https://botorch.readthedocs.io/en/latest/models.html#module-botorch.models.fully_bayesian}{here} for more details.
Uses \CRANpkg{reticulate} to interface with Python.
}
\section{Dictionary}{

This \link[mlr3:Learner]{Learner} can be instantiated via \link[mlr3:mlr_sugar]{lrn()}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{lrn("regr.botorch_fullybayesian")
}\if{html}{\out{</div>}}
}

\section{Meta Information}{

\itemize{
\item Task type: \dQuote{regr}
\item Predict Types: \dQuote{response}, \dQuote{se}
\item Feature Types: \dQuote{integer}, \dQuote{numeric}
\item Required Packages: \CRANpkg{mlr3}, \CRANpkg{mlr3extralearners}, \CRANpkg{reticulate}
}
}

\section{Parameters}{
\tabular{lllll}{
   Id \tab Type \tab Default \tab Levels \tab Range \cr
   device \tab character \tab cpu \tab cpu, cuda \tab - \cr
   input_transform \tab character \tab - \tab normalize, standardize, log10, warp, none \tab - \cr
   outcome_transform \tab character \tab - \tab standardize, none \tab - \cr
   warmup_steps \tab integer \tab - \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   num_samples \tab integer \tab - \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   thinning \tab integer \tab - \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   max_tree_depth \tab integer \tab - \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   disable_progbar \tab logical \tab - \tab TRUE, FALSE \tab - \cr
   jit_compile \tab logical \tab - \tab TRUE, FALSE \tab - \cr
}
}

\author{
Marc Becker
}
\section{Super classes}{
\code{\link[mlr3:Learner]{mlr3::Learner}} -> \code{\link[mlr3:LearnerRegr]{mlr3::LearnerRegr}} -> \code{LearnerRegrBotorchFullyBayesian}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{marshaled}}{(\code{logical(1)})
Whether the learner has been marshaled.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LearnerRegrBotorchFullyBayesian-new}{\code{LearnerRegrBotorchFullyBayesian$new()}}
\item \href{#method-LearnerRegrBotorchFullyBayesian-marshal}{\code{LearnerRegrBotorchFullyBayesian$marshal()}}
\item \href{#method-LearnerRegrBotorchFullyBayesian-unmarshal}{\code{LearnerRegrBotorchFullyBayesian$unmarshal()}}
\item \href{#method-LearnerRegrBotorchFullyBayesian-clone}{\code{LearnerRegrBotorchFullyBayesian$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="base_learner"><a href='../../mlr3/html/Learner.html#method-Learner-base_learner'><code>mlr3::Learner$base_learner()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="configure"><a href='../../mlr3/html/Learner.html#method-Learner-configure'><code>mlr3::Learner$configure()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="encapsulate"><a href='../../mlr3/html/Learner.html#method-Learner-encapsulate'><code>mlr3::Learner$encapsulate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="format"><a href='../../mlr3/html/Learner.html#method-Learner-format'><code>mlr3::Learner$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="help"><a href='../../mlr3/html/Learner.html#method-Learner-help'><code>mlr3::Learner$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict"><a href='../../mlr3/html/Learner.html#method-Learner-predict'><code>mlr3::Learner$predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict_newdata"><a href='../../mlr3/html/Learner.html#method-Learner-predict_newdata'><code>mlr3::Learner$predict_newdata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="print"><a href='../../mlr3/html/Learner.html#method-Learner-print'><code>mlr3::Learner$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="reset"><a href='../../mlr3/html/Learner.html#method-Learner-reset'><code>mlr3::Learner$reset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="selected_features"><a href='../../mlr3/html/Learner.html#method-Learner-selected_features'><code>mlr3::Learner$selected_features()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="train"><a href='../../mlr3/html/Learner.html#method-Learner-train'><code>mlr3::Learner$train()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="LearnerRegr" data-id="predict_newdata_fast"><a href='../../mlr3/html/LearnerRegr.html#method-LearnerRegr-predict_newdata_fast'><code>mlr3::LearnerRegr$predict_newdata_fast()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerRegrBotorchFullyBayesian-new"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerRegrBotorchFullyBayesian-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerRegrBotorchFullyBayesian$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerRegrBotorchFullyBayesian-marshal"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerRegrBotorchFullyBayesian-marshal}{}}}
\subsection{Method \code{marshal()}}{
Marshal the learner's model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerRegrBotorchFullyBayesian$marshal(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(any)\cr
Additional arguments passed to \code{\link[mlr3:marshaling]{mlr3::marshal_model()}}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerRegrBotorchFullyBayesian-unmarshal"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerRegrBotorchFullyBayesian-unmarshal}{}}}
\subsection{Method \code{unmarshal()}}{
Unmarshal the learner's model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerRegrBotorchFullyBayesian$unmarshal(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(any)\cr
Additional arguments passed to \code{\link[mlr3:marshaling]{mlr3::unmarshal_model()}}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerRegrBotorchFullyBayesian-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerRegrBotorchFullyBayesian-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerRegrBotorchFullyBayesian$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
