% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner_partykit_regr_cforest.R
\name{mlr_learners_regr.cforest}
\alias{mlr_learners_regr.cforest}
\alias{LearnerRegrCForest}
\title{Regression Conditional Random Forest Learner}
\description{
Calls \link[partykit:cforest]{partykit::cforest} from package \CRANpkg{partykit}.
}
\section{Dictionary}{
 This \link{Learner} can be instantiated via the
\link[mlr3misc:Dictionary]{dictionary} \link{mlr_learners} or with the associated
sugar function \code{\link[=lrn]{lrn()}}:\preformatted{mlr_learners$get("regr.cforest")
lrn("regr.cforest")
}
}

\section{Traits}{

\itemize{
\item Packages: partykit, sandwich, coin
\item Predict Types: response
\item Feature Types: integer, numeric, factor, ordered
\item Properties: oob_error, weights
}
}

\examples{
if (mlr3misc::require_namespaces(c("partykit", "sandwich", "coin"))) {
  learner = mlr3::lrn("regr.cforest")
  print(learner)

  # available parameters:
  learner$param_set$ids()
}
}
\references{
Hothorn T, Zeileis A (2015).
“partykit: A Modular Toolkit for Recursive Partytioning in R.”
Journal of Machine Learning Research, 16(118), 3905-3909.
\url{http://jmlr.org/papers/v16/hothorn15a.html}

Hothorn T, Hornik K, Zeileis A (2006).
“Unbiased Recursive Partitioning: A Conditional Inference Framework.”
Journal of Computational and Graphical Statistics, 15(3), 651–674.
\doi{10.1198/106186006x133933}
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}:
\link[mlr3:mlr_learners]{mlr3::mlr_learners}
}
\author{
sumny
}
\section{Super classes}{
\code{\link[mlr3:Learner]{mlr3::Learner}} -> \code{\link[mlr3:LearnerRegr]{mlr3::LearnerRegr}} -> \code{LearnerRegrCForest}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LearnerRegrCForest$new()}}
\item \href{#method-oob_error}{\code{LearnerRegrCForest$oob_error()}}
\item \href{#method-clone}{\code{LearnerRegrCForest$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="format">}\href{../../mlr3/html/Learner.html#method-format}{\code{mlr3::Learner$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="help">}\href{../../mlr3/html/Learner.html#method-help}{\code{mlr3::Learner$help()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict">}\href{../../mlr3/html/Learner.html#method-predict}{\code{mlr3::Learner$predict()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict_newdata">}\href{../../mlr3/html/Learner.html#method-predict_newdata}{\code{mlr3::Learner$predict_newdata()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="print">}\href{../../mlr3/html/Learner.html#method-print}{\code{mlr3::Learner$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="reset">}\href{../../mlr3/html/Learner.html#method-reset}{\code{mlr3::Learner$reset()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="train">}\href{../../mlr3/html/Learner.html#method-train}{\code{mlr3::Learner$train()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerRegrCForest$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-oob_error"></a>}}
\if{latex}{\out{\hypertarget{method-oob_error}{}}}
\subsection{Method \code{oob_error()}}{
The out-of-bag error, calculated using the OOB predictions from
\code{partykit}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerRegrCForest$oob_error()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{numeric(1)}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerRegrCForest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
