% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner_frbs_regr_frbs.R
\name{mlr_learners_regr.frbs}
\alias{mlr_learners_regr.frbs}
\alias{LearnerRegrFrbs}
\title{Regression Fuzzy Rule-based System Learner}
\description{
Fuzzy rule-based regression learner supporting multiple inference engines such as WM, HYFIS,
ANFIS, DENFIS, SBC, and several genetic fuzzy systems.
Calls \code{frbs::frbs.learn()} from \CRANpkg{frbs}.
}
\section{Dictionary}{

This \link[mlr3:Learner]{Learner} can be instantiated via \link[mlr3:lrn]{lrn()}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{lrn("regr.frbs")
}\if{html}{\out{</div>}}
}

\section{Meta Information}{

\itemize{
\item Task type: \dQuote{regr}
\item Predict Types: \dQuote{response}
\item Feature Types: \dQuote{integer}, \dQuote{numeric}
\item Required Packages: \CRANpkg{mlr3}, \CRANpkg{mlr3extralearners}, \CRANpkg{frbs}
}
}

\section{Parameters}{
\tabular{lllll}{
   Id \tab Type \tab Default \tab Levels \tab Range \cr
   alpha.heuristic \tab numeric \tab 1 \tab  \tab \eqn{[2.22044604925031e-16, \infty)}{[2.22044604925031e-16, Inf)} \cr
   d \tab numeric \tab 2 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   Dthr \tab numeric \tab 0.1 \tab  \tab \eqn{[0, 1]}{[0, 1]} \cr
   eps.high \tab numeric \tab 0.5 \tab  \tab \eqn{[2.22044604925031e-16, \infty)}{[2.22044604925031e-16, Inf)} \cr
   eps.low \tab numeric \tab 0.15 \tab  \tab \eqn{[2.22044604925031e-16, \infty)}{[2.22044604925031e-16, Inf)} \cr
   epsilon \tab numeric \tab 0.9 \tab  \tab \eqn{[0, 1]}{[0, 1]} \cr
   max.gen \tab integer \tab 10 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   max.iter \tab integer \tab 10 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   max.tune \tab integer \tab 10 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   method.type \tab character \tab WM \tab WM, SBC, HYFIS, ANFIS, DENFIS, FS.HGD, FIR.DM, GFS.FR.MOGUL, GFS.THRIFT, GFS.LT.RS \tab - \cr
   mode.tuning \tab character \tab GLOBAL \tab GLOBAL, LOCAL \tab - \cr
   num.labels \tab integer \tab 7 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   persen_cross \tab numeric \tab 1 \tab  \tab \eqn{[0, 1]}{[0, 1]} \cr
   persen_mutant \tab numeric \tab 1 \tab  \tab \eqn{[0, 1]}{[0, 1]} \cr
   popu.size \tab integer \tab 10 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   r.a \tab numeric \tab 0.5 \tab  \tab \eqn{[2.22044604925031e-16, \infty)}{[2.22044604925031e-16, Inf)} \cr
   range.data \tab untyped \tab NULL \tab  \tab - \cr
   rule.selection \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   step.size \tab numeric \tab 0.01 \tab  \tab \eqn{[0, 1]}{[0, 1]} \cr
   type.defuz \tab character \tab WAM \tab WAM, FIRST.MAX, LAST.MAX, MEAN.MAX, COG \tab - \cr
   type.implication.func \tab character \tab ZADEH \tab DIENES_RESHER, LUKASIEWICZ, ZADEH, GOGUEN, GODEL, SHARP, MIZUMOTO, DUBOIS_PRADE, MIN \tab - \cr
   type.mf \tab character \tab GAUSSIAN \tab TRIANGLE, TRAPEZOID, GAUSSIAN, SIGMOID, BELL \tab - \cr
   type.snorm \tab character \tab MAX \tab MAX, HAMACHER, YAGER, PRODUCT, BOUNDED \tab - \cr
   type.tnorm \tab character \tab MIN \tab MIN, HAMACHER, YAGER, PRODUCT, BOUNDED \tab - \cr
}
}

\examples{
\dontshow{if (learner_is_runnable("regr.frbs")) withAutoprint(\{ # examplesIf}
# Define the Learner
learner = lrn("regr.frbs")
print(learner)

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)


# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
\dontshow{\}) # examplesIf}
}
\references{
Riza LS, Bergmeir C, Herrera F, Benitez JM (2015).
\dQuote{frbs: Fuzzy Rule-Based Systems for Classification and Regression in R.}
\emph{Journal of Statistical Software}, \bold{65}(6), 1--30.
\url{https://www.jstatsoft.org/v65/i06/}.
}
\seealso{
\itemize{
\item \link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}.
\item \code{as.data.table(mlr_learners)} for a table of available \link[mlr3:Learner]{Learners} in the running session (depending on the loaded packages).
\item Chapter in the \href{https://mlr3book.mlr-org.com/}{mlr3book}: \url{https://mlr3book.mlr-org.com/chapters/chapter2/data_and_basic_modeling.html#sec-learners}
\item \CRANpkg{mlr3learners} for a selection of recommended learners.
\item \CRANpkg{mlr3cluster} for unsupervised clustering learners.
\item \CRANpkg{mlr3pipelines} to combine learners with pre- and postprocessing steps.
\item \CRANpkg{mlr3tuning} for tuning of hyperparameters, \CRANpkg{mlr3tuningspaces} for established default tuning spaces.
}
}
\author{
awinterstetter
}
\section{Super classes}{
\code{\link[mlr3:Learner]{mlr3::Learner}} -> \code{\link[mlr3:LearnerRegr]{mlr3::LearnerRegr}} -> \code{LearnerRegrFrbs}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-LearnerRegrFrbs-initialize}{\code{LearnerRegrFrbs$new()}}
    \item \href{#method-LearnerRegrFrbs-clone}{\code{LearnerRegrFrbs$clone()}}
  }
}
\if{html}{\out{<details><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="base_learner"><a href='../../mlr3/html/Learner.html#method-Learner-base_learner'><code>mlr3::Learner$base_learner()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="configure"><a href='../../mlr3/html/Learner.html#method-Learner-configure'><code>mlr3::Learner$configure()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="encapsulate"><a href='../../mlr3/html/Learner.html#method-Learner-encapsulate'><code>mlr3::Learner$encapsulate()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="format"><a href='../../mlr3/html/Learner.html#method-Learner-format'><code>mlr3::Learner$format()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="help"><a href='../../mlr3/html/Learner.html#method-Learner-help'><code>mlr3::Learner$help()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict"><a href='../../mlr3/html/Learner.html#method-Learner-predict'><code>mlr3::Learner$predict()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict_newdata"><a href='../../mlr3/html/Learner.html#method-Learner-predict_newdata'><code>mlr3::Learner$predict_newdata()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="print"><a href='../../mlr3/html/Learner.html#method-Learner-print'><code>mlr3::Learner$print()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="reset"><a href='../../mlr3/html/Learner.html#method-Learner-reset'><code>mlr3::Learner$reset()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="selected_features"><a href='../../mlr3/html/Learner.html#method-Learner-selected_features'><code>mlr3::Learner$selected_features()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="train"><a href='../../mlr3/html/Learner.html#method-Learner-train'><code>mlr3::Learner$train()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="LearnerRegr" data-id="predict_newdata_fast"><a href='../../mlr3/html/LearnerRegr.html#method-LearnerRegr-predict_newdata_fast'><code>mlr3::LearnerRegr$predict_newdata_fast()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerRegrFrbs-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerRegrFrbs-initialize}{}}}
\subsection{\code{LearnerRegrFrbs$new()}}{
  Creates a new instance of this \link[R6:R6Class]{R6} class.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LearnerRegrFrbs$new()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerRegrFrbs-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerRegrFrbs-clone}{}}}
\subsection{\code{LearnerRegrFrbs$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LearnerRegrFrbs$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
