% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner_SpatialML_regr_gwrf.R
\name{mlr_learners_regr.gwrf}
\alias{mlr_learners_regr.gwrf}
\alias{LearnerRegrGWRF}
\title{Geographically Weighted Random Forest Learner}
\description{
Geographically Weighted Random Forest for regression.
Calls \code{SpatialML::grf()} from the SpatialML package.
Note that SpatialML is not on CRAN. To install it, use:
\code{remotes::install_version("SpatialML", version = "0.1.6")}
}
\author{
Manh Hung LE
}
\section{Super classes}{
\code{\link[mlr3:Learner]{mlr3::Learner}} -> \code{\link[mlr3:LearnerRegr]{mlr3::LearnerRegr}} -> \code{LearnerRegrGWRF}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-LearnerRegrGWRF-initialize}{\code{LearnerRegrGWRF$new()}}
    \item \href{#method-LearnerRegrGWRF-importance}{\code{LearnerRegrGWRF$importance()}}
    \item \href{#method-LearnerRegrGWRF-oob_error}{\code{LearnerRegrGWRF$oob_error()}}
    \item \href{#method-LearnerRegrGWRF-clone}{\code{LearnerRegrGWRF$clone()}}
  }
}
\if{html}{\out{<details><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="base_learner"><a href='../../mlr3/html/Learner.html#method-Learner-base_learner'><code>mlr3::Learner$base_learner()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="configure"><a href='../../mlr3/html/Learner.html#method-Learner-configure'><code>mlr3::Learner$configure()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="encapsulate"><a href='../../mlr3/html/Learner.html#method-Learner-encapsulate'><code>mlr3::Learner$encapsulate()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="format"><a href='../../mlr3/html/Learner.html#method-Learner-format'><code>mlr3::Learner$format()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="help"><a href='../../mlr3/html/Learner.html#method-Learner-help'><code>mlr3::Learner$help()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict"><a href='../../mlr3/html/Learner.html#method-Learner-predict'><code>mlr3::Learner$predict()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict_newdata"><a href='../../mlr3/html/Learner.html#method-Learner-predict_newdata'><code>mlr3::Learner$predict_newdata()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="print"><a href='../../mlr3/html/Learner.html#method-Learner-print'><code>mlr3::Learner$print()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="reset"><a href='../../mlr3/html/Learner.html#method-Learner-reset'><code>mlr3::Learner$reset()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="selected_features"><a href='../../mlr3/html/Learner.html#method-Learner-selected_features'><code>mlr3::Learner$selected_features()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="train"><a href='../../mlr3/html/Learner.html#method-Learner-train'><code>mlr3::Learner$train()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="LearnerRegr" data-id="predict_newdata_fast"><a href='../../mlr3/html/LearnerRegr.html#method-LearnerRegr-predict_newdata_fast'><code>mlr3::LearnerRegr$predict_newdata_fast()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerRegrGWRF-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerRegrGWRF-initialize}{}}}
\subsection{\code{LearnerRegrGWRF$new()}}{
  Creates a new instance of this \link[R6:R6Class]{R6} class.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LearnerRegrGWRF$new()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerRegrGWRF-importance"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerRegrGWRF-importance}{}}}
\subsection{\code{LearnerRegrGWRF$importance()}}{
  The importance scores are extracted from the slot \code{Local.Variable.Importance}.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LearnerRegrGWRF$importance()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Named \code{numeric()}.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerRegrGWRF-oob_error"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerRegrGWRF-oob_error}{}}}
\subsection{\code{LearnerRegrGWRF$oob_error()}}{
  OOB errors are extracted from the model slot \code{Global.Model$prediction.error}.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LearnerRegrGWRF$oob_error()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    \code{numeric(1)}.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerRegrGWRF-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerRegrGWRF-clone}{}}}
\subsection{\code{LearnerRegrGWRF$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LearnerRegrGWRF$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
