% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner_h2o_regr_glm.R
\name{mlr_learners_regr.h2o.glm}
\alias{mlr_learners_regr.h2o.glm}
\alias{LearnerRegrH2OGLM}
\title{Generalized Linear Regression H2O Learner}
\description{
Generalized linear model.
Calls \code{\link[h2o:h2o.glm]{h2o::h2o.glm()}} from package \CRANpkg{h2o}.
}
\section{Dictionary}{

This \link[mlr3:Learner]{Learner} can be instantiated via \link[mlr3:mlr_sugar]{lrn()}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{lrn("regr.h2o.glm")
}\if{html}{\out{</div>}}
}

\section{Meta Information}{

\itemize{
\item Task type: \dQuote{regr}
\item Predict Types: \dQuote{response}
\item Feature Types: \dQuote{integer}, \dQuote{numeric}, \dQuote{factor}
\item Required Packages: \CRANpkg{mlr3}, \CRANpkg{h2o}
}
}

\section{Parameters}{
\tabular{lllll}{
   Id \tab Type \tab Default \tab Levels \tab Range \cr
   seed \tab integer \tab -1 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   random_columns \tab untyped \tab NULL \tab  \tab - \cr
   ignore_const_cols \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
   score_each_iteration \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   score_iteration_interval \tab integer \tab -1 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   solver \tab character \tab IRLSM \tab AUTO, IRLSM, L_BFGS, COORDINATE_DESCENT_NAIVE, COORDINATE_DESCENT, GRADIENT_DESCENT_LH, GRADIENT_DESCENT_SQERR \tab - \cr
   rand_family \tab untyped \tab NULL \tab  \tab - \cr
   alpha \tab untyped \tab 0.5 \tab  \tab - \cr
   lambda \tab untyped \tab 1e-05 \tab  \tab - \cr
   lambda_search \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   early_stopping \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
   nlambdas \tab integer \tab - \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   standardize \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
   missing_values_handling \tab character \tab - \tab MeanImputation, Skip, PlugValues \tab - \cr
   plug_values \tab untyped \tab NULL \tab  \tab - \cr
   compute_p_values \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   dispersion_parameter_method \tab character \tab - \tab deviance, pearson, ml \tab - \cr
   init_dispersion_parameter \tab numeric \tab 1 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   remove_collinear_columns \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   intercept \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
   non_negative \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   max_iterations \tab integer \tab 50 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   objective_epsilon \tab numeric \tab -1 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   beta_epsilon \tab numeric \tab 0 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   gradient_epsilon \tab numeric \tab -1 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   link \tab character \tab identity \tab family_default, identity, log, inverse \tab - \cr
   rand_link \tab untyped \tab NULL \tab  \tab - \cr
   startval \tab untyped \tab NULL \tab  \tab - \cr
   calc_like \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   HGLM \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   cold_start \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   lambda_min_ratio \tab numeric \tab - \tab  \tab \eqn{[0, 1]}{[0, 1]} \cr
   beta_constraints \tab untyped \tab NULL \tab  \tab - \cr
   max_active_predictors \tab integer \tab -1 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   interactions \tab untyped \tab NULL \tab  \tab - \cr
   interaction_pairs \tab untyped \tab NULL \tab  \tab - \cr
   obj_reg \tab numeric \tab -1 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   stopping_rounds \tab integer \tab 0 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   stopping_metric \tab character \tab AUTO \tab AUTO, deviance, MSE, RMSE, MAE, RMSLE, custom, custom_increasing \tab - \cr
   stopping_tolerance \tab numeric \tab 0.001 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   max_runtime_secs \tab numeric \tab 0 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   custom_metric_func \tab untyped \tab NULL \tab  \tab - \cr
   generate_scoring_history \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   dispersion_epsilon \tab numeric \tab 1e-04 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   max_iterations_dispersion \tab integer \tab 3000 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   build_null_model \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   fix_dispersion_parameter \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   generate_variable_inflation_factors \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   dispersion_learning_rate \tab numeric \tab 0.5 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   influence \tab untyped \tab NULL \tab  \tab - \cr
}
}

\section{Model family}{

This learner calls \code{\link[h2o:h2o.glm]{h2o::h2o.glm()}} with \code{family = "gaussian"}
}

\examples{
\dontshow{if (learner_is_runnable("regr.h2o.glm")) withAutoprint(\{ # examplesIf}
# Define the Learner
learner = lrn("regr.h2o.glm")
print(learner)

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)


# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
\dontshow{\}) # examplesIf}
}
\references{
Fryda T, LeDell E, Gill N, Aiello S, Fu A, Candel A, Click C, Kraljevic T, Nykodym T, Aboyoun P, Kurka M, Malohlava M, Poirier S, Wong W (2025).
\emph{h2o: R Interface for the 'H2O' Scalable Machine Learning Platform}.
R package version 3.46.0.9, \url{https://github.com/h2oai/h2o-3}.
}
\seealso{
\itemize{
\item \link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}.
\item \code{as.data.table(mlr_learners)} for a table of available \link[mlr3:Learner]{Learners} in the running session (depending on the loaded packages).
\item Chapter in the \href{https://mlr3book.mlr-org.com/}{mlr3book}: \url{https://mlr3book.mlr-org.com/basics.html#learners}
\item \CRANpkg{mlr3learners} for a selection of recommended learners.
\item \CRANpkg{mlr3cluster} for unsupervised clustering learners.
\item \CRANpkg{mlr3pipelines} to combine learners with pre- and postprocessing steps.
\item \CRANpkg{mlr3tuning} for tuning of hyperparameters, \CRANpkg{mlr3tuningspaces} for established default tuning spaces.
}
}
\author{
awinterstetter
}
\section{Super classes}{
\code{\link[mlr3:Learner]{mlr3::Learner}} -> \code{\link[mlr3:LearnerRegr]{mlr3::LearnerRegr}} -> \code{LearnerRegrH2OGLM}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LearnerRegrH2OGLM-new}{\code{LearnerRegrH2OGLM$new()}}
\item \href{#method-LearnerRegrH2OGLM-clone}{\code{LearnerRegrH2OGLM$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="base_learner"><a href='../../mlr3/html/Learner.html#method-Learner-base_learner'><code>mlr3::Learner$base_learner()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="configure"><a href='../../mlr3/html/Learner.html#method-Learner-configure'><code>mlr3::Learner$configure()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="encapsulate"><a href='../../mlr3/html/Learner.html#method-Learner-encapsulate'><code>mlr3::Learner$encapsulate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="format"><a href='../../mlr3/html/Learner.html#method-Learner-format'><code>mlr3::Learner$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="help"><a href='../../mlr3/html/Learner.html#method-Learner-help'><code>mlr3::Learner$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict"><a href='../../mlr3/html/Learner.html#method-Learner-predict'><code>mlr3::Learner$predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict_newdata"><a href='../../mlr3/html/Learner.html#method-Learner-predict_newdata'><code>mlr3::Learner$predict_newdata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="print"><a href='../../mlr3/html/Learner.html#method-Learner-print'><code>mlr3::Learner$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="reset"><a href='../../mlr3/html/Learner.html#method-Learner-reset'><code>mlr3::Learner$reset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="selected_features"><a href='../../mlr3/html/Learner.html#method-Learner-selected_features'><code>mlr3::Learner$selected_features()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="train"><a href='../../mlr3/html/Learner.html#method-Learner-train'><code>mlr3::Learner$train()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="LearnerRegr" data-id="predict_newdata_fast"><a href='../../mlr3/html/LearnerRegr.html#method-LearnerRegr-predict_newdata_fast'><code>mlr3::LearnerRegr$predict_newdata_fast()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerRegrH2OGLM-new"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerRegrH2OGLM-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerRegrH2OGLM$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerRegrH2OGLM-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerRegrH2OGLM-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerRegrH2OGLM$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
