% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner_laGP_regr_laGP.R
\name{mlr_learners_regr.laGP}
\alias{mlr_learners_regr.laGP}
\alias{LearnerRegrLaGP}
\title{Regression Local Approximate Gaussian Process Learner}
\description{
Local approximate Gaussian process for regression.
Calls \code{\link[laGP:aGP]{laGP::aGP()}} from \CRANpkg{laGP}.
}
\note{
Parameters \code{start} and \code{end} tune the initial and maximum neighborhood sizes
used for the local GP fit, \code{d} can fix the length-scale parameters, \code{g}
configures the nugget/regularization term, and \code{method} selects the search
heuristic (\code{alc}, \code{alcray}, \code{efi}, \code{mspe}, \code{nn}). Arguments \code{close} and
\code{numrays} further refine the ray-based search used by \code{method = "alcray"},
while \code{verb} controls the output verbosity.
}
\section{Dictionary}{

This \link[mlr3:Learner]{Learner} can be instantiated via \link[mlr3:mlr_sugar]{lrn()}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{lrn("regr.laGP")
}\if{html}{\out{</div>}}
}

\section{Meta Information}{

\itemize{
\item Task type: \dQuote{regr}
\item Predict Types: \dQuote{response}, \dQuote{se}
\item Feature Types: \dQuote{integer}, \dQuote{numeric}
\item Required Packages: \CRANpkg{mlr3}, \CRANpkg{mlr3extralearners}, \CRANpkg{laGP}
}
}

\section{Parameters}{
\tabular{lllll}{
   Id \tab Type \tab Default \tab Levels \tab Range \cr
   start \tab integer \tab 6 \tab  \tab \eqn{[6, \infty)}{[6, Inf)} \cr
   end \tab integer \tab 50 \tab  \tab \eqn{[6, \infty)}{[6, Inf)} \cr
   d \tab untyped \tab NULL \tab  \tab - \cr
   g \tab untyped \tab 1/10000 \tab  \tab - \cr
   method \tab character \tab alc \tab alc, alcray, efi, mspe, nn \tab - \cr
   close \tab integer \tab - \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   numrays \tab integer \tab - \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   verb \tab numeric \tab 0 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
}
}

\examples{
<<<<<<< HEAD
\dontshow{if (learner_is_runnable("regr.laGP")) withAutoprint(\{ # examplesIf}
# Define the Learner
learner = lrn("regr.laGP")
=======
\dontshow{if (learner_is_runnable("regr.laGP")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Define the Learner (use a small neighborhood size for this task)
learner = lrn("regr.laGP", end = 15)
>>>>>>> a236489dddcdbc448a805b2daa1bd61325b2b347
print(learner)

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
\dontshow{\}) # examplesIf}
}
\references{
Gramacy, B. R (2016).
\dQuote{laGP: Large-Scale Spatial Modeling via Local Approximate Gaussian Processes in R.}
\emph{Journal of Statistical Software}, \bold{72}(1), 1--46.
\doi{10.18637/jss.v072.i01}.
}
\seealso{
\itemize{
\item \link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}.
\item \code{as.data.table(mlr_learners)} for a table of available \link[mlr3:Learner]{Learners} in the running session (depending on the loaded packages).
\item Chapter in the \href{https://mlr3book.mlr-org.com/}{mlr3book}: \url{https://mlr3book.mlr-org.com/basics.html#learners}
\item \CRANpkg{mlr3learners} for a selection of recommended learners.
\item \CRANpkg{mlr3cluster} for unsupervised clustering learners.
\item \CRANpkg{mlr3pipelines} to combine learners with pre- and postprocessing steps.
\item \CRANpkg{mlr3tuning} for tuning of hyperparameters, \CRANpkg{mlr3tuningspaces} for established default tuning spaces.
}
}
\author{
awinterstetter
}
\section{Super classes}{
\code{\link[mlr3:Learner]{mlr3::Learner}} -> \code{\link[mlr3:LearnerRegr]{mlr3::LearnerRegr}} -> \code{LearnerRegrLaGP}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LearnerRegrLaGP-new}{\code{LearnerRegrLaGP$new()}}
\item \href{#method-LearnerRegrLaGP-clone}{\code{LearnerRegrLaGP$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="base_learner"><a href='../../mlr3/html/Learner.html#method-Learner-base_learner'><code>mlr3::Learner$base_learner()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="configure"><a href='../../mlr3/html/Learner.html#method-Learner-configure'><code>mlr3::Learner$configure()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="encapsulate"><a href='../../mlr3/html/Learner.html#method-Learner-encapsulate'><code>mlr3::Learner$encapsulate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="format"><a href='../../mlr3/html/Learner.html#method-Learner-format'><code>mlr3::Learner$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="help"><a href='../../mlr3/html/Learner.html#method-Learner-help'><code>mlr3::Learner$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict"><a href='../../mlr3/html/Learner.html#method-Learner-predict'><code>mlr3::Learner$predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict_newdata"><a href='../../mlr3/html/Learner.html#method-Learner-predict_newdata'><code>mlr3::Learner$predict_newdata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="print"><a href='../../mlr3/html/Learner.html#method-Learner-print'><code>mlr3::Learner$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="reset"><a href='../../mlr3/html/Learner.html#method-Learner-reset'><code>mlr3::Learner$reset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="selected_features"><a href='../../mlr3/html/Learner.html#method-Learner-selected_features'><code>mlr3::Learner$selected_features()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="train"><a href='../../mlr3/html/Learner.html#method-Learner-train'><code>mlr3::Learner$train()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="LearnerRegr" data-id="predict_newdata_fast"><a href='../../mlr3/html/LearnerRegr.html#method-LearnerRegr-predict_newdata_fast'><code>mlr3::LearnerRegr$predict_newdata_fast()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerRegrLaGP-new"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerRegrLaGP-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerRegrLaGP$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerRegrLaGP-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerRegrLaGP-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerRegrLaGP$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
