% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner_LiblineaR_regr_liblinearl2l1svr.R
\name{mlr_learners_regr.liblinearl2l1svr}
\alias{mlr_learners_regr.liblinearl2l1svr}
\alias{LearnerRegrLiblineaRL2L1SVR}
\title{L2-Regularized L1-Loss Support Vector Regression Learner}
\description{
Calls \link[LiblineaR:LiblineaR]{LiblineaR::LiblineaR} from package \CRANpkg{LiblineaR}.
}
\details{
Calls \code{\link[LiblineaR:LiblineaR]{LiblineaR::LiblineaR()}} with \code{type = 13}.
}
\section{Dictionary}{
 This \link{Learner} can be instantiated via the
\link[mlr3misc:Dictionary]{dictionary} \link{mlr_learners} or with the associated
sugar function \code{\link[=lrn]{lrn()}}:\preformatted{mlr_learners$get("regr.liblinearl2l1svr")
lrn("regr.liblinearl2l1svr")
}
}

\section{Traits}{

\itemize{
\item Packages: LiblineaR
\item Predict Types: response
\item Feature Types: integer, numeric
\item Properties:
}
}

\section{Custom mlr3 defaults}{

\itemize{
\item \code{svr_eps}:
\itemize{
\item Actual default: \code{NULL}
\item Adjusted default: 0.001
\item Reason for change: \code{svr_eps} is type dependent and the "type" is handled
by the mlr3learner. The default value is set to th default of the respective
"type".
}
\item \code{epsilon}:
\itemize{
\item Actual default: 0.01
\item Adjusted default: Removed
\item Reason for change: For regr SVR learners paramter \code{svr_eps} overwrites
param \code{epsilon}.
}
}
}

\examples{
# stop example failing with warning if package not installed
learner = suppressWarnings(mlr3::lrn("regr.liblinearl2l1svr"))
print(learner)

# available parameters:
learner$param_set$ids()
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}:
\link[mlr3:mlr_learners]{mlr3::mlr_learners}
}
\author{
be-marc
}
\section{Super classes}{
\code{\link[mlr3:Learner]{mlr3::Learner}} -> \code{\link[mlr3:LearnerRegr]{mlr3::LearnerRegr}} -> \code{LearnerRegrLiblineaRL2L1SVR}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LearnerRegrLiblineaRL2L1SVR$new()}}
\item \href{#method-clone}{\code{LearnerRegrLiblineaRL2L1SVR$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="format">}\href{../../mlr3/html/Learner.html#method-format}{\code{mlr3::Learner$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="help">}\href{../../mlr3/html/Learner.html#method-help}{\code{mlr3::Learner$help()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict">}\href{../../mlr3/html/Learner.html#method-predict}{\code{mlr3::Learner$predict()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict_newdata">}\href{../../mlr3/html/Learner.html#method-predict_newdata}{\code{mlr3::Learner$predict_newdata()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="print">}\href{../../mlr3/html/Learner.html#method-print}{\code{mlr3::Learner$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="reset">}\href{../../mlr3/html/Learner.html#method-reset}{\code{mlr3::Learner$reset()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="train">}\href{../../mlr3/html/Learner.html#method-train}{\code{mlr3::Learner$train()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerRegrLiblineaRL2L1SVR$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerRegrLiblineaRL2L1SVR$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
