% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner_prioritylasso_regr_priority_lasso.R
\name{mlr_learners_regr.priority_lasso}
\alias{mlr_learners_regr.priority_lasso}
\alias{LearnerRegrPriorityLasso}
\title{Regression Priority Lasso Learner}
\description{
Patient outcome prediction based on multi-omics data taking practitioners’
preferences into account.
Calls \code{prioritylasso::prioritylasso()} from \CRANpkg{prioritylasso}.
}
\section{Initial parameter values}{

\itemize{
\item \code{family} is set to \code{"gaussian"} and cannot be changed
\item \code{type.measure} set to \code{"mse"} (cross-validation measure) and cannot be changed
}
}

\section{Dictionary}{

This \link[mlr3:Learner]{Learner} can be instantiated via \link[mlr3:mlr_sugar]{lrn()}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{lrn("regr.priority_lasso")
}\if{html}{\out{</div>}}
}

\section{Meta Information}{

\itemize{
\item Task type: \dQuote{regr}
\item Predict Types: \dQuote{response}
\item Feature Types: \dQuote{logical}, \dQuote{integer}, \dQuote{numeric}
\item Required Packages: \CRANpkg{mlr3}, \CRANpkg{prioritylasso}
}
}

\section{Parameters}{
\tabular{lllll}{
   Id \tab Type \tab Default \tab Levels \tab Range \cr
   blocks \tab untyped \tab - \tab  \tab - \cr
   max.coef \tab untyped \tab NULL \tab  \tab - \cr
   block1.penalization \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
   lambda.type \tab character \tab lambda.min \tab lambda.min, lambda.1se \tab - \cr
   standardize \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
   nfolds \tab integer \tab 10 \tab  \tab \eqn{[3, \infty)}{[3, Inf)} \cr
   foldid \tab untyped \tab NULL \tab  \tab - \cr
   cvoffset \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   cvoffsetnfolds \tab integer \tab 10 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   scale.y \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   return.x \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
   lambda \tab untyped \tab NULL \tab  \tab - \cr
   grouped \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
   trace.it \tab integer \tab 0 \tab  \tab \eqn{[0, 1]}{[0, 1]} \cr
   type.gaussian \tab character \tab - \tab covariance, naive \tab - \cr
   include.allintercepts \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   use.blocks \tab untyped \tab "all" \tab  \tab - \cr
}
}

\section{Scope and supported arguments}{

This learner intentionally exposes a focused subset of training and prediction arguments.
It is designed to work well out of the box, without requiring extensive parameter tuning.
Some arguments from \code{cv.glmnet()}, \code{glmnet()}, and \code{predict.prioritylasso()} are not included,
because they are not consistently supported or forwarded through the full train/predict path
(e.g. handling missing test data, or performing relaxed lasso fits).

Please open an issue if there is a need for supporting more learner parameters.
}

\examples{
\dontshow{if (learner_is_runnable("regr.priority_lasso")) withAutoprint(\{ # examplesIf}
# Simulate regression data
set.seed(1L)
x = matrix(rnorm(50L * 500L), nrow = 50L, ncol = 500L)
data = as.data.frame(x)
data$y = rnorm(50L)

# Define a Task
task = mlr3::as_task_regr(data, target = "y")

# Create train and test set
ids = partition(task)

# Define the Learner and set parameter values
learner = lrn("regr.priority_lasso",
  blocks = list(bp1 = 1:75, bp2 = 76:200, bp3 = 201:500),
  max.coef = c(Inf, 8, 5),
  block1.penalization = TRUE,
  lambda.type = "lambda.min",
  standardize = TRUE,
  nfolds = 5,
  cvoffset = FALSE)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

# Selected features
learner$selected_features()

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
\dontshow{\}) # examplesIf}
}
\references{
Simon K, Vindi J, Roman H, Tobias H, Anne-Laure B (2018).
\dQuote{Priority-Lasso: a simple hierarchical approach to the prediction of clinical outcome using multi-omics data.}
\emph{BMC Bioinformatics}, \bold{19}.
\doi{10.1186/s12859-018-2344-6}.
}
\seealso{
\itemize{
\item \link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}.
\item \code{as.data.table(mlr_learners)} for a table of available \link[mlr3:Learner]{Learners} in the running session (depending on the loaded packages).
\item Chapter in the \href{https://mlr3book.mlr-org.com/}{mlr3book}: \url{https://mlr3book.mlr-org.com/chapters/chapter2/data_and_basic_modeling.html#sec-learners}
\item \CRANpkg{mlr3learners} for a selection of recommended learners.
\item \CRANpkg{mlr3cluster} for unsupervised clustering learners.
\item \CRANpkg{mlr3pipelines} to combine learners with pre- and postprocessing steps.
\item \CRANpkg{mlr3tuning} for tuning of hyperparameters, \CRANpkg{mlr3tuningspaces} for established default tuning spaces.
}
}
\author{
HarutyunyanLiana

bblodfon
}
\section{Super classes}{
\code{\link[mlr3:Learner]{mlr3::Learner}} -> \code{\link[mlr3:LearnerRegr]{mlr3::LearnerRegr}} -> \code{LearnerRegrPriorityLasso}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LearnerRegrPriorityLasso-new}{\code{LearnerRegrPriorityLasso$new()}}
\item \href{#method-LearnerRegrPriorityLasso-selected_features}{\code{LearnerRegrPriorityLasso$selected_features()}}
\item \href{#method-LearnerRegrPriorityLasso-clone}{\code{LearnerRegrPriorityLasso$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="base_learner"><a href='../../mlr3/html/Learner.html#method-Learner-base_learner'><code>mlr3::Learner$base_learner()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="configure"><a href='../../mlr3/html/Learner.html#method-Learner-configure'><code>mlr3::Learner$configure()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="encapsulate"><a href='../../mlr3/html/Learner.html#method-Learner-encapsulate'><code>mlr3::Learner$encapsulate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="format"><a href='../../mlr3/html/Learner.html#method-Learner-format'><code>mlr3::Learner$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="help"><a href='../../mlr3/html/Learner.html#method-Learner-help'><code>mlr3::Learner$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict"><a href='../../mlr3/html/Learner.html#method-Learner-predict'><code>mlr3::Learner$predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict_newdata"><a href='../../mlr3/html/Learner.html#method-Learner-predict_newdata'><code>mlr3::Learner$predict_newdata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="print"><a href='../../mlr3/html/Learner.html#method-Learner-print'><code>mlr3::Learner$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="reset"><a href='../../mlr3/html/Learner.html#method-Learner-reset'><code>mlr3::Learner$reset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="train"><a href='../../mlr3/html/Learner.html#method-Learner-train'><code>mlr3::Learner$train()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="LearnerRegr" data-id="predict_newdata_fast"><a href='../../mlr3/html/LearnerRegr.html#method-LearnerRegr-predict_newdata_fast'><code>mlr3::LearnerRegr$predict_newdata_fast()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerRegrPriorityLasso-new"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerRegrPriorityLasso-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerRegrPriorityLasso$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerRegrPriorityLasso-selected_features"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerRegrPriorityLasso-selected_features}{}}}
\subsection{Method \code{selected_features()}}{
Selected features when coef is positive
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerRegrPriorityLasso$selected_features()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character()}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerRegrPriorityLasso-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerRegrPriorityLasso-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerRegrPriorityLasso$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
