% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner_tabnet_regr_tabnet.R
\name{mlr_learners_regr.tabnet}
\alias{mlr_learners_regr.tabnet}
\alias{LearnerRegrTabNet}
\title{Regression TabNet Learner}
\description{
Regression TabNet Learner

Regression TabNet Learner
}
\section{Dictionary}{
 This \link{Learner} can be instantiated via the
\link[mlr3misc:Dictionary]{dictionary} \link{mlr_learners} or with the associated
sugar function \code{\link[=lrn]{lrn()}}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{mlr_learners$get("regr.tabnet")
lrn("regr.tabnet")
}\if{html}{\out{</div>}}
}

\section{Meta Information}{

\verb{r paste(mlr3misc::rd_info(mlr3::lrn("regr.tabnet")), collapse = "\\n")}
}

\section{Parameters}{

\verb{r paste(mlr3misc::rd_info(mlr3::lrn("regr.tabnet")$param_set), collapse = "\\n")}
}

\examples{
\dontrun{
library(mlr3)
library(mlr3torch)

task = tsk("boston_housing")

# Creating a learner & training on example task
lrn = lrn("regr.tabnet")

lrn$param_set$values$epochs = 10
lrn$train(task)

# Predict on training data, get RMSE
predictions = lrn$predict(task)
predictions$score(msr("regr.rmse"))
}
}
\references{
\verb{r format_bib("arik2021tabnet")}
}
\seealso{
\itemize{
\item \link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}.
\item \code{as.data.table(mlr_learners)} for a table of available \link[=Learner]{Learners} in the running session (depending on the loaded packages).
\item Chapter in the \href{https://mlr3book.mlr-org.com/}{mlr3book}: \url{https://mlr3book.mlr-org.com/basics.html#learners}
\item \CRANpkg{mlr3learners} for a selection of recommended learners.
\item \CRANpkg{mlr3cluster} for unsupervised clustering learners.
\item \CRANpkg{mlr3pipelines} to combine learners with pre- and postprocessing steps.
\item \CRANpkg{mlr3tuning} for tuning of hyperparameters, \CRANpkg{mlr3tuningspaces} for established default tuning spaces.
}
}
\author{
Lukas Burk
}
\section{Super classes}{
\code{\link[mlr3:Learner]{mlr3::Learner}} -> \code{\link[mlr3:LearnerRegr]{mlr3::LearnerRegr}} -> \code{LearnerRegrTabnet}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LearnerRegrTabnet-new}{\code{LearnerRegrTabNet$new()}}
\item \href{#method-LearnerRegrTabnet-importance}{\code{LearnerRegrTabNet$importance()}}
\item \href{#method-LearnerRegrTabnet-clone}{\code{LearnerRegrTabNet$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="base_learner"><a href='../../mlr3/html/Learner.html#method-Learner-base_learner'><code>mlr3::Learner$base_learner()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="format"><a href='../../mlr3/html/Learner.html#method-Learner-format'><code>mlr3::Learner$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="help"><a href='../../mlr3/html/Learner.html#method-Learner-help'><code>mlr3::Learner$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict"><a href='../../mlr3/html/Learner.html#method-Learner-predict'><code>mlr3::Learner$predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict_newdata"><a href='../../mlr3/html/Learner.html#method-Learner-predict_newdata'><code>mlr3::Learner$predict_newdata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="print"><a href='../../mlr3/html/Learner.html#method-Learner-print'><code>mlr3::Learner$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="reset"><a href='../../mlr3/html/Learner.html#method-Learner-reset'><code>mlr3::Learner$reset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="train"><a href='../../mlr3/html/Learner.html#method-Learner-train'><code>mlr3::Learner$train()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerRegrTabnet-new"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerRegrTabnet-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerRegrTabNet$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerRegrTabnet-importance"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerRegrTabnet-importance}{}}}
\subsection{Method \code{importance()}}{
The importance scores are extracted from the slot \code{.$model$fit$importances}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerRegrTabNet$importance()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Named \code{numeric()}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerRegrTabnet-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerRegrTabnet-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerRegrTabNet$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
