% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner_partykit_surv_cforest.R
\name{mlr_learners_surv.cforest}
\alias{mlr_learners_surv.cforest}
\alias{LearnerSurvCForest}
\title{Survival Conditional Random Forest Learner}
\description{
A random forest based on conditional inference trees (\link[partykit:ctree]{ctree}).
Calls \code{\link[partykit:cforest]{partykit::cforest()}} from \CRANpkg{partykit}.
}
\section{Dictionary}{

This \link{Learner} can be instantiated via the \link[mlr3misc:Dictionary]{dictionary} \link{mlr_learners} or with the associated sugar function \code{\link[=lrn]{lrn()}}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{mlr_learners$get("surv.cforest")
lrn("surv.cforest")
}\if{html}{\out{</div>}}
}

\section{Meta Information}{

\itemize{
\item Task type: \dQuote{surv}
\item Predict Types: \dQuote{crank}, \dQuote{distr}
\item Feature Types: \dQuote{integer}, \dQuote{numeric}, \dQuote{factor}, \dQuote{ordered}
\item Required Packages: \CRANpkg{mlr3}, \CRANpkg{mlr3proba}, \CRANpkg{mlr3extralearners}, \CRANpkg{partykit}, \CRANpkg{sandwich}, \CRANpkg{coin}
}
}

\section{Parameters}{
\tabular{lllll}{
   Id \tab Type \tab Default \tab Levels \tab Range \cr
   ntree \tab integer \tab 500 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   replace \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   fraction \tab numeric \tab 0.632 \tab  \tab \eqn{[0, 1]}{[0, 1]} \cr
   mtry \tab integer \tab - \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   mtryratio \tab numeric \tab - \tab  \tab \eqn{[0, 1]}{[0, 1]} \cr
   applyfun \tab untyped \tab - \tab  \tab - \cr
   cores \tab integer \tab NULL \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   trace \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   offset \tab untyped \tab - \tab  \tab - \cr
   cluster \tab untyped \tab - \tab  \tab - \cr
   na.action \tab untyped \tab ::     , stats  , na.pass \tab  \tab - \cr
   scores \tab untyped \tab - \tab  \tab - \cr
   teststat \tab character \tab quadratic \tab quadratic, maximum \tab - \cr
   splitstat \tab character \tab quadratic \tab quadratic, maximum \tab - \cr
   splittest \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   testtype \tab character \tab Univariate \tab Bonferroni, MonteCarlo, Univariate, Teststatistic \tab - \cr
   nmax \tab untyped \tab - \tab  \tab - \cr
   alpha \tab numeric \tab 0.05 \tab  \tab \eqn{[0, 1]}{[0, 1]} \cr
   mincriterion \tab numeric \tab 0.95 \tab  \tab \eqn{[0, 1]}{[0, 1]} \cr
   logmincriterion \tab numeric \tab -0.05129329 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   minsplit \tab integer \tab 20 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   minbucket \tab integer \tab 7 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   minprob \tab numeric \tab 0.01 \tab  \tab \eqn{[0, 1]}{[0, 1]} \cr
   stump \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   lookahead \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   MIA \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   nresample \tab integer \tab 9999 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   tol \tab numeric \tab 1.490116e-08 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   maxsurrogate \tab integer \tab 0 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   numsurrogate \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   maxdepth \tab integer \tab Inf \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   multiway \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   splittry \tab integer \tab 2 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   intersplit \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   majority \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   caseweights \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
   saveinfo \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   update \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   splitflavour \tab character \tab ctree \tab ctree, exhaustive \tab - \cr
   maxvar \tab integer \tab - \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   OOB \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   simplify \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
   scale \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
   maxpts \tab integer \tab 25000 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   abseps \tab numeric \tab 0.001 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   releps \tab numeric \tab 0 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
}
}

\section{Initial parameter values}{

\itemize{
\item \code{mtry}:
\itemize{
\item This hyperparameter can alternatively be set via the added hyperparameter \code{mtryratio}
as \code{mtry = max(ceiling(mtryratio * n_features), 1)}.
Note that \code{mtry} and \code{mtryratio} are mutually exclusive.
}
}
}

\examples{
learner = mlr3::lrn("surv.cforest")
print(learner)

# available parameters:
learner$param_set$ids()
}
\references{
Hothorn T, Zeileis A (2015).
\dQuote{partykit: A Modular Toolkit for Recursive Partytioning in R.}
\emph{Journal of Machine Learning Research}, \bold{16}(118), 3905-3909.
\url{http://jmlr.org/papers/v16/hothorn15a.html}.

Hothorn T, Hornik K, Zeileis A (2006).
\dQuote{Unbiased Recursive Partitioning: A Conditional Inference Framework.}
\emph{Journal of Computational and Graphical Statistics}, \bold{15}(3), 651--674.
\doi{10.1198/106186006x133933}, \url{https://doi.org/10.1198/106186006x133933}.
}
\seealso{
\itemize{
\item \link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}.
\item \code{as.data.table(mlr_learners)} for a table of available \link[=Learner]{Learners} in the running session (depending on the loaded packages).
\item Chapter in the \href{https://mlr3book.mlr-org.com/}{mlr3book}: \url{https://mlr3book.mlr-org.com/basics.html#learners}
\item \CRANpkg{mlr3learners} for a selection of recommended learners.
\item \CRANpkg{mlr3cluster} for unsupervised clustering learners.
\item \CRANpkg{mlr3pipelines} to combine learners with pre- and postprocessing steps.
\item \CRANpkg{mlr3tuning} for tuning of hyperparameters, \CRANpkg{mlr3tuningspaces} for established default tuning spaces.
}
}
\author{
RaphaelS1
}
\section{Super classes}{
\code{\link[mlr3:Learner]{mlr3::Learner}} -> \code{\link[mlr3proba:LearnerSurv]{mlr3proba::LearnerSurv}} -> \code{LearnerSurvCForest}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LearnerSurvCForest-new}{\code{LearnerSurvCForest$new()}}
\item \href{#method-LearnerSurvCForest-clone}{\code{LearnerSurvCForest$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="base_learner"><a href='../../mlr3/html/Learner.html#method-Learner-base_learner'><code>mlr3::Learner$base_learner()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="format"><a href='../../mlr3/html/Learner.html#method-Learner-format'><code>mlr3::Learner$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="help"><a href='../../mlr3/html/Learner.html#method-Learner-help'><code>mlr3::Learner$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict"><a href='../../mlr3/html/Learner.html#method-Learner-predict'><code>mlr3::Learner$predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict_newdata"><a href='../../mlr3/html/Learner.html#method-Learner-predict_newdata'><code>mlr3::Learner$predict_newdata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="print"><a href='../../mlr3/html/Learner.html#method-Learner-print'><code>mlr3::Learner$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="reset"><a href='../../mlr3/html/Learner.html#method-Learner-reset'><code>mlr3::Learner$reset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="train"><a href='../../mlr3/html/Learner.html#method-Learner-train'><code>mlr3::Learner$train()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerSurvCForest-new"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerSurvCForest-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerSurvCForest$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerSurvCForest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerSurvCForest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerSurvCForest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
