% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner_survivalmodels_surv_dnnsurv.R
\name{mlr_learners_surv.dnnsurv}
\alias{mlr_learners_surv.dnnsurv}
\alias{LearnerSurvDNNSurv}
\title{Survival DNNSurv Learner}
\description{
Calls \link[RaphaelS1/survivalmodels:dnnsurv]{RaphaelS1/survivalmodels::dnnsurv} from package \CRANpkg{RaphaelS1/survivalmodels}.
}
\details{
Custom nets can be used in this learner either using the
\link[survivalmodels:build_keras_net]{survivalmodels::build_keras_net} utility function or using \CRANpkg{keras}.
The number of output channels should be of length \code{1} and number of input channels is
the number of features plus number of cuts.
}
\section{Dictionary}{
 This \link{Learner} can be instantiated via the
\link[mlr3misc:Dictionary]{dictionary} \link{mlr_learners} or with the associated
sugar function \code{\link[=lrn]{lrn()}}:\preformatted{mlr_learners$get("surv.dnnsurv")
lrn("surv.dnnsurv")
}
}

\section{Traits}{

\itemize{
\item Packages: RaphaelS1/survivalmodels, keras, pseudo, tensorflow, distr6
\item Predict Types: crank, distr
\item Feature Types: integer, numeric
\item Properties:
}
}

\section{Custom mlr3 defaults}{

\itemize{
\item \code{verbose}:
\itemize{
\item Actual default: \code{1L}
\item Adjusted default: \code{0L}
\item Reason for change: Prevents plotting.
}
}
}

\examples{
# stop example failing with warning if package not installed
learner = suppressWarnings(mlr3::lrn("surv.dnnsurv"))
print(learner)

# available parameters:
learner$param_set$ids()
}
\references{
Zhao, L., & Feng, D. (2020).
DNNSurv: Deep Neural Networks for Survival Analysis Using Pseudo Values.
https://arxiv.org/abs/1908.02337
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}:
\link[mlr3:mlr_learners]{mlr3::mlr_learners}
}
\author{
RaphaelS1
}
\section{Super classes}{
\code{\link[mlr3:Learner]{mlr3::Learner}} -> \code{\link[mlr3proba:LearnerSurv]{mlr3proba::LearnerSurv}} -> \code{LearnerSurvDNNSurv}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LearnerSurvDNNSurv$new()}}
\item \href{#method-clone}{\code{LearnerSurvDNNSurv$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="format">}\href{../../mlr3/html/Learner.html#method-format}{\code{mlr3::Learner$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="help">}\href{../../mlr3/html/Learner.html#method-help}{\code{mlr3::Learner$help()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict">}\href{../../mlr3/html/Learner.html#method-predict}{\code{mlr3::Learner$predict()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict_newdata">}\href{../../mlr3/html/Learner.html#method-predict_newdata}{\code{mlr3::Learner$predict_newdata()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="print">}\href{../../mlr3/html/Learner.html#method-print}{\code{mlr3::Learner$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="reset">}\href{../../mlr3/html/Learner.html#method-reset}{\code{mlr3::Learner$reset()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="train">}\href{../../mlr3/html/Learner.html#method-train}{\code{mlr3::Learner$train()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerSurvDNNSurv$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerSurvDNNSurv$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
