% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner_flexsurv_surv_flexsurvreg.R
\name{mlr_learners_surv.flexreg}
\alias{mlr_learners_surv.flexreg}
\alias{LearnerSurvFlexReg}
\title{Survival Flexible Parametric Learner}
\description{
Flexible parametric learner.
Calls \code{\link[flexsurv:flexsurvreg]{flexsurv::flexsurvreg()}} from \CRANpkg{flexsurv}.
}
\section{Dictionary}{

This \link[mlr3:Learner]{Learner} can be instantiated via \link[mlr3:mlr_sugar]{lrn()}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{lrn("surv.flexreg")
}\if{html}{\out{</div>}}
}

\section{Meta Information}{

\itemize{
\item Task type: \dQuote{surv}
\item Predict Types: \dQuote{crank}, \dQuote{distr}, \dQuote{lp}, \dQuote{response}
\item Feature Types: \dQuote{logical}, \dQuote{integer}, \dQuote{numeric}, \dQuote{factor}
\item Required Packages: \CRANpkg{mlr3}, \CRANpkg{mlr3proba}, \CRANpkg{mlr3extralearners}, \CRANpkg{flexsurv}, \CRANpkg{survival}
}
}

\section{Parameters}{
\tabular{lllll}{
   Id \tab Type \tab Default \tab Levels \tab Range \cr
   formula \tab untyped \tab - \tab  \tab - \cr
   anc \tab untyped \tab - \tab  \tab - \cr
   bhazard \tab untyped \tab - \tab  \tab - \cr
   rtrunc \tab untyped \tab - \tab  \tab - \cr
   dist \tab character \tab - \tab gengamma, gengamma.orig, genf, genf.orig, weibull, weibullph, gamma, exp, exponential, llogis, \link{...} \tab - \cr
   inits \tab untyped \tab - \tab  \tab - \cr
   fixedpars \tab untyped \tab - \tab  \tab - \cr
   cl \tab numeric \tab 0.95 \tab  \tab \eqn{[0, 1]}{[0, 1]} \cr
   hessian \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
   hess.control \tab untyped \tab - \tab  \tab - \cr
   maxiter \tab integer \tab 30 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   rel.tolerance \tab numeric \tab 1e-09 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   toler.chol \tab numeric \tab 1e-10 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   debug \tab integer \tab 0 \tab  \tab \eqn{[0, 1]}{[0, 1]} \cr
   outer.max \tab integer \tab 10 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   times \tab untyped \tab - \tab  \tab - \cr
}
}

\section{Prediction types}{

This learner returns three prediction types:
\enumerate{
\item \code{lp}: a vector containing the linear predictors (relative risk scores),
where each score corresponds to a specific test observation.
Calculated using \code{predict.flexsurvreg(type = "lp", ...)}.
The interpretation whether higher lp corresponds to higher risk or higher
survival, depends on the distribution choice (see \code{dist} parameter documentation).
\item \code{distr}: a survival matrix in two dimensions, where observations are
represented in rows and time points in columns.
Calculated using \code{predict.flexsurvreg(type = "survival", ...)}.
\item \code{crank}: same as \code{lp}.
\item \code{response}: mean survival time calculated using \code{predict.flexsurvreg(type = "response", ...)}.
}
}

\examples{
\dontshow{if (learner_is_runnable("surv.flexreg")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(survival)

# Define the task, split to train/test set
task = tsk("lung")
set.seed(42)
part = partition(task)

# Define the learner
learner = lrn("surv.flexreg",
  formula = survival::Surv(time, status) ~ age + ph.karno + sex,
  anc = list(shape = ~ sex), dist = "weibull")

# Train the learner on the training ids
learner$train(task, part$train)
print(learner$model)

# Make predictions for the test rows
predictions = learner$predict(task, part$test)
print(predictions)

# Score the predictions
predictions$score()
\dontshow{\}) # examplesIf}
}
\references{
Jackson, H. C (2016).
\dQuote{flexsurv: A Platform for Parametric Survival Modeling in R.}
\emph{Journal of Statistical Software}, \bold{70}(8), 1--33.
\doi{10.18637/JSS.V070.I08}, \url{https://www.jstatsoft.org/index.php/jss/article/view/v070i08}.
}
\seealso{
\itemize{
\item \link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}.
\item \code{as.data.table(mlr_learners)} for a table of available \link[mlr3:Learner]{Learners} in the running session (depending on the loaded packages).
\item Chapter in the \href{https://mlr3book.mlr-org.com/}{mlr3book}: \url{https://mlr3book.mlr-org.com/basics.html#learners}
\item \CRANpkg{mlr3learners} for a selection of recommended learners.
\item \CRANpkg{mlr3cluster} for unsupervised clustering learners.
\item \CRANpkg{mlr3pipelines} to combine learners with pre- and postprocessing steps.
\item \CRANpkg{mlr3tuning} for tuning of hyperparameters, \CRANpkg{mlr3tuningspaces} for established default tuning spaces.
}
}
\author{
bblodfon
}
\section{Super classes}{
\code{\link[mlr3:Learner]{mlr3::Learner}} -> \code{\link[mlr3proba:LearnerSurv]{mlr3proba::LearnerSurv}} -> \code{LearnerSurvFlexReg}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LearnerSurvFlexReg-new}{\code{LearnerSurvFlexReg$new()}}
\item \href{#method-LearnerSurvFlexReg-clone}{\code{LearnerSurvFlexReg$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="base_learner"><a href='../../mlr3/html/Learner.html#method-Learner-base_learner'><code>mlr3::Learner$base_learner()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="configure"><a href='../../mlr3/html/Learner.html#method-Learner-configure'><code>mlr3::Learner$configure()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="encapsulate"><a href='../../mlr3/html/Learner.html#method-Learner-encapsulate'><code>mlr3::Learner$encapsulate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="format"><a href='../../mlr3/html/Learner.html#method-Learner-format'><code>mlr3::Learner$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="help"><a href='../../mlr3/html/Learner.html#method-Learner-help'><code>mlr3::Learner$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict"><a href='../../mlr3/html/Learner.html#method-Learner-predict'><code>mlr3::Learner$predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict_newdata"><a href='../../mlr3/html/Learner.html#method-Learner-predict_newdata'><code>mlr3::Learner$predict_newdata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="print"><a href='../../mlr3/html/Learner.html#method-Learner-print'><code>mlr3::Learner$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="reset"><a href='../../mlr3/html/Learner.html#method-Learner-reset'><code>mlr3::Learner$reset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="selected_features"><a href='../../mlr3/html/Learner.html#method-Learner-selected_features'><code>mlr3::Learner$selected_features()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="train"><a href='../../mlr3/html/Learner.html#method-Learner-train'><code>mlr3::Learner$train()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerSurvFlexReg-new"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerSurvFlexReg-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerSurvFlexReg$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerSurvFlexReg-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerSurvFlexReg-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerSurvFlexReg$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
