% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner_mgcv_surv_gam_cox.R
\name{mlr_learners_surv.gam.cox}
\alias{mlr_learners_surv.gam.cox}
\alias{LearnerSurvGamCox}
\title{Cox Proportional Hazards Generalized Additive Learner}
\description{
Cox proportional hazards model fitted via generalized additive modeling.
Internally calls \code{mgcv::gam()} from the \CRANpkg{mgcv} package with
\code{family = mgcv::cox.ph()}. The censoring indicator is passed via the
\code{weights} argument.
}
\section{Prediction types}{

Three types of prediction are returned for this learner:
\enumerate{
\item \code{lp}: a vector of linear predictors (relative risk scores), one per
observation. Calls \code{mgcv::predict.gam()} with \code{type = "link"}.
\item \code{crank}: same as \code{lp}.
\item \code{distr}: a survival matrix in two dimensions, where observations are
represented in rows and time points in columns.
By default, the Breslow estimator is used via \code{mlr3proba::breslow()}.
}
}

\section{Dictionary}{

This \link[mlr3:Learner]{Learner} can be instantiated via \link[mlr3:mlr_sugar]{lrn()}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{lrn("surv.gam.cox")
}\if{html}{\out{</div>}}
}

\section{Meta Information}{

\itemize{
\item Task type: \dQuote{surv}
\item Predict Types: \dQuote{crank}, \dQuote{distr}, \dQuote{lp}
\item Feature Types: \dQuote{logical}, \dQuote{integer}, \dQuote{numeric}, \dQuote{factor}
\item Required Packages: \CRANpkg{mlr3}, \CRANpkg{mlr3proba}, \CRANpkg{mlr3extralearners}, \CRANpkg{mgcv}
}
}

\section{Parameters}{
\tabular{lllll}{
   Id \tab Type \tab Default \tab Levels \tab Range \cr
   formula \tab untyped \tab - \tab  \tab - \cr
   method \tab character \tab GCV.Cp \tab GCV.Cp, GACV.Cp, REML, P-REML, ML, P-ML \tab - \cr
   optimizer \tab untyped \tab c("outer", "newton") \tab  \tab - \cr
   scale \tab numeric \tab 0 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   select \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   knots \tab untyped \tab NULL \tab  \tab - \cr
   sp \tab untyped \tab NULL \tab  \tab - \cr
   min.sp \tab untyped \tab NULL \tab  \tab - \cr
   H \tab untyped \tab NULL \tab  \tab - \cr
   gamma \tab numeric \tab 1 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   paraPen \tab untyped \tab NULL \tab  \tab - \cr
   G \tab untyped \tab NULL \tab  \tab - \cr
   in.out \tab untyped \tab NULL \tab  \tab - \cr
   drop.unused.levels \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
   drop.intercept \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   nei \tab untyped \tab - \tab  \tab - \cr
   nthreads \tab integer \tab 1 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   ncv.threads \tab integer \tab 1 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   irls.reg \tab numeric \tab 0 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   epsilon \tab numeric \tab 1e-07 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   maxit \tab integer \tab 200 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   mgcv.tol \tab numeric \tab 1e-07 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   mgcv.half \tab integer \tab 15 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   trace \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   rank.tol \tab numeric \tab 1.490116e-08 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   nlm \tab untyped \tab list() \tab  \tab - \cr
   optim \tab untyped \tab list() \tab  \tab - \cr
   newton \tab untyped \tab list() \tab  \tab - \cr
   idLinksBases \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
   scalePenalty \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
   efs.lspmax \tab integer \tab 15 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   efs.tol \tab numeric \tab 0.1 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   scale.est \tab character \tab fletcher \tab fletcher, pearson, deviance \tab - \cr
   edge.correct \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   block.size \tab integer \tab 1000 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   unconditional \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
}
}

\section{Formula}{

A gam formula specific to the task at hand is required for the \code{formula}
parameter (see example and \code{?mgcv::formula.gam}). Beware, if no formula is provided, a fallback formula is
used that will make the gam behave like a glm (this behavior is required
for the unit tests). Only features specified in the formula will be used,
superseding columns with col_roles "feature" in the task.
}

\section{Offset}{

If a \code{Task} contains a column with the \code{offset} role, it is automatically
incorporated during training via the \code{offset} argument in \code{mgcv::gam()}.
No offset is applied during prediction for this learner.
}

\examples{
\dontshow{if (learner_is_runnable("surv.gam.cox")) withAutoprint(\{ # examplesIf}
# simple example
t = tsk("lung")
l = lrn("surv.gam.cox")
l$param_set$set_values(formula = time ~ s(age, k = 5) + ph.karno + sex)
l$train(t)
l$model
\dontshow{\}) # examplesIf}
}
\references{
Wood, Simon (2012).
\dQuote{mgcv: Mixed GAM Computation Vehicle with GCV/AIC/REML smoothness estimation.}

Wood, N. S, Pya, Natalya, Safken, Benjamin (2016).
\dQuote{Smoothing Parameter and Model Selection for General Smooth Models.}
\emph{Journal of the American Statistical Association}, \bold{111}(516), 1548--1563.
\doi{10.1080/01621459.2016.1180986}.
}
\author{
bblodfon
}
\section{Super classes}{
\code{\link[mlr3:Learner]{mlr3::Learner}} -> \code{\link[mlr3proba:LearnerSurv]{mlr3proba::LearnerSurv}} -> \code{LearnerSurvGamCox}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LearnerSurvGamCox-new}{\code{LearnerSurvGamCox$new()}}
\item \href{#method-LearnerSurvGamCox-clone}{\code{LearnerSurvGamCox$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="base_learner"><a href='../../mlr3/html/Learner.html#method-Learner-base_learner'><code>mlr3::Learner$base_learner()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="configure"><a href='../../mlr3/html/Learner.html#method-Learner-configure'><code>mlr3::Learner$configure()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="encapsulate"><a href='../../mlr3/html/Learner.html#method-Learner-encapsulate'><code>mlr3::Learner$encapsulate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="format"><a href='../../mlr3/html/Learner.html#method-Learner-format'><code>mlr3::Learner$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="help"><a href='../../mlr3/html/Learner.html#method-Learner-help'><code>mlr3::Learner$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict"><a href='../../mlr3/html/Learner.html#method-Learner-predict'><code>mlr3::Learner$predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict_newdata"><a href='../../mlr3/html/Learner.html#method-Learner-predict_newdata'><code>mlr3::Learner$predict_newdata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="print"><a href='../../mlr3/html/Learner.html#method-Learner-print'><code>mlr3::Learner$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="reset"><a href='../../mlr3/html/Learner.html#method-Learner-reset'><code>mlr3::Learner$reset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="selected_features"><a href='../../mlr3/html/Learner.html#method-Learner-selected_features'><code>mlr3::Learner$selected_features()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="train"><a href='../../mlr3/html/Learner.html#method-Learner-train'><code>mlr3::Learner$train()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerSurvGamCox-new"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerSurvGamCox-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerSurvGamCox$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerSurvGamCox-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerSurvGamCox-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerSurvGamCox$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
