% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner_surv_survival_nelson.R
\name{mlr_learners_surv.nelson}
\alias{mlr_learners_surv.nelson}
\alias{LearnerSurvNelson}
\title{Survival Nelson-Aalen Estimator Learner}
\description{
A \link[mlr3proba:LearnerSurv]{mlr3proba::LearnerSurv} implementing \code{survfit} from package
\CRANpkg{survival}.
Calls \code{\link[survival:survfit]{survival::survfit()}}.
}
\details{
\itemize{
\item \code{distr} is predicted by estimating the cumulative hazard function from \code{survival::survfit()}.
\item \code{crank} is predicted as the expectation of \code{distr}
}
}
\section{Dictionary}{
 This \link{Learner} can be instantiated via the
\link[mlr3misc:Dictionary]{dictionary} \link{mlr_learners} or with the associated
sugar function \code{\link[=lrn]{lrn()}}:\preformatted{mlr_learners$get("surv.nelson")
lrn("surv.nelson")
}
}

\section{Traits}{

\itemize{
\item Packages: survival, pracma
\item Predict Types: crank, distr
\item Feature Types: logical, integer, numeric, character, factor, ordered
\item Properties: missings
}
}

\examples{
if (mlr3misc::require_namespaces(c("survival", "pracma"))) {
  learner = mlr3::lrn("surv.nelson")
  print(learner)

  # available parameters:
  learner$param_set$ids()
}
}
\references{
Nelson W (1969).
Hazard Plotting for Incomplete Failure Data.
Journal of Quality Technology, 1(1), 27–52. doi: 10.1080/00224065.1969.11980344.

Nelson, W. (1972).
Theory and Applications of Hazard Plotting for Censored Failure Data.
Technometrics, 14(4), 945–966.

Aalen, O. (1978).
Nonparametric Inference for a Family of Counting Processes.
The Annals of Statistics, 6(4), 701–726.
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}:
\link[mlr3:mlr_learners]{mlr3::mlr_learners}
}
\author{
RaphaelS1
}
\section{Super classes}{
\code{\link[mlr3:Learner]{mlr3::Learner}} -> \code{\link[mlr3proba:LearnerSurv]{mlr3proba::LearnerSurv}} -> \code{LearnerSurvNelson}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LearnerSurvNelson$new()}}
\item \href{#method-clone}{\code{LearnerSurvNelson$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="format">}\href{../../mlr3/html/Learner.html#method-format}{\code{mlr3::Learner$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="help">}\href{../../mlr3/html/Learner.html#method-help}{\code{mlr3::Learner$help()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict">}\href{../../mlr3/html/Learner.html#method-predict}{\code{mlr3::Learner$predict()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict_newdata">}\href{../../mlr3/html/Learner.html#method-predict_newdata}{\code{mlr3::Learner$predict_newdata()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="print">}\href{../../mlr3/html/Learner.html#method-print}{\code{mlr3::Learner$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="reset">}\href{../../mlr3/html/Learner.html#method-reset}{\code{mlr3::Learner$reset()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="train">}\href{../../mlr3/html/Learner.html#method-train}{\code{mlr3::Learner$train()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerSurvNelson$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerSurvNelson$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
