% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner_survivalmodels_surv_parametric.R
\name{mlr_learners_surv.parametric}
\alias{mlr_learners_surv.parametric}
\alias{LearnerSurvParametric}
\title{Survival Fully Parametric Learner}
\description{
Parametric survival model.
Calls \code{\link[survivalmodels:parametric]{survivalmodels::parametric()}} from package 'survivalmodels'.
}
\details{
This learner allows you to choose a \strong{distribution} and a \strong{model form} to compose
a predicted survival probability distribution \eqn{S(t)}.

The predict method is implemented in \code{\link[survivalmodels:predict.parametric]{survivalmodels::predict.parametric()}}.
Our implementation is more efficient for composition to distributions than
\code{\link[survival:predict.survreg]{survival::predict.survreg()}}.

The available model forms with their respective survival functions, are as follows:
\itemize{
\item Accelerated Failure Time (\code{aft}) \deqn{S(t) = S_0(\frac{t}{exp(lp)})}{S(t) = S0(t/exp(lp))}
\item Proportional Hazards (\code{ph}) \deqn{S(t) = S_0(t)^{exp(lp)}}{S(t) = S0(t)^exp(lp)}
\item Proportional Odds (\code{po}) \deqn{S(t) =
\frac{S_0(t)}{exp(-lp) + (1-exp(-lp)) S_0(t)}}{S(t) = S0(t) / [exp(-lp) + S0(t) (1-exp(-lp))]}
\item Tobit (\code{tobit}) \deqn{S(t) = 1 - \Phi((t - lp)/s)}
}

where \eqn{S_0}{S0} is the estimated baseline survival distribution (in
this case with a given parametric form), \eqn{lp} is the predicted linear
predictor, \eqn{\Phi} is the cdf of a \eqn{N(0, 1)} distribution, and \eqn{s} is
the fitted scale parameter.

Whilst any combination of distribution and model form is possible, this does
not mean it will necessarily create a sensible or interpretable prediction.
The following combinations are 'sensible' (we note that ones mostly used in
the literature):
\itemize{
\item dist = "gaussian"; form = "tobit";
\item dist = "weibull"; form = "ph"; (fairly used)
\item dist = "exponential"; form = "ph";
\item dist = "weibull"; form = "aft"; (fairly used, \strong{default option})
\item dist = "exponential"; form = "aft";
\item dist = "loglogistic"; form = "aft"; (fairly used)
\item dist = "lognormal"; form = "aft";
\item dist = "loglogistic"; form = "po";
}
}
\section{Prediction types}{

This learner returns three prediction types:
\enumerate{
\item \code{lp}: a vector of linear predictors (relative risk scores), one per test
observation.
\code{lp} is predicted using the formula \eqn{lp = X\beta} where \eqn{X} are the
variables in the test data set and \eqn{\beta} are the fitted coefficients.
\item \code{crank}: same as \code{lp}.
\item \code{distr}: a survival matrix in two dimensions, where observations are
represented in rows and time points in columns.
The distribution \code{distr} is composed using the \code{lp} predictions and specifying
a model form in the \code{form} hyper-parameter, see Details.
The survival matrix uses the unique time points from the training set.
The parameter \code{ntime} allows to adjust the granularity of these time points
to any number (e.g. \code{150}).
}
}

\section{Dictionary}{

This \link[mlr3:Learner]{Learner} can be instantiated via \link[mlr3:mlr_sugar]{lrn()}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{lrn("surv.parametric")
}\if{html}{\out{</div>}}
}

\section{Meta Information}{

\itemize{
\item Task type: \dQuote{surv}
\item Predict Types: \dQuote{crank}, \dQuote{distr}, \dQuote{lp}
\item Feature Types: \dQuote{logical}, \dQuote{integer}, \dQuote{numeric}, \dQuote{factor}
\item Required Packages: \CRANpkg{mlr3}, \CRANpkg{mlr3proba}, \CRANpkg{mlr3extralearners}, \CRANpkg{survival}
}
}

\section{Parameters}{
\tabular{lllll}{
   Id \tab Type \tab Default \tab Levels \tab Range \cr
   form \tab character \tab aft \tab aft, ph, po, tobit \tab - \cr
   na.action \tab untyped \tab - \tab  \tab - \cr
   dist \tab character \tab weibull \tab weibull, exponential, gaussian, lognormal, loglogistic \tab - \cr
   parms \tab untyped \tab - \tab  \tab - \cr
   init \tab untyped \tab - \tab  \tab - \cr
   scale \tab numeric \tab 0 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   maxiter \tab integer \tab 30 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   rel.tolerance \tab numeric \tab 1e-09 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   toler.chol \tab numeric \tab 1e-10 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   debug \tab integer \tab 0 \tab  \tab \eqn{[0, 1]}{[0, 1]} \cr
   outer.max \tab integer \tab 10 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   robust \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   score \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   cluster \tab untyped \tab - \tab  \tab - \cr
   ntime \tab integer \tab NULL \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   round_time \tab integer \tab 2 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
}
}

\section{Installation}{

Package 'survivalmodels' is not on CRAN and has to be install from GitHub via
\code{remotes::install_github("RaphaelS1/survivalmodels")}.
}

\references{
Kalbfleisch, D J, Prentice, L R (2011).
\emph{The statistical analysis of failure time data}.
John Wiley & Sons.
}
\seealso{
\itemize{
\item \link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}.
\item \code{as.data.table(mlr_learners)} for a table of available \link[mlr3:Learner]{Learners} in the running session (depending on the loaded packages).
\item Chapter in the \href{https://mlr3book.mlr-org.com/}{mlr3book}: \url{https://mlr3book.mlr-org.com/basics.html#learners}
\item \CRANpkg{mlr3learners} for a selection of recommended learners.
\item \CRANpkg{mlr3cluster} for unsupervised clustering learners.
\item \CRANpkg{mlr3pipelines} to combine learners with pre- and postprocessing steps.
\item \CRANpkg{mlr3tuning} for tuning of hyperparameters, \CRANpkg{mlr3tuningspaces} for established default tuning spaces.
}
}
\author{
bblodfon
}
\section{Super classes}{
\code{\link[mlr3:Learner]{mlr3::Learner}} -> \code{\link[mlr3proba:LearnerSurv]{mlr3proba::LearnerSurv}} -> \code{LearnerSurvParametric}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LearnerSurvParametric-new}{\code{LearnerSurvParametric$new()}}
\item \href{#method-LearnerSurvParametric-clone}{\code{LearnerSurvParametric$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="base_learner"><a href='../../mlr3/html/Learner.html#method-Learner-base_learner'><code>mlr3::Learner$base_learner()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="configure"><a href='../../mlr3/html/Learner.html#method-Learner-configure'><code>mlr3::Learner$configure()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="encapsulate"><a href='../../mlr3/html/Learner.html#method-Learner-encapsulate'><code>mlr3::Learner$encapsulate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="format"><a href='../../mlr3/html/Learner.html#method-Learner-format'><code>mlr3::Learner$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="help"><a href='../../mlr3/html/Learner.html#method-Learner-help'><code>mlr3::Learner$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict"><a href='../../mlr3/html/Learner.html#method-Learner-predict'><code>mlr3::Learner$predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict_newdata"><a href='../../mlr3/html/Learner.html#method-Learner-predict_newdata'><code>mlr3::Learner$predict_newdata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="print"><a href='../../mlr3/html/Learner.html#method-Learner-print'><code>mlr3::Learner$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="reset"><a href='../../mlr3/html/Learner.html#method-Learner-reset'><code>mlr3::Learner$reset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="selected_features"><a href='../../mlr3/html/Learner.html#method-Learner-selected_features'><code>mlr3::Learner$selected_features()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="train"><a href='../../mlr3/html/Learner.html#method-Learner-train'><code>mlr3::Learner$train()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerSurvParametric-new"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerSurvParametric-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
returned \code{risk} from survivalmodels is hp-style
ie higher value => higher risk
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerSurvParametric$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerSurvParametric-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerSurvParametric-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerSurvParametric$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
