% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner_survdnn_surv_survdnn.R
\name{mlr_learners_surv.survdnn}
\alias{mlr_learners_surv.survdnn}
\alias{LearnerSurvDNN}
\title{SurvDNN Survival Learner}
\description{
Deep neural network survival models from package \CRANpkg{survdnn}, aimed at
tabular (low to moderate-dimensional) covariate settings using torch-based
multilayer perceptrons. The learner wraps \code{\link[survdnn:survdnn]{survdnn::survdnn()}}.
}
\section{Prediction types}{

This learner supports the following prediction types:
\describe{
\item{\code{lp}}{A numeric vector of linear predictors, one per observation.
For \code{loss} \code{"cox"} / \code{"cox_l2"} this is a log-risk score (higher implies worse prognosis).
For \code{"aft"}, \code{\link[survdnn:predict.survdnn]{survdnn::predict.survdnn()}} returns the predicted log-time location \eqn{\mu(x)}
(higher implies better prognosis), therefore the learner internally negates it such that higher
values imply higher risk (consistent with \CRANpkg{mlr3} conventions). For \code{"coxtime"},
this is \eqn{g(t_0, x)} evaluated at a reference time.}
\item{\code{crank}}{same as \code{lp}.}
\item{\code{distr}}{A survival matrix (rows = observations, columns = time points) based on
\code{predict(type = "survival")}. By default, predictions are evaluated on the unique event times
of the training data.}
}
}

\section{Dictionary}{

This \link[mlr3:Learner]{Learner} can be instantiated via \link[mlr3:mlr_sugar]{lrn()}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{lrn("surv.survdnn")
}\if{html}{\out{</div>}}
}

\section{Meta Information}{

\itemize{
\item Task type: \dQuote{surv}
\item Predict Types: \dQuote{crank}, \dQuote{distr}, \dQuote{lp}
\item Feature Types: \dQuote{integer}, \dQuote{numeric}, \dQuote{factor}, \dQuote{ordered}
\item Required Packages: \CRANpkg{mlr3}, \CRANpkg{mlr3proba}, \CRANpkg{mlr3extralearners}, \CRANpkg{survdnn}, \CRANpkg{torch}
}
}

\section{Parameters}{
\tabular{lllll}{
   Id \tab Type \tab Default \tab Levels \tab Range \cr
   hidden \tab untyped \tab c(32L, 16L) \tab  \tab - \cr
   activation \tab character \tab relu \tab relu, leaky_relu, tanh, sigmoid, gelu, elu, softplus \tab - \cr
   lr \tab numeric \tab 1e-04 \tab  \tab \eqn{[1e-06, 1]}{[1e-06, 1]} \cr
   epochs \tab integer \tab 300 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   loss \tab character \tab cox \tab cox, cox_l2, aft, coxtime \tab - \cr
   optimizer \tab character \tab adam \tab adam, adamw, sgd, rmsprop, adagrad \tab - \cr
   optim_args \tab untyped \tab list() \tab  \tab - \cr
   verbose \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   dropout \tab numeric \tab 0.3 \tab  \tab \eqn{[0, 1]}{[0, 1]} \cr
   batch_norm \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
   callbacks \tab untyped \tab NULL \tab  \tab - \cr
   .seed \tab integer \tab NULL \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   .device \tab character \tab auto \tab auto, cpu, cuda \tab - \cr
   na_action \tab character \tab omit \tab omit, fail \tab - \cr
}
}

\examples{
\dontshow{if (learner_is_runnable("surv.survdnn")) withAutoprint(\{ # examplesIf}
# Define the Learner
learner = lrn("surv.survdnn")
print(learner)

# Define a Task
task = tsk("grace")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)


# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
\dontshow{\}) # examplesIf}
}
\seealso{
\itemize{
\item \link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}.
\item \code{as.data.table(mlr_learners)} for a table of available \link[mlr3:Learner]{Learners} in the running session (depending on the loaded packages).
\item Chapter in the \href{https://mlr3book.mlr-org.com/}{mlr3book}: \url{https://mlr3book.mlr-org.com/basics.html#learners}
\item \CRANpkg{mlr3learners} for a selection of recommended learners.
\item \CRANpkg{mlr3cluster} for unsupervised clustering learners.
\item \CRANpkg{mlr3pipelines} to combine learners with pre- and postprocessing steps.
\item \CRANpkg{mlr3tuning} for tuning of hyperparameters, \CRANpkg{mlr3tuningspaces} for established default tuning spaces.
}
}
\author{
ielbadisy
}
\section{Super classes}{
\code{\link[mlr3:Learner]{mlr3::Learner}} -> \code{\link[mlr3proba:LearnerSurv]{mlr3proba::LearnerSurv}} -> \code{LearnerSurvDNN}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{marshaled}}{(\code{logical(1)})
Whether the learner has been marshaled.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LearnerSurvDNN-new}{\code{LearnerSurvDNN$new()}}
\item \href{#method-LearnerSurvDNN-marshal}{\code{LearnerSurvDNN$marshal()}}
\item \href{#method-LearnerSurvDNN-unmarshal}{\code{LearnerSurvDNN$unmarshal()}}
\item \href{#method-LearnerSurvDNN-clone}{\code{LearnerSurvDNN$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="base_learner"><a href='../../mlr3/html/Learner.html#method-Learner-base_learner'><code>mlr3::Learner$base_learner()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="configure"><a href='../../mlr3/html/Learner.html#method-Learner-configure'><code>mlr3::Learner$configure()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="encapsulate"><a href='../../mlr3/html/Learner.html#method-Learner-encapsulate'><code>mlr3::Learner$encapsulate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="format"><a href='../../mlr3/html/Learner.html#method-Learner-format'><code>mlr3::Learner$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="help"><a href='../../mlr3/html/Learner.html#method-Learner-help'><code>mlr3::Learner$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict"><a href='../../mlr3/html/Learner.html#method-Learner-predict'><code>mlr3::Learner$predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict_newdata"><a href='../../mlr3/html/Learner.html#method-Learner-predict_newdata'><code>mlr3::Learner$predict_newdata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="print"><a href='../../mlr3/html/Learner.html#method-Learner-print'><code>mlr3::Learner$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="reset"><a href='../../mlr3/html/Learner.html#method-Learner-reset'><code>mlr3::Learner$reset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="selected_features"><a href='../../mlr3/html/Learner.html#method-Learner-selected_features'><code>mlr3::Learner$selected_features()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="train"><a href='../../mlr3/html/Learner.html#method-Learner-train'><code>mlr3::Learner$train()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerSurvDNN-new"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerSurvDNN-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerSurvDNN$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerSurvDNN-marshal"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerSurvDNN-marshal}{}}}
\subsection{Method \code{marshal()}}{
Marshal the learner's model (required for encapsulation / saveRDS).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerSurvDNN$marshal(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerSurvDNN-unmarshal"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerSurvDNN-unmarshal}{}}}
\subsection{Method \code{unmarshal()}}{
Unmarshal the learner's model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerSurvDNN$unmarshal(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerSurvDNN-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerSurvDNN-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerSurvDNN$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
