% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner_survivalsvm_surv_svm.R
\name{mlr_learners_surv.svm}
\alias{mlr_learners_surv.svm}
\alias{LearnerSurvSVM}
\title{Survival Support Vector Machine Learner}
\description{
Survival Support Vector Machine Learner

Survival Support Vector Machine Learner
}
\details{
Four possible SVMs can be implemented, dependent on the \code{type} parameter. These correspond
to predicting the survival time via regression (\code{regression}), predicting a continuous rank
(\code{vanbelle1}, \code{vanbelle2}), or a hybrid of the two (\code{hybrid}).
Whichever \code{type} is chosen determines how the \code{crank} predict type is calculated,
but in any case all can be considered a valid continuous ranking.
}
\section{Custom mlr3 defaults}{

\itemize{
\item \code{type}
\itemize{
\item original: \code{regression}
\item new: \code{hybrid}
\item reason: \code{regression} is a subet of \code{hybrid}. Replacement allows control of other parameters
}
\item \code{diff.meth}
\itemize{
\item original: none
\item new: \code{makediff3}
\item reason: Required for \code{hybrid} and this is in line with the Van Belle papers.
}
\item \code{gamma.mu}
\itemize{
\item original: none
\item new: c(0, 0))
\item reason: Arbitrary choice, must be tuned.
}
}
}

\references{
Belle VV, Pelckmans K, Huffel SV, Suykens JAK (2010).
“Improved performance on high-dimensional survival data by application of Survival-SVM.”
Bioinformatics, 27(1), 87–94.
doi: 10.1093/bioinformatics/btq617.

Belle VV, Pelckmans K, Huffel SV, Suykens JA (2011).
“Support vector methods for survival analysis: a comparison between ranking and regression
approaches."
Artificial Intelligence in Medicine, 53(2), 107–118.
doi: 10.1016/j.artmed.2011.06.006.

Shivaswamy, P. K., Chu, W., & Jansche, M. (2007).
A support vector approach to censored targets.
In Proceedings - IEEE International Conference on Data Mining, ICDM (pp. 655–660).
https://doi.org/10.1109/ICDM.2007.93
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}:
\link[mlr3:mlr_learners]{mlr3::mlr_learners}
}
\author{
RaphaelS1
}
\section{Super classes}{
\code{\link[mlr3:Learner]{mlr3::Learner}} -> \code{\link[mlr3proba:LearnerSurv]{mlr3proba::LearnerSurv}} -> \code{LearnerSurvSVM}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LearnerSurvSVM$new()}}
\item \href{#method-clone}{\code{LearnerSurvSVM$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="format">}\href{../../mlr3/html/Learner.html#method-format}{\code{mlr3::Learner$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="help">}\href{../../mlr3/html/Learner.html#method-help}{\code{mlr3::Learner$help()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict">}\href{../../mlr3/html/Learner.html#method-predict}{\code{mlr3::Learner$predict()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict_newdata">}\href{../../mlr3/html/Learner.html#method-predict_newdata}{\code{mlr3::Learner$predict_newdata()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="print">}\href{../../mlr3/html/Learner.html#method-print}{\code{mlr3::Learner$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="reset">}\href{../../mlr3/html/Learner.html#method-reset}{\code{mlr3::Learner$reset()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="train">}\href{../../mlr3/html/Learner.html#method-train}{\code{mlr3::Learner$train()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerSurvSVM$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerSurvSVM$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
