% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_cforest.R
\name{train_cforest}
\alias{train_cforest}
\title{all 3 cforest learners have the same .train call: DRY}
\usage{
train_cforest(self, task)
}
\description{
all 3 cforest learners have the same .train call: DRY
}
