% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_learner.R
\name{create_learner}
\alias{create_learner}
\title{Create a New Learner}
\usage{
create_learner(
  path = ".",
  classname,
  type,
  key = tolower(classname),
  algorithm,
  package,
  caller,
  feature_types,
  predict_types,
  properties,
  gh_name = "Unknown",
  label = toproper(algorithm),
  data_formats = "data.table"
)
}
\arguments{
\item{path}{(\code{character(1)})\cr
The path to a folder. This is where the files will be created.
In case the folder is an R package, the learner file will be create in path/R and the
test files will be created in path/tests/testthat. Otherwise all the files will be created
in path.}

\item{classname}{(\code{character(1)})\cr
Suffix for R6 class name passed to LearnerType 'classname'.}

\item{type}{(\code{character(1)})\cr
See \code{mlr3::mlr_reflections$task_types$type}.}

\item{key}{(\code{character(1)})\cr
key for learner, if not provided defaults to the \code{classname} in all lower case.
In combination with \code{type} it creates the learner's id.}

\item{algorithm}{(\code{character(1)})\cr
Brief description of the algorithm, like "Linear Model" or "Random Forest".
Is used for the title of the help package and as the label (if no other label is provided).}

\item{package}{(\code{character(1)})\cr
Package from which the learner is implemented.}

\item{caller}{\code{character(1)}\cr
Training function called from the upstream package.}

\item{feature_types}{(\code{character()})\cr
Feature types that can be handled by the learner, see
\code{mlr3::mlr_reflections$task_feature_types}.}

\item{predict_types}{(\code{character()})\cr
Prediction types that can be made by the learner, see
\code{mlr3::mlr_reflections$learner_predict_types}.}

\item{properties}{(\code{character()})\cr
Properties that can be handled by the learner, see \code{mlr3::mlr_reflections$learner_properties}.}

\item{gh_name}{(\code{character(1)})\cr
Your GitHub handle, used to add you as the maintainer of the learner.
Defaults to \code{"Unknown"}.}

\item{label}{(\code{character(1)})\cr
Label for the learner, default is the value of the parameter \code{algorithm}.}

\item{data_formats}{(\code{character(1)})\cr
The data format the learner can deal with (see \code{mlr_reflections$data_formats}).
Default is \code{"data.table"}.}
}
\description{
Helper function to create a template for a learner, as well as the test and parameter test.
For more details see the \href{https://mlr3book.mlr-org.com/extending.html}{mlr3book}.
}
\examples{
\dontrun{
path = tempfile()
dir.create(path)
create_learner(
  path = path,
  classname = "Rpart",
  type = "classif",
  key = "rpart",
  algorithm = "Decision Tree",
  package = "rpart",
  caller = "rpart",
  feature_types = c("logical", "integer", "numeric", "factor", "ordered"),
  predict_types = c("response", "prob"),
  properties = c("importance", "missings", "multiclass", "twoclass", "weights"),
  gh_name = "RaphaelS1",
  label = "Regression and Partition Tree",
  data_formats = "data.table"
)
}
}
