% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_install_catboost.R
\name{install_catboost}
\alias{install_catboost}
\title{Install catboost}
\usage{
install_catboost(
  version = NULL,
  INSTALL_opts = c("--no-multiarch", "--no-test-load"),
  ...
)
}
\arguments{
\item{version}{\code{(character(1))}
Version to install, if \code{NULL} installs latest.
Can e.g. be \code{"1.0.6"}.}

\item{INSTALL_opts}{\code{(character())} \cr
Passed to \link[remotes:install_url]{remotes::install_url}}

\item{...}{\code{ANY} \cr
Other arguments passed to \link[remotes:install_url]{remotes::install_url}}
}
\description{
Helper function to install catboost.
Installs the latest version if no version is provided.
}
