% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_proba.R
\name{install_keras}
\alias{install_keras}
\title{Install Keras and the TensorFlow backend}
\usage{
install_keras(
  method = c("auto", "virtualenv", "conda"),
  conda = "auto",
  version = "default",
  tensorflow = "default",
  extra_packages = c("tensorflow-hub"),
  ...
)
}
\arguments{
\item{method, conda, version, tensorflow, extra_packages, ...}{See \link[keras:install_keras]{keras::install_keras}.}
}
\description{
See \link[keras:install_keras]{keras::install_keras} for more details.
}
