% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_install_learners.R
\name{install_learners}
\alias{install_learners}
\title{Install Learner Dependencies}
\usage{
install_learners(.keys, ...)
}
\arguments{
\item{.keys}{\code{character()} \cr Keys passed to \link[mlr3:mlr_learners]{mlr_learners} specifying
learners to install.}

\item{...}{\code{ANY} \cr Additional options to pass to \link[utils:install.packages]{utils::install.packages} or
\link[remotes:install_github]{remotes::install_github}.}
}
\description{
Install required dependencies for specified learners. Works for packages on
GitHub and CRAN, otherwise must be manually installed.
}
