% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{install_learners}
\alias{install_learners}
\title{Install Learner Dependencies}
\usage{
install_learners(.keys, ...)
}
\arguments{
\item{.keys}{\code{character()} \cr Keys passed to \link[mlr3:mlr_learners]{mlr_learners} specifying
learners to install.}

\item{...}{\code{ANY} \cr Additional options to pass to \link{install.packages}.}
}
\description{
Install required dependencies for specified learners.
}
\examples{
\dontrun{
install_learners(c("regr.gbm", "classif.kknn"))
}
}
