% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3extralearners-package}
\alias{mlr3extralearners}
\alias{mlr3extralearners-package}
\title{mlr3extralearners: Extra Learners For mlr3}
\description{
Extra learners for use in mlr3.
}
\author{
\strong{Maintainer}: Raphael Sonabend \email{raphael.sonabend.15@ucl.ac.uk} (\href{https://orcid.org/0000-0001-9225-4654}{ORCID})

Authors:
\itemize{
  \item Patrick Schratz \email{patrick.schratz@gmail.com} (\href{https://orcid.org/0000-0003-0748-6624}{ORCID})
}

Other contributors:
\itemize{
  \item Lorenz A. Kapsner \email{lorenz.kapsner@gmail.com} (\href{https://orcid.org/0000-0003-1866-860X}{ORCID}) [contributor]
}

}
