% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner_C50_classif_C50.R
\name{mlr_learners_classif.C50}
\alias{mlr_learners_classif.C50}
\alias{LearnerClassifC50}
\title{Classification C5.0 Learner}
\description{
Decision Tree Algorithm.
Calls \code{\link[C50:C5.0]{C50::C5.0.formula()}} from \CRANpkg{C50}.
}
\section{Dictionary}{

This \link{Learner} can be instantiated via the \link[mlr3misc:Dictionary]{dictionary} \link{mlr_learners} or with the associated sugar function \code{\link[=lrn]{lrn()}}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{mlr_learners$get("classif.C50")
lrn("classif.C50")
}\if{html}{\out{</div>}}
}

\section{Meta Information}{

\itemize{
\item Task type: \dQuote{classif}
\item Predict Types: \dQuote{response}, \dQuote{prob}
\item Feature Types: \dQuote{numeric}, \dQuote{factor}, \dQuote{ordered}
\item Required Packages: \CRANpkg{mlr3}, \CRANpkg{mlr3extralearners}, \CRANpkg{C50}
}
}

\section{Parameters}{
\tabular{lllll}{
   Id \tab Type \tab Default \tab Levels \tab Range \cr
   trials \tab integer \tab 1 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   rules \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   costs \tab untyped \tab  \tab  \tab - \cr
   subset \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
   bands \tab integer \tab - \tab  \tab \eqn{[0, 1000]}{[0, 1000]} \cr
   winnow \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   noGlobalPruning \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   CF \tab numeric \tab 0.25 \tab  \tab \eqn{[0, 1]}{[0, 1]} \cr
   minCases \tab integer \tab 2 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   fuzzyThreshold \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   sample \tab numeric \tab 0 \tab  \tab \eqn{[0, 0.999]}{[0, 0.999]} \cr
   seed \tab integer \tab - \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   earlyStopping \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
   label \tab untyped \tab outcome \tab  \tab - \cr
   na.action \tab untyped \tab ::     , stats  , na.pass \tab  \tab - \cr
}
}

\references{
Quinlan, Ross J (2014).
\emph{C4. 5: programs for machine learning}.
Elsevier.
}
\author{
henrifnk
}
\section{Super classes}{
\code{\link[mlr3:Learner]{mlr3::Learner}} -> \code{\link[mlr3:LearnerClassif]{mlr3::LearnerClassif}} -> \code{LearnerClassifC50}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LearnerClassifC50-new}{\code{LearnerClassifC50$new()}}
\item \href{#method-LearnerClassifC50-clone}{\code{LearnerClassifC50$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="base_learner"><a href='../../mlr3/html/Learner.html#method-Learner-base_learner'><code>mlr3::Learner$base_learner()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="format"><a href='../../mlr3/html/Learner.html#method-Learner-format'><code>mlr3::Learner$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="help"><a href='../../mlr3/html/Learner.html#method-Learner-help'><code>mlr3::Learner$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict"><a href='../../mlr3/html/Learner.html#method-Learner-predict'><code>mlr3::Learner$predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict_newdata"><a href='../../mlr3/html/Learner.html#method-Learner-predict_newdata'><code>mlr3::Learner$predict_newdata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="print"><a href='../../mlr3/html/Learner.html#method-Learner-print'><code>mlr3::Learner$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="reset"><a href='../../mlr3/html/Learner.html#method-Learner-reset'><code>mlr3::Learner$reset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="train"><a href='../../mlr3/html/Learner.html#method-Learner-train'><code>mlr3::Learner$train()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerClassifC50-new"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerClassifC50-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerClassifC50$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerClassifC50-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerClassifC50-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerClassifC50$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
