% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner_partykit_classif_cforest.R
\name{mlr_learners_classif.cforest}
\alias{mlr_learners_classif.cforest}
\alias{LearnerClassifCForest}
\title{Classification Conditional Random Forest Learner}
\description{
Calls \link[partykit:cforest]{partykit::cforest} from package \CRANpkg{partykit}.
}
\section{Dictionary}{
 This \link{Learner} can be instantiated via the
\link[mlr3misc:Dictionary]{dictionary} \link{mlr_learners} or with the associated
sugar function \code{\link[=lrn]{lrn()}}:\preformatted{mlr_learners$get("classif.cforest")
lrn("classif.cforest")
}
}

\section{Meta Information}{

\itemize{
\item Task type: \dQuote{classif}
\item Predict Types: \dQuote{response}, \dQuote{prob}
\item Feature Types: \dQuote{integer}, \dQuote{numeric}, \dQuote{factor}, \dQuote{ordered}
\item Required Packages: \CRANpkg{mlr3}, \CRANpkg{mlr3extralearners}, \CRANpkg{partykit}, \CRANpkg{sandwich}, \CRANpkg{coin}
}
}

\section{Parameters}{
\tabular{lllll}{
   Id \tab Type \tab Default \tab Levels \tab Range \cr
   ntree \tab integer \tab 500 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   replace \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   fraction \tab numeric \tab 0.632 \tab  \tab \eqn{[0, 1]}{[0, 1]} \cr
   mtry \tab integer \tab - \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   mtryratio \tab numeric \tab - \tab  \tab \eqn{[0, 1]}{[0, 1]} \cr
   applyfun \tab list \tab - \tab  \tab - \cr
   cores \tab integer \tab NULL \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   trace \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   offset \tab list \tab - \tab  \tab - \cr
   cluster \tab list \tab - \tab  \tab - \cr
   scores \tab list \tab - \tab  \tab - \cr
   teststat \tab character \tab quadratic \tab quadratic, maximum \tab - \cr
   splitstat \tab character \tab quadratic \tab quadratic, maximum \tab - \cr
   splittest \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   testtype \tab character \tab Univariate \tab Bonferroni, MonteCarlo, Univariate, Teststatistic \tab - \cr
   nmax \tab list \tab - \tab  \tab - \cr
   pargs \tab list \tab - \tab  \tab - \cr
   alpha \tab numeric \tab 0.05 \tab  \tab \eqn{[0, 1]}{[0, 1]} \cr
   mincriterion \tab numeric \tab 0 \tab  \tab \eqn{[0, 1]}{[0, 1]} \cr
   logmincriterion \tab numeric \tab 0 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   minsplit \tab integer \tab 20 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   minbucket \tab integer \tab 7 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   minprob \tab numeric \tab 0.01 \tab  \tab \eqn{[0, 1]}{[0, 1]} \cr
   stump \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   lookahead \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   MIA \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   nresample \tab integer \tab 9999 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   tol \tab numeric \tab 1.490116e-08 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   maxsurrogate \tab integer \tab 0 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   numsurrogate \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   maxdepth \tab integer \tab Inf \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   multiway \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   splittry \tab integer \tab 2 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   intersplit \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   majority \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   caseweights \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
   saveinfo \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   update \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   splitflavour \tab character \tab ctree \tab ctree, exhaustive \tab - \cr
   maxvar \tab integer \tab - \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   OOB \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   simplify \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
   scale \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
   nperm \tab integer \tab 1 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   risk \tab character \tab loglik \tab loglik, misclassification \tab - \cr
   conditional \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   threshold \tab numeric \tab 0.2 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
}
}

\section{Custom mlr3 defaults}{

\itemize{
\item \code{mtry}:
\itemize{
\item This hyperparameter can alternatively be set via the added hyperparameter \code{mtryratio}
as \code{mtry = max(ceiling(mtryratio * n_features), 1)}.
Note that \code{mtry} and \code{mtryratio} are mutually exclusive.
}
}
}

\examples{
if (requireNamespace("partykit", quietly = TRUE) && requireNamespace("sandwich", quietly = TRUE) && requireNamespace("coin", quietly = TRUE)) {
  learner = mlr3::lrn("classif.cforest")
  print(learner)

  # available parameters:
  learner$param_set$ids()
}
}
\references{
`r format_bib(c("hothorn_2015", "hothorn_2006"))
}
\seealso{
\itemize{
\item \link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}.
\item \code{as.data.table(mlr_learners)} for a table of available \link[=Learner]{Learners} in the running session (depending on the loaded packages).
\item Chapter in the \href{https://mlr3book.mlr-org.com/}{mlr3book}: \url{https://mlr3book.mlr-org.com/basics.html#learners}
\item \CRANpkg{mlr3learners} for a selection of recommended learners.
\item \CRANpkg{mlr3cluster} for unsupervised clustering learners.
\item \CRANpkg{mlr3pipelines} to combine learners with pre- and postprocessing steps.
\item \CRANpkg{mlr3tuning} for tuning of hyperparameters, \CRANpkg{mlr3tuningspaces} for established default tuning spaces.
}
}
\author{
sumny
}
\section{Super classes}{
\code{\link[mlr3:Learner]{mlr3::Learner}} -> \code{\link[mlr3:LearnerClassif]{mlr3::LearnerClassif}} -> \code{LearnerClassifCForest}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LearnerClassifCForest$new()}}
\item \href{#method-oob_error}{\code{LearnerClassifCForest$oob_error()}}
\item \href{#method-clone}{\code{LearnerClassifCForest$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="base_learner">}\href{../../mlr3/html/Learner.html#method-base_learner}{\code{mlr3::Learner$base_learner()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="format">}\href{../../mlr3/html/Learner.html#method-format}{\code{mlr3::Learner$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="help">}\href{../../mlr3/html/Learner.html#method-help}{\code{mlr3::Learner$help()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict">}\href{../../mlr3/html/Learner.html#method-predict}{\code{mlr3::Learner$predict()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict_newdata">}\href{../../mlr3/html/Learner.html#method-predict_newdata}{\code{mlr3::Learner$predict_newdata()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="print">}\href{../../mlr3/html/Learner.html#method-print}{\code{mlr3::Learner$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="reset">}\href{../../mlr3/html/Learner.html#method-reset}{\code{mlr3::Learner$reset()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="train">}\href{../../mlr3/html/Learner.html#method-train}{\code{mlr3::Learner$train()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerClassifCForest$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-oob_error"></a>}}
\if{latex}{\out{\hypertarget{method-oob_error}{}}}
\subsection{Method \code{oob_error()}}{
The out-of-bag error, calculated using the OOB predictions from
\code{partykit}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerClassifCForest$oob_error()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{numeric(1)}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerClassifCForest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
