% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner_randomPlantedForest_classif_rpf.R
\name{mlr_learners_classif.rpf}
\alias{mlr_learners_classif.rpf}
\alias{LearnerClassifRandomPlantedForest}
\title{Classification Random Planted Forest Learner}
\description{
Random Planted Forest: A directly interpretable tree ensemble.

Calls \code{\link[randomPlantedForest:rpf]{randomPlantedForest::rpf()}} from 'randomPlantedForest'.
}
\section{Initial parameter values}{

\itemize{
\item \code{loss}:
\itemize{
\item Actual default: \code{"L2"}.
\item Initial value: \code{"exponential"}.
\item Reason for change: Using \code{"L2"} (or \code{"L1"}) loss does not guarantee predictions are valid
probabilities and more akin to the linear predictor of a GLM.
}
}
}

\section{Custom mlr3 parameters}{

\itemize{
\item \code{max_interaction}:
\itemize{
\item This hyperparameter can alternatively be set via \code{max_interaction_ratio} as
\code{max_interaction = max(ceiling(max_interaction_ratio * n_features), 1)}.
The parameter \code{max_interaction_limit} can optionally be set as an upper bound, such that
\code{max_interaction_ratio * min(n_features, max_interaction_limit)} is used instead.
This is analogous to \code{mtry.ratio} in \code{\link[mlr3learners:mlr_learners_classif.ranger]{classif.ranger}}, with
\code{max_interaction_limit} as an additional constraint.
The parameter \code{max_interaction_limit} is initialized to \code{Inf}.
}
}
}

\section{Dictionary}{

This \link[mlr3:Learner]{Learner} can be instantiated via \link[=lrn]{lrn()}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{lrn("classif.rpf")
}\if{html}{\out{</div>}}
}

\section{Meta Information}{

\itemize{
\item Task type: \dQuote{classif}
\item Predict Types: \dQuote{response}, \dQuote{prob}
\item Feature Types: \dQuote{logical}, \dQuote{integer}, \dQuote{numeric}, \dQuote{factor}, \dQuote{ordered}
\item Required Packages: \CRANpkg{mlr3}, \CRANpkg{randomPlantedForest}
}
}

\section{Parameters}{
\tabular{lllll}{
   Id \tab Type \tab Default \tab Levels \tab Range \cr
   max_interaction \tab integer \tab 1 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   max_interaction_ratio \tab numeric \tab - \tab  \tab \eqn{[0, 1]}{[0, 1]} \cr
   max_interaction_limit \tab integer \tab - \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   ntrees \tab integer \tab 50 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   splits \tab integer \tab 30 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   split_try \tab integer \tab 10 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   t_try \tab numeric \tab 0.4 \tab  \tab \eqn{[0, 1]}{[0, 1]} \cr
   loss \tab character \tab L2 \tab L1, L2, logit, exponential \tab - \cr
   delta \tab numeric \tab 1 \tab  \tab \eqn{[0, 1]}{[0, 1]} \cr
   epsilon \tab numeric \tab 0.1 \tab  \tab \eqn{[0, 1]}{[0, 1]} \cr
   deterministic \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   nthreads \tab integer \tab 1 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   cv \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   purify \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
}
}

\section{Installation}{

Package 'randomPlantedForest' is not on CRAN and has to be installed from GitHub via
\code{remotes::install_github("PlantedML/randomPlantedForest")}.
}

\examples{
\dontshow{if (learner_is_runnable("classif.rpf")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Define the Learner
learner = lrn("classif.rpf")
print(learner)

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

print(learner$model)


# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
\dontshow{\}) # examplesIf}
}
\references{
Hiabu, Munir, Mammen, Enno, Meyer, T. J (2023).
\dQuote{Random Planted Forest: a directly interpretable tree ensemble.}
\emph{arXiv preprint arXiv:2012.14563}.
\doi{10.48550/ARXIV.2012.14563}.
}
\seealso{
\itemize{
\item \link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}.
\item \code{as.data.table(mlr_learners)} for a table of available \link[mlr3:Learner]{Learners} in the running session (depending on the loaded packages).
\item Chapter in the \href{https://mlr3book.mlr-org.com/}{mlr3book}: \url{https://mlr3book.mlr-org.com/basics.html#learners}
\item \CRANpkg{mlr3learners} for a selection of recommended learners.
\item \CRANpkg{mlr3cluster} for unsupervised clustering learners.
\item \CRANpkg{mlr3pipelines} to combine learners with pre- and postprocessing steps.
\item \CRANpkg{mlr3tuning} for tuning of hyperparameters, \CRANpkg{mlr3tuningspaces} for established default tuning spaces.
}
}
\author{
jemus42
}
\section{Super classes}{
\code{\link[mlr3:Learner]{mlr3::Learner}} -> \code{\link[mlr3:LearnerClassif]{mlr3::LearnerClassif}} -> \code{LearnerClassifRandomPlantedForest}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LearnerClassifRandomPlantedForest-new}{\code{LearnerClassifRandomPlantedForest$new()}}
\item \href{#method-LearnerClassifRandomPlantedForest-clone}{\code{LearnerClassifRandomPlantedForest$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="base_learner"><a href='../../mlr3/html/Learner.html#method-Learner-base_learner'><code>mlr3::Learner$base_learner()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="configure"><a href='../../mlr3/html/Learner.html#method-Learner-configure'><code>mlr3::Learner$configure()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="encapsulate"><a href='../../mlr3/html/Learner.html#method-Learner-encapsulate'><code>mlr3::Learner$encapsulate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="format"><a href='../../mlr3/html/Learner.html#method-Learner-format'><code>mlr3::Learner$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="help"><a href='../../mlr3/html/Learner.html#method-Learner-help'><code>mlr3::Learner$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict"><a href='../../mlr3/html/Learner.html#method-Learner-predict'><code>mlr3::Learner$predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict_newdata"><a href='../../mlr3/html/Learner.html#method-Learner-predict_newdata'><code>mlr3::Learner$predict_newdata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="print"><a href='../../mlr3/html/Learner.html#method-Learner-print'><code>mlr3::Learner$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="reset"><a href='../../mlr3/html/Learner.html#method-Learner-reset'><code>mlr3::Learner$reset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="selected_features"><a href='../../mlr3/html/Learner.html#method-Learner-selected_features'><code>mlr3::Learner$selected_features()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="train"><a href='../../mlr3/html/Learner.html#method-Learner-train'><code>mlr3::Learner$train()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="LearnerClassif" data-id="predict_newdata_fast"><a href='../../mlr3/html/LearnerClassif.html#method-LearnerClassif-predict_newdata_fast'><code>mlr3::LearnerClassif$predict_newdata_fast()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerClassifRandomPlantedForest-new"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerClassifRandomPlantedForest-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerClassifRandomPlantedForest$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerClassifRandomPlantedForest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerClassifRandomPlantedForest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerClassifRandomPlantedForest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
