% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner_ks_dens_kde_ks.R
\name{mlr_learners_dens.kde_ks}
\alias{mlr_learners_dens.kde_ks}
\alias{LearnerDensKDEks}
\title{Density KS Kernel Learner}
\description{
Kernel density estimator.
Calls \code{\link[ks:kde]{ks::kde()}} from \CRANpkg{ks}.
}
\section{Dictionary}{

This \link{Learner} can be instantiated via the \link[mlr3misc:Dictionary]{dictionary} \link{mlr_learners} or with the associated sugar function \code{\link[=lrn]{lrn()}}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{mlr_learners$get("dens.kde_ks")
lrn("dens.kde_ks")
}\if{html}{\out{</div>}}
}

\section{Meta Information}{

\itemize{
\item Task type: \dQuote{dens}
\item Predict Types: \dQuote{pdf}
\item Feature Types: \dQuote{integer}, \dQuote{numeric}
\item Required Packages: \CRANpkg{mlr3}, \CRANpkg{mlr3proba}, \CRANpkg{mlr3extralearners}, \CRANpkg{ks}
}
}

\section{Parameters}{
\tabular{lllll}{
   Id \tab Type \tab Default \tab Levels \tab Range \cr
   h \tab numeric \tab - \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   H \tab untyped \tab - \tab  \tab - \cr
   gridsize \tab untyped \tab - \tab  \tab - \cr
   gridtype \tab untyped \tab - \tab  \tab - \cr
   xmin \tab numeric \tab - \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   xmax \tab numeric \tab - \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   supp \tab numeric \tab 3.7 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   binned \tab numeric \tab - \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   bgridsize \tab untyped \tab - \tab  \tab - \cr
   positive \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   adj.positive \tab untyped \tab - \tab  \tab - \cr
   w \tab untyped \tab - \tab  \tab - \cr
   compute.cont \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
   approx.cont \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
   unit.interval \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   verbose \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   density \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
}
}

\examples{
learner = mlr3::lrn("dens.kde_ks")
print(learner)

# available parameters:
learner$param_set$ids()
}
\references{
Gramacki, Artur, Gramacki, Jarosław (2017).
\dQuote{FFT-based fast computation of multivariate kernel density estimators with unconstrained bandwidth matrices.}
\emph{Journal of Computational and Graphical Statistics}, \bold{26}(2), 459--462.
}
\seealso{
\itemize{
\item \link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}.
\item \code{as.data.table(mlr_learners)} for a table of available \link[=Learner]{Learners} in the running session (depending on the loaded packages).
\item Chapter in the \href{https://mlr3book.mlr-org.com/}{mlr3book}: \url{https://mlr3book.mlr-org.com/basics.html#learners}
\item \CRANpkg{mlr3learners} for a selection of recommended learners.
\item \CRANpkg{mlr3cluster} for unsupervised clustering learners.
\item \CRANpkg{mlr3pipelines} to combine learners with pre- and postprocessing steps.
\item \CRANpkg{mlr3tuning} for tuning of hyperparameters, \CRANpkg{mlr3tuningspaces} for established default tuning spaces.
}
}
\author{
RaphaelS1
}
\section{Super classes}{
\code{\link[mlr3:Learner]{mlr3::Learner}} -> \code{\link[mlr3proba:LearnerDens]{mlr3proba::LearnerDens}} -> \code{LearnerDensKDEks}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LearnerDensKDEks-new}{\code{LearnerDensKDEks$new()}}
\item \href{#method-LearnerDensKDEks-clone}{\code{LearnerDensKDEks$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="base_learner"><a href='../../mlr3/html/Learner.html#method-Learner-base_learner'><code>mlr3::Learner$base_learner()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="format"><a href='../../mlr3/html/Learner.html#method-Learner-format'><code>mlr3::Learner$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="help"><a href='../../mlr3/html/Learner.html#method-Learner-help'><code>mlr3::Learner$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict"><a href='../../mlr3/html/Learner.html#method-Learner-predict'><code>mlr3::Learner$predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict_newdata"><a href='../../mlr3/html/Learner.html#method-Learner-predict_newdata'><code>mlr3::Learner$predict_newdata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="print"><a href='../../mlr3/html/Learner.html#method-Learner-print'><code>mlr3::Learner$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="reset"><a href='../../mlr3/html/Learner.html#method-Learner-reset'><code>mlr3::Learner$reset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="train"><a href='../../mlr3/html/Learner.html#method-Learner-train'><code>mlr3::Learner$train()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerDensKDEks-new"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerDensKDEks-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerDensKDEks$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerDensKDEks-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerDensKDEks-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerDensKDEks$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
