% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner_np_dens_mixed.R
\name{mlr_learners_dens.mixed}
\alias{mlr_learners_dens.mixed}
\alias{LearnerDensMixed}
\title{Density Mixed Data Kernel Learner}
\description{
Density estimator for discrete and continuous variables.
Calls \code{\link[np:np.density]{np::npudens()}} from \CRANpkg{np}.
}
\section{Dictionary}{

This \link{Learner} can be instantiated via the \link[mlr3misc:Dictionary]{dictionary} \link{mlr_learners} or with the associated sugar function \code{\link[=lrn]{lrn()}}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{mlr_learners$get("dens.mixed")
lrn("dens.mixed")
}\if{html}{\out{</div>}}
}

\section{Meta Information}{

\itemize{
\item Task type: \dQuote{dens}
\item Predict Types: \dQuote{pdf}
\item Feature Types: \dQuote{integer}, \dQuote{numeric}
\item Required Packages: \CRANpkg{mlr3}, \CRANpkg{mlr3proba}, \CRANpkg{mlr3extralearners}, \CRANpkg{np}
}
}

\section{Parameters}{
\tabular{lllll}{
   Id \tab Type \tab Default \tab Levels \tab Range \cr
   bws \tab untyped \tab - \tab  \tab - \cr
   ckertype \tab character \tab gaussian \tab gaussian, epanechnikov, uniform \tab - \cr
   bwscaling \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   bwmethod \tab character \tab cv.ml \tab cv.ml, cv.ls, normal-reference \tab - \cr
   bwtype \tab character \tab fixed \tab fixed, generalized_nn, adaptive_nn \tab - \cr
   bandwidth.compute \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   ckerorder \tab integer \tab 2 \tab  \tab \eqn{[2, 8]}{[2, 8]} \cr
   remin \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
   itmax \tab integer \tab 10000 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   nmulti \tab integer \tab - \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   ftol \tab numeric \tab 1.490116e-07 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   tol \tab numeric \tab 0.0001490116 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   small \tab numeric \tab 1.490116e-05 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   lbc.dir \tab numeric \tab 0.5 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   dfc.dir \tab numeric \tab 0.5 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   cfac.dir \tab untyped \tab *            , 2.5          , (3 - sqrt(5)) \tab  \tab - \cr
   initc.dir \tab numeric \tab 1 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   lbd.dir \tab numeric \tab 0.1 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   hbd.dir \tab numeric \tab 1 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   dfac.dir \tab untyped \tab *            , 0.25         , (3 - sqrt(5)) \tab  \tab - \cr
   initd.dir \tab numeric \tab 1 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   lbc.init \tab numeric \tab 0.1 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   hbc.init \tab numeric \tab 2 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   cfac.init \tab numeric \tab 0.5 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   lbd.init \tab numeric \tab 0.1 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   hbd.init \tab numeric \tab 0.9 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   dfac.init \tab numeric \tab 0.37 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   ukertype \tab character \tab - \tab aitchisonaitken, liracine \tab - \cr
   okertype \tab character \tab - \tab wangvanryzin, liracine \tab - \cr
}
}

\examples{
learner = mlr3::lrn("dens.mixed")
print(learner)

# available parameters:
learner$param_set$ids()
}
\references{
Li, Qi, Racine, Jeff (2003).
\dQuote{Nonparametric estimation of distributions with categorical and continuous data.}
\emph{journal of multivariate analysis}, \bold{86}(2), 266--292.
}
\seealso{
\itemize{
\item \link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}.
\item \code{as.data.table(mlr_learners)} for a table of available \link[=Learner]{Learners} in the running session (depending on the loaded packages).
\item Chapter in the \href{https://mlr3book.mlr-org.com/}{mlr3book}: \url{https://mlr3book.mlr-org.com/basics.html#learners}
\item \CRANpkg{mlr3learners} for a selection of recommended learners.
\item \CRANpkg{mlr3cluster} for unsupervised clustering learners.
\item \CRANpkg{mlr3pipelines} to combine learners with pre- and postprocessing steps.
\item \CRANpkg{mlr3tuning} for tuning of hyperparameters, \CRANpkg{mlr3tuningspaces} for established default tuning spaces.
}
}
\author{
RaphaelS1
}
\section{Super classes}{
\code{\link[mlr3:Learner]{mlr3::Learner}} -> \code{\link[mlr3proba:LearnerDens]{mlr3proba::LearnerDens}} -> \code{LearnerDensMixed}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LearnerDensMixed-new}{\code{LearnerDensMixed$new()}}
\item \href{#method-LearnerDensMixed-clone}{\code{LearnerDensMixed$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="base_learner"><a href='../../mlr3/html/Learner.html#method-Learner-base_learner'><code>mlr3::Learner$base_learner()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="format"><a href='../../mlr3/html/Learner.html#method-Learner-format'><code>mlr3::Learner$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="help"><a href='../../mlr3/html/Learner.html#method-Learner-help'><code>mlr3::Learner$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict"><a href='../../mlr3/html/Learner.html#method-Learner-predict'><code>mlr3::Learner$predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict_newdata"><a href='../../mlr3/html/Learner.html#method-Learner-predict_newdata'><code>mlr3::Learner$predict_newdata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="print"><a href='../../mlr3/html/Learner.html#method-Learner-print'><code>mlr3::Learner$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="reset"><a href='../../mlr3/html/Learner.html#method-Learner-reset'><code>mlr3::Learner$reset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="train"><a href='../../mlr3/html/Learner.html#method-Learner-train'><code>mlr3::Learner$train()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerDensMixed-new"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerDensMixed-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerDensMixed$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerDensMixed-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerDensMixed-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerDensMixed$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
