% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner_mgcv_regr_bam.R
\name{mlr_learners_regr.bam}
\alias{mlr_learners_regr.bam}
\alias{LearnerRegrBam}
\title{Regression Fast Generalized Additive Model Learner}
\description{
Fast generalized additive models for large datasets (BAM).
Calls \code{mgcv::bam()} from package \CRANpkg{mgcv}.
}
\section{Formula}{

A gam formula specific to the task at hand is required for the \code{formula}
parameter (see example and \code{mgcv::formula.gam}). Beware, if no formula is provided, a fallback formula is
used that will make the model behave like a glm (this behavior is required
for the unit tests). Only features specified in the formula will be used,
superseding columns with col_roles "feature" in the task.
}

\section{Dictionary}{

This \link[mlr3:Learner]{Learner} can be instantiated via \link[mlr3:mlr_sugar]{lrn()}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{lrn("regr.bam")
}\if{html}{\out{</div>}}
}

\section{Meta Information}{

\itemize{
\item Task type: \dQuote{regr}
\item Predict Types: \dQuote{response}, \dQuote{se}
\item Feature Types: \dQuote{logical}, \dQuote{integer}, \dQuote{numeric}, \dQuote{factor}
\item Required Packages: \CRANpkg{mlr3}, \CRANpkg{mgcv}
}
}

\section{Parameters}{
\tabular{lllll}{
   Id \tab Type \tab Default \tab Levels \tab Range \cr
   formula \tab untyped \tab - \tab  \tab - \cr
   family \tab character \tab gaussian \tab gaussian, poisson \tab - \cr
   method \tab character \tab fREML \tab fREML, REML, GCV.Cp \tab - \cr
   discrete \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
   nthreads \tab integer \tab 1 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   chunk.size \tab integer \tab 10000 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   sparse \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   samfrac \tab numeric \tab 1 \tab  \tab \eqn{[0, 1]}{[0, 1]} \cr
   select \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   gamma \tab numeric \tab 1 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   knots \tab untyped \tab NULL \tab  \tab - \cr
   sp \tab untyped \tab NULL \tab  \tab - \cr
   min.sp \tab untyped \tab NULL \tab  \tab - \cr
   scale \tab numeric \tab 0 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   paraPen \tab untyped \tab NULL \tab  \tab - \cr
   G \tab untyped \tab NULL \tab  \tab - \cr
   in.out \tab untyped \tab NULL \tab  \tab - \cr
   drop.unused.levels \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
   drop.intercept \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   block.size \tab integer \tab 50000 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
}
}

\examples{
if (requireNamespace("mgcv", quietly = TRUE)) {
  task = mlr3::tsk("mtcars")

  learner = LearnerRegrBam$new()

  learner$param_set$set_values(
    formula = mpg ~ s(disp, k = 3) + s(hp, k = 4) + cyl,
    method = "fREML"
  )

  learner$train(task)
  print(learner$model)

  pred = learner$predict(task)
  print(pred)
}
}
\references{
Hastie, J T, Tibshirani, J R (2017).
\emph{Generalized additive models}.
Routledge.

Wood, Simon (2012).
\dQuote{mgcv: Mixed GAM Computation Vehicle with GCV/AIC/REML smoothness estimation.}
}
\author{
Tino-Rg
}
\section{Super classes}{
\code{\link[mlr3:Learner]{mlr3::Learner}} -> \code{\link[mlr3:LearnerRegr]{mlr3::LearnerRegr}} -> \code{LearnerRegrBam}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LearnerRegrBam-new}{\code{LearnerRegrBam$new()}}
\item \href{#method-LearnerRegrBam-clone}{\code{LearnerRegrBam$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="base_learner"><a href='../../mlr3/html/Learner.html#method-Learner-base_learner'><code>mlr3::Learner$base_learner()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="configure"><a href='../../mlr3/html/Learner.html#method-Learner-configure'><code>mlr3::Learner$configure()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="encapsulate"><a href='../../mlr3/html/Learner.html#method-Learner-encapsulate'><code>mlr3::Learner$encapsulate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="format"><a href='../../mlr3/html/Learner.html#method-Learner-format'><code>mlr3::Learner$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="help"><a href='../../mlr3/html/Learner.html#method-Learner-help'><code>mlr3::Learner$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict"><a href='../../mlr3/html/Learner.html#method-Learner-predict'><code>mlr3::Learner$predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict_newdata"><a href='../../mlr3/html/Learner.html#method-Learner-predict_newdata'><code>mlr3::Learner$predict_newdata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="print"><a href='../../mlr3/html/Learner.html#method-Learner-print'><code>mlr3::Learner$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="reset"><a href='../../mlr3/html/Learner.html#method-Learner-reset'><code>mlr3::Learner$reset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="selected_features"><a href='../../mlr3/html/Learner.html#method-Learner-selected_features'><code>mlr3::Learner$selected_features()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="train"><a href='../../mlr3/html/Learner.html#method-Learner-train'><code>mlr3::Learner$train()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="LearnerRegr" data-id="predict_newdata_fast"><a href='../../mlr3/html/LearnerRegr.html#method-LearnerRegr-predict_newdata_fast'><code>mlr3::LearnerRegr$predict_newdata_fast()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerRegrBam-new"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerRegrBam-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerRegrBam$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerRegrBam-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerRegrBam-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerRegrBam$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
