% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner_earth_regr_earth.R
\name{mlr_learners_regr.earth}
\alias{mlr_learners_regr.earth}
\alias{LearnerRegrEarth}
\title{Implementation of Multivariate Adaptive Regression Splines}
\description{
This is an alternative implementation of MARS (Multivariate Adaptive Regression Splines).
MARS is trademarked and thus not used as the name. The name "earth" stands for
"Enhanced Adaptive Regression Through Hinges".

Calls \code{\link[earth:earth]{earth::earth()}} from \CRANpkg{earth}.
}
\details{
Methods for variance estimations are not yet implemented.
}
\section{Dictionary}{

This \link{Learner} can be instantiated via the \link[mlr3misc:Dictionary]{dictionary} \link{mlr_learners} or with the associated sugar function \code{\link[=lrn]{lrn()}}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{mlr_learners$get("regr.earth")
lrn("regr.earth")
}\if{html}{\out{</div>}}
}

\section{Meta Information}{

\itemize{
\item Task type: \dQuote{regr}
\item Predict Types: \dQuote{response}, \dQuote{se}
\item Feature Types: \dQuote{integer}, \dQuote{numeric}, \dQuote{factor}
\item Required Packages: \CRANpkg{mlr3}, \CRANpkg{mlr3extralearners}, \CRANpkg{earth}
}
}

\section{Parameters}{
\tabular{lllll}{
   Id \tab Type \tab Default \tab Levels \tab Range \cr
   wp \tab untyped \tab  \tab  \tab - \cr
   offset \tab untyped \tab  \tab  \tab - \cr
   keepxy \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   trace \tab character \tab 0 \tab 0, .3, .5, 1, 2, 3, 4, 5 \tab - \cr
   degree \tab integer \tab 1 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   penalty \tab numeric \tab 2 \tab  \tab \eqn{[-1, \infty)}{[-1, Inf)} \cr
   nk \tab untyped \tab  \tab  \tab - \cr
   thresh \tab numeric \tab 0.001 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   minspan \tab numeric \tab 0 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   endspan \tab numeric \tab 0 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   newvar.penalty \tab numeric \tab 0 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   fast.k \tab integer \tab 20 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   fast.beta \tab integer \tab 1 \tab  \tab \eqn{[0, 1]}{[0, 1]} \cr
   linpreds \tab untyped \tab FALSE \tab  \tab - \cr
   allowed \tab untyped \tab - \tab  \tab - \cr
   pmethod \tab character \tab backward \tab backward, none, exhaustive, forward, seqrep, cv \tab - \cr
   nprune \tab integer \tab - \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   nfold \tab integer \tab 0 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   ncross \tab integer \tab 1 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   stratify \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
   varmod.method \tab character \tab none \tab none, const, lm, rlm, earth, gam, power, power0, x.lm, x.rlm, \link{...} \tab - \cr
   varmod.exponent \tab numeric \tab 1 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   varmod.conv \tab numeric \tab 1 \tab  \tab \eqn{[0, 1]}{[0, 1]} \cr
   varmod.clamp \tab numeric \tab 0.1 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   varmod.minspan \tab numeric \tab -3 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   Scale.y \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   Adjust.endspan \tab numeric \tab 2 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   Auto.linpreds \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
   Force.weights \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   Use.beta.cache \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
   Force.xtx.prune \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   Get.leverages \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
   Exhaustive.tol \tab numeric \tab 1e-10 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
}
}

\examples{
learner = mlr3::lrn("regr.earth")
print(learner)

# available parameters:
learner$param_set$ids()
}
\references{
Milborrow, Stephen, Hastie, T, Tibshirani, R (2014).
\dQuote{Earth: multivariate adaptive regression spline models.}
\emph{R package version}, \bold{3}(7).

Friedman, H J (1991).
\dQuote{Multivariate adaptive regression splines.}
\emph{The annals of statistics}, \bold{19}(1), 1--67.
}
\seealso{
\itemize{
\item \link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}.
\item \code{as.data.table(mlr_learners)} for a table of available \link[=Learner]{Learners} in the running session (depending on the loaded packages).
\item Chapter in the \href{https://mlr3book.mlr-org.com/}{mlr3book}: \url{https://mlr3book.mlr-org.com/basics.html#learners}
\item \CRANpkg{mlr3learners} for a selection of recommended learners.
\item \CRANpkg{mlr3cluster} for unsupervised clustering learners.
\item \CRANpkg{mlr3pipelines} to combine learners with pre- and postprocessing steps.
\item \CRANpkg{mlr3tuning} for tuning of hyperparameters, \CRANpkg{mlr3tuningspaces} for established default tuning spaces.
}
}
\author{
pkopper
}
\section{Super classes}{
\code{\link[mlr3:Learner]{mlr3::Learner}} -> \code{\link[mlr3:LearnerRegr]{mlr3::LearnerRegr}} -> \code{LearnerRegrEarth}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LearnerRegrEarth-new}{\code{LearnerRegrEarth$new()}}
\item \href{#method-LearnerRegrEarth-clone}{\code{LearnerRegrEarth$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="base_learner"><a href='../../mlr3/html/Learner.html#method-Learner-base_learner'><code>mlr3::Learner$base_learner()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="format"><a href='../../mlr3/html/Learner.html#method-Learner-format'><code>mlr3::Learner$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="help"><a href='../../mlr3/html/Learner.html#method-Learner-help'><code>mlr3::Learner$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict"><a href='../../mlr3/html/Learner.html#method-Learner-predict'><code>mlr3::Learner$predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict_newdata"><a href='../../mlr3/html/Learner.html#method-Learner-predict_newdata'><code>mlr3::Learner$predict_newdata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="print"><a href='../../mlr3/html/Learner.html#method-Learner-print'><code>mlr3::Learner$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="reset"><a href='../../mlr3/html/Learner.html#method-Learner-reset'><code>mlr3::Learner$reset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="train"><a href='../../mlr3/html/Learner.html#method-Learner-train'><code>mlr3::Learner$train()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerRegrEarth-new"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerRegrEarth-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerRegrEarth$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerRegrEarth-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerRegrEarth-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerRegrEarth$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
