% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner_ExhaustiveSearch_regr_lm.R
\name{mlr_learners_regr.exhaustive_search}
\alias{mlr_learners_regr.exhaustive_search}
\alias{LearnerRegrExhaustiveSearch}
\title{Linear Regression with Exhaustive Search}
\description{
Linear Regression, where the features used in the model are selected by running Exhaustive Search.

Calls \code{\link[ExhaustiveSearch:ExhaustiveSearch]{ExhaustiveSearch::ExhaustiveSearch()}} from \CRANpkg{ExhaustiveSearch}.
}
\section{Initial parameter values}{

\itemize{
\item \code{family}:
\itemize{
\item Actual default: NULL
\item Adjusted default: "gaussian"
\item Reason for change: To comply with mlr3 architecture,
we differentiate between classification and regression learners.
}
\item \code{nThreads}:
\itemize{
\item Actual default: NULL
\item Adjusted default: 1
\item Reason for change: Suppressing the automatic internal parallelization if
\code{cv.folds} > 0.
}
\item \code{quietly}:
\itemize{
\item Actual default: FALSE
\item Adjusted default: TRUE
\item Reason for change: Suppression of constant printing to console
}
}
}

\section{Dictionary}{

This \link[mlr3:Learner]{Learner} can be instantiated via \link[mlr3:mlr_sugar]{lrn()}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{lrn("regr.exhaustive_search")
}\if{html}{\out{</div>}}
}

\section{Meta Information}{

\itemize{
\item Task type: \dQuote{regr}
\item Predict Types: \dQuote{response}, \dQuote{se}
\item Feature Types: \dQuote{logical}, \dQuote{integer}, \dQuote{numeric}
\item Required Packages: \CRANpkg{mlr3}, \CRANpkg{mlr3extralearners}, \CRANpkg{ExhaustiveSearch}
}
}

\section{Parameters}{
\tabular{lllll}{
   Id \tab Type \tab Default \tab Levels \tab Range \cr
   family \tab character \tab - \tab gaussian, binomial \tab - \cr
   performanceMeasure \tab character \tab - \tab MSE, AIC \tab - \cr
   combsUpTo \tab integer \tab - \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   nResults \tab integer \tab 5000 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   nThreads \tab integer \tab - \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   testSetIDs \tab integer \tab - \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   errorVal \tab untyped \tab -1 \tab  \tab - \cr
   quietly \tab logical \tab - \tab TRUE, FALSE \tab - \cr
   checkLarge \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
}
}

\examples{
\dontshow{if (learner_is_runnable("regr.exhaustive_search")) withAutoprint(\{ # examplesIf}
# define learner
learner = LearnerRegrExhaustiveSearch$new()
learner$predict_type = "se"

# define task
tsk_cars = tsk("mtcars")

# train learner
learner$train(tsk_cars)

# extract selected features
learner$selected_features()

# predict on training task
learner$predict(tsk_cars)
\dontshow{\}) # examplesIf}
}
\seealso{
\itemize{
\item \link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}.
\item \code{as.data.table(mlr_learners)} for a table of available \link[mlr3:Learner]{Learners} in the running session (depending on the loaded packages).
\item Chapter in the \href{https://mlr3book.mlr-org.com/}{mlr3book}: \url{https://mlr3book.mlr-org.com/basics.html#learners}
\item \CRANpkg{mlr3learners} for a selection of recommended learners.
\item \CRANpkg{mlr3cluster} for unsupervised clustering learners.
\item \CRANpkg{mlr3pipelines} to combine learners with pre- and postprocessing steps.
\item \CRANpkg{mlr3tuning} for tuning of hyperparameters, \CRANpkg{mlr3tuningspaces} for established default tuning spaces.
}
}
\author{
ngerman
}
\section{Super classes}{
\code{\link[mlr3:Learner]{mlr3::Learner}} -> \code{\link[mlr3:LearnerRegr]{mlr3::LearnerRegr}} -> \code{LearnerRegrExhaustiveSearch}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LearnerRegrExhaustiveSearch-new}{\code{LearnerRegrExhaustiveSearch$new()}}
\item \href{#method-LearnerRegrExhaustiveSearch-selected_features}{\code{LearnerRegrExhaustiveSearch$selected_features()}}
\item \href{#method-LearnerRegrExhaustiveSearch-clone}{\code{LearnerRegrExhaustiveSearch$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="base_learner"><a href='../../mlr3/html/Learner.html#method-Learner-base_learner'><code>mlr3::Learner$base_learner()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="configure"><a href='../../mlr3/html/Learner.html#method-Learner-configure'><code>mlr3::Learner$configure()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="encapsulate"><a href='../../mlr3/html/Learner.html#method-Learner-encapsulate'><code>mlr3::Learner$encapsulate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="format"><a href='../../mlr3/html/Learner.html#method-Learner-format'><code>mlr3::Learner$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="help"><a href='../../mlr3/html/Learner.html#method-Learner-help'><code>mlr3::Learner$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict"><a href='../../mlr3/html/Learner.html#method-Learner-predict'><code>mlr3::Learner$predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict_newdata"><a href='../../mlr3/html/Learner.html#method-Learner-predict_newdata'><code>mlr3::Learner$predict_newdata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="print"><a href='../../mlr3/html/Learner.html#method-Learner-print'><code>mlr3::Learner$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="reset"><a href='../../mlr3/html/Learner.html#method-Learner-reset'><code>mlr3::Learner$reset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="train"><a href='../../mlr3/html/Learner.html#method-Learner-train'><code>mlr3::Learner$train()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="LearnerRegr" data-id="predict_newdata_fast"><a href='../../mlr3/html/LearnerRegr.html#method-LearnerRegr-predict_newdata_fast'><code>mlr3::LearnerRegr$predict_newdata_fast()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerRegrExhaustiveSearch-new"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerRegrExhaustiveSearch-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerRegrExhaustiveSearch$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerRegrExhaustiveSearch-selected_features"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerRegrExhaustiveSearch-selected_features}{}}}
\subsection{Method \code{selected_features()}}{
Extracts selected features of this learner.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerRegrExhaustiveSearch$selected_features()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerRegrExhaustiveSearch-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerRegrExhaustiveSearch-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerRegrExhaustiveSearch$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
