% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner_SpatialML_regr_gwrf.R
\name{mlr_learners_regr.gwrf}
\alias{mlr_learners_regr.gwrf}
\alias{LearnerRegrGWRF}
\title{Regression Geographically Weighted Random Forest Learner}
\description{
Geographically weighted random forest for regression.
Calls \code{SpatialML::grf()} from package SpatialML.

The learner requires a spatial task with coordinates, e.g., a \code{TaskRegrST} from package \CRANpkg{mlr3spatiotempcv}.
SpatialML has been archived on CRAN.
It can be installed with \code{remotes::install_version("SpatialML", version = "0.1.6")}.

Predictions are computed with \code{SpatialML::predict.grf()} using only the local models,
i.e., \code{local.w = 1} and \code{global.w = 0}.
}
\section{Initial parameter values}{

\itemize{
\item \code{kernel} is initialized to \code{"adaptive"} because \code{SpatialML::grf()} has no default.
\item \code{print.results} is set to \code{FALSE} and not exposed as a hyperparameter.
}
}

\section{Custom mlr3 parameters}{

\itemize{
\item \code{min.node.size}, \code{max.depth}, \code{replace}, and \code{sample.fraction} are passed to \code{ranger::ranger()}.
}
}

\section{Dictionary}{

This \link[mlr3:Learner]{Learner} can be instantiated via \link[mlr3:lrn]{lrn()}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{lrn("regr.gwrf")
}\if{html}{\out{</div>}}
}

\section{Meta Information}{

\itemize{
\item Task type: \dQuote{regr}
\item Predict Types: \dQuote{response}
\item Feature Types: \dQuote{integer}, \dQuote{numeric}, \dQuote{factor}
\item Required Packages: \CRANpkg{mlr3}, \CRANpkg{mlr3extralearners}, \CRANpkg{SpatialML}
}
}

\section{Parameters}{
\tabular{lllll}{
   Id \tab Type \tab Default \tab Levels \tab Range \cr
   bw \tab numeric \tab - \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   kernel \tab character \tab - \tab adaptive, fixed \tab - \cr
   ntree \tab integer \tab 500 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   mtry \tab integer \tab - \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   importance \tab character \tab impurity \tab impurity, permutation \tab - \cr
   nthreads \tab integer \tab - \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   geo.weighted \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
   min.node.size \tab integer \tab 5 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   max.depth \tab integer \tab - \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   replace \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
   sample.fraction \tab numeric \tab - \tab  \tab \eqn{[0, 1]}{[0, 1]} \cr
}
}

\references{
Georganos, Stefanos, Grippa, Tais, Gadiaga N, Assane, Linard, Catherine, Lennert, Moritz, Vanhuysse, Sabine, Mboga, Nicholus, Wolff, Eleonore, Kalogirou, Stamatis (2021).
\dQuote{Geographical random forests: a spatial extension of the random forest algorithm to address spatial heterogeneity in remote sensing and population modelling.}
\emph{Geocarto International}, \bold{36}(2), 121--136.
\doi{10.1080/10106049.2019.1595177}.
}
\seealso{
\itemize{
\item \link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}.
\item \code{as.data.table(mlr_learners)} for a table of available \link[mlr3:Learner]{Learners} in the running session (depending on the loaded packages).
\item Chapter in the \href{https://mlr3book.mlr-org.com/}{mlr3book}: \url{https://mlr3book.mlr-org.com/chapters/chapter2/data_and_basic_modeling.html#sec-learners}
\item \CRANpkg{mlr3learners} for a selection of recommended learners.
\item \CRANpkg{mlr3cluster} for unsupervised clustering learners.
\item \CRANpkg{mlr3pipelines} to combine learners with pre- and postprocessing steps.
\item \CRANpkg{mlr3tuning} for tuning of hyperparameters, \CRANpkg{mlr3tuningspaces} for established default tuning spaces.
}
}
\author{
Manh Hung LE
}
\section{Super classes}{
\code{\link[mlr3:Learner]{mlr3::Learner}} -> \code{\link[mlr3:LearnerRegr]{mlr3::LearnerRegr}} -> \code{LearnerRegrGWRF}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-LearnerRegrGWRF-initialize}{\code{LearnerRegrGWRF$new()}}
    \item \href{#method-LearnerRegrGWRF-importance}{\code{LearnerRegrGWRF$importance()}}
    \item \href{#method-LearnerRegrGWRF-oob_error}{\code{LearnerRegrGWRF$oob_error()}}
    \item \href{#method-LearnerRegrGWRF-clone}{\code{LearnerRegrGWRF$clone()}}
  }
}
\if{html}{\out{<details><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="base_learner"><a href='../../mlr3/html/Learner.html#method-Learner-base_learner'><code>mlr3::Learner$base_learner()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="configure"><a href='../../mlr3/html/Learner.html#method-Learner-configure'><code>mlr3::Learner$configure()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="encapsulate"><a href='../../mlr3/html/Learner.html#method-Learner-encapsulate'><code>mlr3::Learner$encapsulate()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="format"><a href='../../mlr3/html/Learner.html#method-Learner-format'><code>mlr3::Learner$format()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="help"><a href='../../mlr3/html/Learner.html#method-Learner-help'><code>mlr3::Learner$help()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict"><a href='../../mlr3/html/Learner.html#method-Learner-predict'><code>mlr3::Learner$predict()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict_newdata"><a href='../../mlr3/html/Learner.html#method-Learner-predict_newdata'><code>mlr3::Learner$predict_newdata()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="print"><a href='../../mlr3/html/Learner.html#method-Learner-print'><code>mlr3::Learner$print()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="reset"><a href='../../mlr3/html/Learner.html#method-Learner-reset'><code>mlr3::Learner$reset()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="selected_features"><a href='../../mlr3/html/Learner.html#method-Learner-selected_features'><code>mlr3::Learner$selected_features()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="train"><a href='../../mlr3/html/Learner.html#method-Learner-train'><code>mlr3::Learner$train()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="LearnerRegr" data-id="predict_newdata_fast"><a href='../../mlr3/html/LearnerRegr.html#method-LearnerRegr-predict_newdata_fast'><code>mlr3::LearnerRegr$predict_newdata_fast()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerRegrGWRF-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerRegrGWRF-initialize}{}}}
\subsection{\code{LearnerRegrGWRF$new()}}{
  Creates a new instance of this \link[R6:R6Class]{R6} class.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LearnerRegrGWRF$new()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerRegrGWRF-importance"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerRegrGWRF-importance}{}}}
\subsection{\code{LearnerRegrGWRF$importance()}}{
  The importance scores are the column means of the slot \code{Local.Variable.Importance}.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LearnerRegrGWRF$importance()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Named \code{numeric()}.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerRegrGWRF-oob_error"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerRegrGWRF-oob_error}{}}}
\subsection{\code{LearnerRegrGWRF$oob_error()}}{
  The OOB error is extracted from the global model slot \code{Global.Model$prediction.error}.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LearnerRegrGWRF$oob_error()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    \code{numeric(1)}.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerRegrGWRF-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerRegrGWRF-clone}{}}}
\subsection{\code{LearnerRegrGWRF$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LearnerRegrGWRF$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
