% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner_aorsf_surv_aorsf.R
\name{mlr_learners_surv.aorsf}
\alias{mlr_learners_surv.aorsf}
\alias{LearnerSurvAorsf}
\title{Accelerated Oblique Random Survival Forest Learner}
\description{
Accelerated oblique random survival forest.
Calls \code{\link[aorsf:orsf]{aorsf::orsf()}} from \CRANpkg{aorsf}.
Note that although the learner has the property \code{"missing"} and it can in principle deal with missing values,
the behaviour has to be configured using the parameter \code{na_action}.
}
\section{Dictionary}{

This \link{Learner} can be instantiated via the \link[mlr3misc:Dictionary]{dictionary} \link{mlr_learners} or with the associated sugar function \code{\link[=lrn]{lrn()}}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{mlr_learners$get("surv.aorsf")
lrn("surv.aorsf")
}\if{html}{\out{</div>}}
}

\section{Meta Information}{

\itemize{
\item Task type: \dQuote{surv}
\item Predict Types: \dQuote{crank}, \dQuote{distr}
\item Feature Types: \dQuote{integer}, \dQuote{numeric}, \dQuote{factor}, \dQuote{ordered}
\item Required Packages: \CRANpkg{mlr3}, \CRANpkg{mlr3proba}, \CRANpkg{mlr3extralearners}, \CRANpkg{aorsf}, \CRANpkg{pracma}
}
}

\section{Parameters}{
\tabular{lllll}{
   Id \tab Type \tab Default \tab Levels \tab Range \cr
   n_tree \tab integer \tab 500 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   n_split \tab integer \tab 5 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   n_retry \tab integer \tab 3 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   mtry \tab integer \tab NULL \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   mtry_ratio \tab numeric \tab - \tab  \tab \eqn{[0, 1]}{[0, 1]} \cr
   control_type \tab character \tab fast \tab fast, cph, net \tab - \cr
   control_fast_do_scale \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
   control_fast_method \tab character \tab efron \tab efron, breslow \tab - \cr
   control_cph_method \tab character \tab efron \tab efron, breslow \tab - \cr
   control_cph_eps \tab numeric \tab 1e-09 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   control_cph_iter_max \tab integer \tab 20 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   control_net_alpha \tab numeric \tab 0.5 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   control_net_df_target \tab integer \tab NULL \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   leaf_min_events \tab integer \tab 1 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   leaf_min_obs \tab integer \tab 5 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   split_min_events \tab integer \tab 5 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   split_min_obs \tab integer \tab 10 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   split_min_stat \tab numeric \tab 3.841459 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   oobag_pred_type \tab character \tab surv \tab none, surv, risk, chf \tab - \cr
   importance \tab character \tab anova \tab none, anova, negate, permute \tab - \cr
   oobag_pred_horizon \tab numeric \tab NULL \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   oobag_eval_every \tab integer \tab NULL \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   attach_data \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
   verbose_progress \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   na_action \tab character \tab fail \tab fail, omit, impute_meanmode \tab - \cr
}
}

\section{Initial parameter values}{

\itemize{
\item \code{mtry}:
\itemize{
\item This hyperparameter can alternatively be set via the added hyperparameter \code{mtry_ratio}
as \code{mtry = max(ceiling(mtry_ratio * n_features), 1)}.
Note that \code{mtry} and \code{mtry_ratio} are mutually exclusive.
}
}
}

\examples{
learner = mlr3::lrn("surv.aorsf")
print(learner)

# available parameters:
learner$param_set$ids()
}
\references{
Jaeger BC, Long DL, Long DM, Sims M, Szychowski JM, Min Y, Mcclure LA, Howard G, Simon N (2019).
\dQuote{Oblique random survival forests.}
\emph{The Annals of Applied Statistics}, \bold{13}(3).
\doi{10.1214/19-aoas1261}.

Jaeger BC, Welden S, Lenoir K, Speiser JL, Segar MW, Pandey A, Pajewski NM (2022).
\dQuote{Accelerated and interpretable oblique random survival forests.}
\emph{arXiv preprint arXiv:2208.01129}.
}
\seealso{
\itemize{
\item \link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}.
\item \code{as.data.table(mlr_learners)} for a table of available \link[=Learner]{Learners} in the running session (depending on the loaded packages).
\item Chapter in the \href{https://mlr3book.mlr-org.com/}{mlr3book}: \url{https://mlr3book.mlr-org.com/basics.html#learners}
\item \CRANpkg{mlr3learners} for a selection of recommended learners.
\item \CRANpkg{mlr3cluster} for unsupervised clustering learners.
\item \CRANpkg{mlr3pipelines} to combine learners with pre- and postprocessing steps.
\item \CRANpkg{mlr3tuning} for tuning of hyperparameters, \CRANpkg{mlr3tuningspaces} for established default tuning spaces.
}
}
\author{
bcjaeger
}
\section{Super classes}{
\code{\link[mlr3:Learner]{mlr3::Learner}} -> \code{\link[mlr3proba:LearnerSurv]{mlr3proba::LearnerSurv}} -> \code{LearnerSurvAorsf}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LearnerSurvAorsf-new}{\code{LearnerSurvAorsf$new()}}
\item \href{#method-LearnerSurvAorsf-oob_error}{\code{LearnerSurvAorsf$oob_error()}}
\item \href{#method-LearnerSurvAorsf-importance}{\code{LearnerSurvAorsf$importance()}}
\item \href{#method-LearnerSurvAorsf-clone}{\code{LearnerSurvAorsf$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="base_learner"><a href='../../mlr3/html/Learner.html#method-Learner-base_learner'><code>mlr3::Learner$base_learner()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="format"><a href='../../mlr3/html/Learner.html#method-Learner-format'><code>mlr3::Learner$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="help"><a href='../../mlr3/html/Learner.html#method-Learner-help'><code>mlr3::Learner$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict"><a href='../../mlr3/html/Learner.html#method-Learner-predict'><code>mlr3::Learner$predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict_newdata"><a href='../../mlr3/html/Learner.html#method-Learner-predict_newdata'><code>mlr3::Learner$predict_newdata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="print"><a href='../../mlr3/html/Learner.html#method-Learner-print'><code>mlr3::Learner$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="reset"><a href='../../mlr3/html/Learner.html#method-Learner-reset'><code>mlr3::Learner$reset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="train"><a href='../../mlr3/html/Learner.html#method-Learner-train'><code>mlr3::Learner$train()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerSurvAorsf-new"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerSurvAorsf-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerSurvAorsf$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerSurvAorsf-oob_error"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerSurvAorsf-oob_error}{}}}
\subsection{Method \code{oob_error()}}{
OOB concordance error extracted from the model slot
\code{eval_oobag$stat_values}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerSurvAorsf$oob_error()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{numeric()}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerSurvAorsf-importance"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerSurvAorsf-importance}{}}}
\subsection{Method \code{importance()}}{
The importance scores are extracted from the model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerSurvAorsf$importance()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Named \code{numeric()}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerSurvAorsf-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerSurvAorsf-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerSurvAorsf$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
