% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner_BART_surv_bart.R
\name{mlr_learners_surv.bart}
\alias{mlr_learners_surv.bart}
\alias{LearnerSurvLearnerSurvBART}
\title{Survival Bayesian Additive Regression Trees Learner}
\description{
Fits a Bayesian Additive Regression Trees (BART) learner to right-censored
survival data. Calls \code{\link[BART:surv.bart]{BART::mc.surv.bart()}} from \CRANpkg{BART}.
}
\section{Prediction types}{

This learner returns two prediction types:
\enumerate{
\item \code{distr}: a 3d survival array with observations as 1st dimension, time
points as 2nd and the posterior draws as 3rd dimension.
Calculated using the internal \code{predict.survbart()} function.
\item \code{crank}: the expected mortality using \code{\link[mlr3proba:dot-surv_return]{mlr3proba::.surv_return()}}. The parameter
\code{which.curve} decides which posterior draw (3rd dimension) will be used for the
calculation of the expected mortality. Note that the median posterior is
by default used for the calculation of survival measures that require a \code{distr}
prediction, see more info on \link[mlr3proba:PredictionSurv]{PredictionSurv}.
}
}

\section{Initial parameter values}{

\itemize{
\item \code{mc.cores} is initialized to 1 to avoid threading conflicts with \CRANpkg{future}.
}
}

\section{Custom mlr3 parameters}{

\itemize{
\item \code{quiet} allows to suppress messages generated by the wrapped C++ code. Is
initialized to \code{TRUE}.
\item \code{importance} allows to choose the type of importance. Default is \code{count},
see documentation of method \verb{$importance()} for more details.
\item \code{which.curve} allows to choose which posterior draw will be used for the
calculation of the \code{crank} prediction. If between (0,1) it is taken as the
quantile of the curves otherwise if greater than 1 it is taken as the curve
index, can also be 'mean'. By default the \strong{median posterior} is used,
i.e. \code{which.curve} is 0.5.
}
}

\section{Dictionary}{

This \link[mlr3:Learner]{Learner} can be instantiated via \link[mlr3:mlr_sugar]{lrn()}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{lrn("surv.bart")
}\if{html}{\out{</div>}}
}

\section{Meta Information}{

\itemize{
\item Task type: \dQuote{surv}
\item Predict Types: \dQuote{crank}, \dQuote{distr}
\item Feature Types: \dQuote{logical}, \dQuote{integer}, \dQuote{numeric}
\item Required Packages: \CRANpkg{mlr3}, \CRANpkg{mlr3proba}, \CRANpkg{BART}
}
}

\section{Parameters}{
\tabular{lllll}{
   Id \tab Type \tab Default \tab Levels \tab Range \cr
   K \tab numeric \tab NULL \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   events \tab untyped \tab NULL \tab  \tab - \cr
   ztimes \tab untyped \tab NULL \tab  \tab - \cr
   zdelta \tab untyped \tab NULL \tab  \tab - \cr
   sparse \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   theta \tab numeric \tab 0 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   omega \tab numeric \tab 1 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   a \tab numeric \tab 0.5 \tab  \tab \eqn{[0.5, 1]}{[0.5, 1]} \cr
   b \tab numeric \tab 1 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   augment \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   rho \tab numeric \tab NULL \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   usequants \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   rm.const \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
   type \tab character \tab pbart \tab pbart, lbart \tab - \cr
   ntype \tab integer \tab - \tab  \tab \eqn{[1, 3]}{[1, 3]} \cr
   k \tab numeric \tab 2 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   power \tab numeric \tab 2 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   base \tab numeric \tab 0.95 \tab  \tab \eqn{[0, 1]}{[0, 1]} \cr
   offset \tab numeric \tab NULL \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   ntree \tab integer \tab 50 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   numcut \tab integer \tab 100 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   ndpost \tab integer \tab 1000 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   nskip \tab integer \tab 250 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   keepevery \tab integer \tab 10 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   printevery \tab integer \tab 100 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   seed \tab integer \tab 99 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   mc.cores \tab integer \tab 2 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   nice \tab integer \tab 19 \tab  \tab \eqn{[0, 19]}{[0, 19]} \cr
   openmp \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
   quiet \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
   importance \tab character \tab count \tab count, prob \tab - \cr
   which.curve \tab numeric \tab - \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
}
}

\examples{
lrn("surv.bart")
}
\references{
Sparapani, Rodney, Spanbauer, Charles, McCulloch, Robert (2021).
\dQuote{Nonparametric machine learning and efficient computation with bayesian additive regression trees: the BART R package.}
\emph{Journal of Statistical Software}, \bold{97}, 1--66.

Chipman, A H, George, I E, McCulloch, E R (2010).
\dQuote{BART: Bayesian additive regression trees.}
\emph{The Annals of Applied Statistics}, \bold{4}(1), 266--298.
}
\seealso{
\itemize{
\item \link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}.
\item \code{as.data.table(mlr_learners)} for a table of available \link[mlr3:Learner]{Learners} in the running session (depending on the loaded packages).
\item Chapter in the \href{https://mlr3book.mlr-org.com/}{mlr3book}: \url{https://mlr3book.mlr-org.com/basics.html#learners}
\item \CRANpkg{mlr3learners} for a selection of recommended learners.
\item \CRANpkg{mlr3cluster} for unsupervised clustering learners.
\item \CRANpkg{mlr3pipelines} to combine learners with pre- and postprocessing steps.
\item \CRANpkg{mlr3tuning} for tuning of hyperparameters, \CRANpkg{mlr3tuningspaces} for established default tuning spaces.
}
}
\author{
bblodfon
}
\section{Super classes}{
\code{\link[mlr3:Learner]{mlr3::Learner}} -> \code{\link[mlr3proba:LearnerSurv]{mlr3proba::LearnerSurv}} -> \code{LearnerSurvLearnerSurvBART}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LearnerSurvLearnerSurvBART-new}{\code{LearnerSurvLearnerSurvBART$new()}}
\item \href{#method-LearnerSurvLearnerSurvBART-importance}{\code{LearnerSurvLearnerSurvBART$importance()}}
\item \href{#method-LearnerSurvLearnerSurvBART-clone}{\code{LearnerSurvLearnerSurvBART$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="base_learner"><a href='../../mlr3/html/Learner.html#method-Learner-base_learner'><code>mlr3::Learner$base_learner()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="configure"><a href='../../mlr3/html/Learner.html#method-Learner-configure'><code>mlr3::Learner$configure()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="encapsulate"><a href='../../mlr3/html/Learner.html#method-Learner-encapsulate'><code>mlr3::Learner$encapsulate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="format"><a href='../../mlr3/html/Learner.html#method-Learner-format'><code>mlr3::Learner$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="help"><a href='../../mlr3/html/Learner.html#method-Learner-help'><code>mlr3::Learner$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict"><a href='../../mlr3/html/Learner.html#method-Learner-predict'><code>mlr3::Learner$predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict_newdata"><a href='../../mlr3/html/Learner.html#method-Learner-predict_newdata'><code>mlr3::Learner$predict_newdata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="print"><a href='../../mlr3/html/Learner.html#method-Learner-print'><code>mlr3::Learner$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="reset"><a href='../../mlr3/html/Learner.html#method-Learner-reset'><code>mlr3::Learner$reset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="selected_features"><a href='../../mlr3/html/Learner.html#method-Learner-selected_features'><code>mlr3::Learner$selected_features()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="train"><a href='../../mlr3/html/Learner.html#method-Learner-train'><code>mlr3::Learner$train()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerSurvLearnerSurvBART-new"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerSurvLearnerSurvBART-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerSurvLearnerSurvBART$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerSurvLearnerSurvBART-importance"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerSurvLearnerSurvBART-importance}{}}}
\subsection{Method \code{importance()}}{
Two types of importance scores are supported based on the value
of the parameter \code{importance}:
\enumerate{
\item \code{prob}: The mean selection probability of each feature in the trees,
extracted from the slot \code{varprob.mean}.
If \code{sparse = FALSE} (default), this is a fixed constant.
Recommended to use this option when \code{sparse = TRUE}.
\item \code{count}: The mean observed count of each feature in the trees (average
number of times the feature was used in a tree decision rule across all
posterior draws), extracted from the slot \code{varcount.mean}.
This is the default importance scores.
}

In both cases, higher values signify more important variables.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerSurvLearnerSurvBART$importance()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Named \code{numeric()}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerSurvLearnerSurvBART-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerSurvLearnerSurvBART-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerSurvLearnerSurvBART$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
