% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner_mboost_surv_glmboost.R
\name{mlr_learners_surv.glmboost}
\alias{mlr_learners_surv.glmboost}
\alias{LearnerSurvGLMBoost}
\title{Boosted Generalized Linear Survival Learner}
\description{
Fits a generalized linear survival model using a boosting algorithm.
Calls \code{\link[mboost:glmboost]{mboost::glmboost()}} from \CRANpkg{mboost}.
}
\details{
\code{distr} prediction made by \code{\link[mboost:survFit]{mboost::survFit()}}.
}
\section{Dictionary}{

This \link{Learner} can be instantiated via the \link[mlr3misc:Dictionary]{dictionary} \link{mlr_learners} or with the associated sugar function \code{\link[=lrn]{lrn()}}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{mlr_learners$get("surv.glmboost")
lrn("surv.glmboost")
}\if{html}{\out{</div>}}
}

\section{Meta Information}{

\itemize{
\item Task type: \dQuote{surv}
\item Predict Types: \dQuote{crank}, \dQuote{distr}, \dQuote{lp}
\item Feature Types: \dQuote{logical}, \dQuote{integer}, \dQuote{numeric}, \dQuote{factor}
\item Required Packages: \CRANpkg{mlr3}, \CRANpkg{mlr3proba}, \CRANpkg{mlr3extralearners}, \CRANpkg{mboost}, \CRANpkg{pracma}
}
}

\section{Parameters}{
\tabular{lllll}{
   Id \tab Type \tab Default \tab Levels \tab Range \cr
   offset \tab numeric \tab - \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   family \tab character \tab coxph \tab coxph, weibull, loglog, lognormal, gehan, cindex, custom \tab - \cr
   custom.family \tab untyped \tab - \tab  \tab - \cr
   nuirange \tab untyped \tab c  , 0  , 100 \tab  \tab - \cr
   center \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
   mstop \tab integer \tab 100 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   nu \tab numeric \tab 0.1 \tab  \tab \eqn{[0, 1]}{[0, 1]} \cr
   risk \tab character \tab inbag \tab inbag, oobag, none \tab - \cr
   oobweights \tab untyped \tab  \tab  \tab - \cr
   stopintern \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   trace \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   sigma \tab numeric \tab 0.1 \tab  \tab \eqn{[0, 1]}{[0, 1]} \cr
   ipcw \tab untyped \tab 1 \tab  \tab - \cr
   na.action \tab untyped \tab ::     , stats  , na.omit \tab  \tab - \cr
   contrasts.arg \tab untyped \tab - \tab  \tab - \cr
}
}

\examples{
learner = mlr3::lrn("surv.glmboost")
print(learner)

# available parameters:
learner$param_set$ids()
}
\references{
Bühlmann, Peter, Yu, Bin (2003).
\dQuote{Boosting with the L 2 loss: regression and classification.}
\emph{Journal of the American Statistical Association}, \bold{98}(462), 324--339.
}
\seealso{
\itemize{
\item \link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}.
\item \code{as.data.table(mlr_learners)} for a table of available \link[=Learner]{Learners} in the running session (depending on the loaded packages).
\item Chapter in the \href{https://mlr3book.mlr-org.com/}{mlr3book}: \url{https://mlr3book.mlr-org.com/basics.html#learners}
\item \CRANpkg{mlr3learners} for a selection of recommended learners.
\item \CRANpkg{mlr3cluster} for unsupervised clustering learners.
\item \CRANpkg{mlr3pipelines} to combine learners with pre- and postprocessing steps.
\item \CRANpkg{mlr3tuning} for tuning of hyperparameters, \CRANpkg{mlr3tuningspaces} for established default tuning spaces.
}
}
\author{
RaphaelS1
}
\section{Super classes}{
\code{\link[mlr3:Learner]{mlr3::Learner}} -> \code{\link[mlr3proba:LearnerSurv]{mlr3proba::LearnerSurv}} -> \code{LearnerSurvGLMBoost}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LearnerSurvGLMBoost-new}{\code{LearnerSurvGLMBoost$new()}}
\item \href{#method-LearnerSurvGLMBoost-clone}{\code{LearnerSurvGLMBoost$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="base_learner"><a href='../../mlr3/html/Learner.html#method-Learner-base_learner'><code>mlr3::Learner$base_learner()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="format"><a href='../../mlr3/html/Learner.html#method-Learner-format'><code>mlr3::Learner$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="help"><a href='../../mlr3/html/Learner.html#method-Learner-help'><code>mlr3::Learner$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict"><a href='../../mlr3/html/Learner.html#method-Learner-predict'><code>mlr3::Learner$predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict_newdata"><a href='../../mlr3/html/Learner.html#method-Learner-predict_newdata'><code>mlr3::Learner$predict_newdata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="print"><a href='../../mlr3/html/Learner.html#method-Learner-print'><code>mlr3::Learner$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="reset"><a href='../../mlr3/html/Learner.html#method-Learner-reset'><code>mlr3::Learner$reset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="train"><a href='../../mlr3/html/Learner.html#method-Learner-train'><code>mlr3::Learner$train()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerSurvGLMBoost-new"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerSurvGLMBoost-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
Importance is supported but fails tests as internally data
is coerced to model
matrix and original names can't be recovered.

Importance is supported but fails tests as internally data
is coerced to model
matrix and original names can't be recovered.

description
Selected features are extracted with the function
\code{\link[mboost:methods]{mboost::variable.names.mboost()}}, with
\code{used.only = TRUE}.
return \code{character()}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerSurvGLMBoost$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerSurvGLMBoost-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerSurvGLMBoost-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerSurvGLMBoost$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
