% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner_obliqueRSF_surv_obliqueRSF.R
\name{mlr_learners_surv.obliqueRSF}
\alias{mlr_learners_surv.obliqueRSF}
\alias{LearnerSurvObliqueRSF}
\title{Survival Oblique Random Survival Forest Learner}
\description{
Calls \link[obliqueRSF:ORSF]{obliqueRSF::ORSF} from package \CRANpkg{obliqueRSF}.
}
\section{Dictionary}{
 This \link{Learner} can be instantiated via the
\link[mlr3misc:Dictionary]{dictionary} \link{mlr_learners} or with the associated
sugar function \code{\link[=lrn]{lrn()}}:\preformatted{mlr_learners$get("surv.obliqueRSF")
lrn("surv.obliqueRSF")
}
}

\section{Traits}{

\itemize{
\item Packages: obliqueRSF, pracma
\item Predict Types: crank, distr
\item Feature Types: integer, numeric, factor, ordered
\item Properties: missings, oob_error
}
}

\section{Custom mlr3 defaults}{

\itemize{
\item \code{verbose}:
\itemize{
\item Actual default: \code{TRUE}
\item Adjusted default: \code{FALSE}
\item Reason for change: mlr3 already has it's own verbose set to \code{TRUE} by default
}
}
}

\examples{
if (mlr3misc::require_namespaces(c("obliqueRSF", "pracma"))) {
  learner = mlr3::lrn("surv.obliqueRSF")
  print(learner)

  # available parameters:
  learner$param_set$ids()
}
}
\references{
Jaeger BC, Long DL, Long DM, Sims M, Szychowski JM, Min Y, Mcclure LA, Howard G, Simon N (2019).
“Oblique random survival forests.” The Annals of Applied Statistics, 13(3), 1847–1883.
ISSN 1932-6157, 1941-7330, doi: 10.1214/19-AOAS1261,
https://projecteuclid.org/euclid.aoas/1571277776.
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}:
\link[mlr3:mlr_learners]{mlr3::mlr_learners}
}
\author{
RaphaelS1
}
\section{Super classes}{
\code{\link[mlr3:Learner]{mlr3::Learner}} -> \code{\link[mlr3proba:LearnerSurv]{mlr3proba::LearnerSurv}} -> \code{LearnerSurvObliqueRSF}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LearnerSurvObliqueRSF$new()}}
\item \href{#method-oob_error}{\code{LearnerSurvObliqueRSF$oob_error()}}
\item \href{#method-clone}{\code{LearnerSurvObliqueRSF$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="format">}\href{../../mlr3/html/Learner.html#method-format}{\code{mlr3::Learner$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="help">}\href{../../mlr3/html/Learner.html#method-help}{\code{mlr3::Learner$help()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict">}\href{../../mlr3/html/Learner.html#method-predict}{\code{mlr3::Learner$predict()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict_newdata">}\href{../../mlr3/html/Learner.html#method-predict_newdata}{\code{mlr3::Learner$predict_newdata()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="print">}\href{../../mlr3/html/Learner.html#method-print}{\code{mlr3::Learner$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="reset">}\href{../../mlr3/html/Learner.html#method-reset}{\code{mlr3::Learner$reset()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="train">}\href{../../mlr3/html/Learner.html#method-train}{\code{mlr3::Learner$train()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerSurvObliqueRSF$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-oob_error"></a>}}
\if{latex}{\out{\hypertarget{method-oob_error}{}}}
\subsection{Method \code{oob_error()}}{
Integrated brier score OOB error extracted from the model slot \code{oob_error}.
Concordance is also available.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerSurvObliqueRSF$oob_error()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{numeric()}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerSurvObliqueRSF$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
