% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner_obliqueRSF_surv_obliqueRSF.R
\name{mlr_learners_surv.obliqueRSF}
\alias{mlr_learners_surv.obliqueRSF}
\alias{LearnerSurvObliqueRSF}
\title{Survival Oblique Random Survival Forest Learner}
\description{
Oblique random forest.
Calls \code{\link[obliqueRSF:ORSF]{obliqueRSF::ORSF()}} from \CRANpkg{obliqueRSF}.
Note that \code{obliqueRSF} has been superseded by \code{aorsf.}
We highly recommend you use aorsf to fit oblique random survival forests: see https://github.com/bcjaeger/aorsf or install from CRAN with install.packages('aorsf').
}
\section{Dictionary}{

This \link{Learner} can be instantiated via the \link[mlr3misc:Dictionary]{dictionary} \link{mlr_learners} or with the associated sugar function \code{\link[=lrn]{lrn()}}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{mlr_learners$get("surv.obliqueRSF")
lrn("surv.obliqueRSF")
}\if{html}{\out{</div>}}
}

\section{Meta Information}{

\itemize{
\item Task type: \dQuote{surv}
\item Predict Types: \dQuote{crank}, \dQuote{distr}
\item Feature Types: \dQuote{integer}, \dQuote{numeric}, \dQuote{factor}, \dQuote{ordered}
\item Required Packages: \CRANpkg{mlr3}, \CRANpkg{mlr3proba}, \CRANpkg{mlr3extralearners}, \CRANpkg{obliqueRSF}, \CRANpkg{pracma}
}
}

\section{Parameters}{
\tabular{lllll}{
   Id \tab Type \tab Default \tab Levels \tab Range \cr
   alpha \tab numeric \tab 0.5 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   ntree \tab integer \tab 100 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   eval_times \tab untyped \tab - \tab  \tab - \cr
   min_events_to_split_node \tab integer \tab 5 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   min_obs_to_split_node \tab integer \tab 10 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   min_obs_in_leaf_node \tab integer \tab 5 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   min_events_in_leaf_node \tab integer \tab 1 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   nsplit \tab integer \tab 25 \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   gamma \tab numeric \tab 0.5 \tab  \tab \eqn{[1e-16, \infty)}{[1e-16, Inf)} \cr
   max_pval_to_split_node \tab numeric \tab 0.5 \tab  \tab \eqn{[0, 1]}{[0, 1]} \cr
   mtry \tab integer \tab - \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   mtry_ratio \tab numeric \tab - \tab  \tab \eqn{[0, 1]}{[0, 1]} \cr
   dfmax \tab integer \tab - \tab  \tab \eqn{[1, \infty)}{[1, Inf)} \cr
   use.cv \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   verbose \tab logical \tab TRUE \tab TRUE, FALSE \tab - \cr
   compute_oob_predictions \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   random_seed \tab integer \tab - \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
}
}

\section{Initial parameter values}{

\itemize{
\item \code{mtry}:
\itemize{
\item This hyperparameter can alternatively be set via the added hyperparameter \code{mtry_ratio}
as \code{mtry = max(ceiling(mtry_ratio * n_features), 1)}.
Note that \code{mtry} and \code{mtry_ratio} are mutually exclusive.
}
}
}

\section{Custom mlr3 defaults}{

\itemize{
\item \code{verbose}:
\itemize{
\item Actual default: \code{TRUE}
\item Adjusted default: \code{FALSE}
\item Reason for change: mlr3 already has it's own verbose set to \code{TRUE} by default
}
}
}

\examples{
learner = mlr3::lrn("surv.obliqueRSF")
print(learner)

# available parameters:
learner$param_set$ids()
}
\references{
Jaeger BC, Long DL, Long DM, Sims M, Szychowski JM, Min Y, Mcclure LA, Howard G, Simon N (2019).
\dQuote{Oblique random survival forests.}
\emph{The Annals of Applied Statistics}, \bold{13}(3).
\doi{10.1214/19-aoas1261}.
}
\seealso{
\itemize{
\item \link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}.
\item \code{as.data.table(mlr_learners)} for a table of available \link[=Learner]{Learners} in the running session (depending on the loaded packages).
\item Chapter in the \href{https://mlr3book.mlr-org.com/}{mlr3book}: \url{https://mlr3book.mlr-org.com/basics.html#learners}
\item \CRANpkg{mlr3learners} for a selection of recommended learners.
\item \CRANpkg{mlr3cluster} for unsupervised clustering learners.
\item \CRANpkg{mlr3pipelines} to combine learners with pre- and postprocessing steps.
\item \CRANpkg{mlr3tuning} for tuning of hyperparameters, \CRANpkg{mlr3tuningspaces} for established default tuning spaces.
}
}
\author{
adibender
}
\section{Super classes}{
\code{\link[mlr3:Learner]{mlr3::Learner}} -> \code{\link[mlr3proba:LearnerSurv]{mlr3proba::LearnerSurv}} -> \code{LearnerSurvObliqueRSF}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LearnerSurvObliqueRSF-new}{\code{LearnerSurvObliqueRSF$new()}}
\item \href{#method-LearnerSurvObliqueRSF-oob_error}{\code{LearnerSurvObliqueRSF$oob_error()}}
\item \href{#method-LearnerSurvObliqueRSF-clone}{\code{LearnerSurvObliqueRSF$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="base_learner"><a href='../../mlr3/html/Learner.html#method-Learner-base_learner'><code>mlr3::Learner$base_learner()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="format"><a href='../../mlr3/html/Learner.html#method-Learner-format'><code>mlr3::Learner$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="help"><a href='../../mlr3/html/Learner.html#method-Learner-help'><code>mlr3::Learner$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict"><a href='../../mlr3/html/Learner.html#method-Learner-predict'><code>mlr3::Learner$predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict_newdata"><a href='../../mlr3/html/Learner.html#method-Learner-predict_newdata'><code>mlr3::Learner$predict_newdata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="print"><a href='../../mlr3/html/Learner.html#method-Learner-print'><code>mlr3::Learner$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="reset"><a href='../../mlr3/html/Learner.html#method-Learner-reset'><code>mlr3::Learner$reset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="train"><a href='../../mlr3/html/Learner.html#method-Learner-train'><code>mlr3::Learner$train()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerSurvObliqueRSF-new"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerSurvObliqueRSF-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerSurvObliqueRSF$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerSurvObliqueRSF-oob_error"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerSurvObliqueRSF-oob_error}{}}}
\subsection{Method \code{oob_error()}}{
Integrated brier score OOB error extracted from the model slot \code{oob_error}.
Concordance is also available.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerSurvObliqueRSF$oob_error()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{numeric()}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerSurvObliqueRSF-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerSurvObliqueRSF-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerSurvObliqueRSF$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
