% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner_survival_surv_parametric.R
\name{mlr_learners_surv.parametric}
\alias{mlr_learners_surv.parametric}
\alias{LearnerSurvParametric}
\title{Survival Fully Parametric Learner}
\description{
Parametric survival model.
Calls \code{\link[survival:survreg]{survival::survreg()}} from \CRANpkg{survival}.
}
\details{
This learner allows you to choose a distribution and a model form to compose a predicted
survival probability distribution.

The internal predict method is implemented in this package as our implementation is more
efficient for composition to distributions than \code{\link[survival:predict.survreg]{survival::predict.survreg()}}.

\code{lp} is predicted using the formula \eqn{lp = X\beta} where \eqn{X} are the variables in the test
data set and \eqn{\beta} are the fitted coefficients.

The distribution \code{distr} is composed using the \code{lp} and specifying a model form in the
\code{type} hyper-parameter. These are as follows, with respective survival functions,
\itemize{
\item Accelerated Failure Time (\code{aft}) \deqn{S(t) = S_0(\frac{t}{exp(lp)})}{S(t) = S0(t/exp(lp))}
\item Proportional Hazards (\code{ph}) \deqn{S(t) = S_0(t)^{exp(lp)}}{S(t) = S0(t)^exp(lp)}
\item Proportional Odds (\code{po}) \deqn{S(t) =
\frac{S_0(t)}{exp(-lp) + (1-exp(-lp)) S_0(t)}}{S(t) = S0(t) / [exp(-lp) + S0(t) (1-exp(-lp))]}
\item Tobit (\code{tobit}) \deqn{S(t) = 1 - F((t - lp)/s)}
}

where \eqn{S_0}{S0} is the estimated baseline survival distribution (in this case
with a given parametric form), \eqn{lp} is the predicted linear predictor, \eqn{F} is the cdf
of a N(0, 1) distribution, and \eqn{s} is the fitted scale parameter.

Whilst any combination of distribution and model form is possible, this does not mean it will
necessarily create a sensible or interpretable prediction. The following combinations are
'sensible':
\itemize{
\item dist = "gaussian"; type = "tobit";
\item dist = "weibull"; type = "ph";
\item dist = "exponential"; type = "ph";
\item dist = "weibull"; type = "aft";
\item dist = "exponential"; type = "aft";
\item dist = "loglogistic"; type = "aft";
\item dist = "lognormal"; type = "aft";
\item dist = "loglogistic"; type = "po";
}
}
\section{Dictionary}{

This \link{Learner} can be instantiated via the \link[mlr3misc:Dictionary]{dictionary} \link{mlr_learners} or with the associated sugar function \code{\link[=lrn]{lrn()}}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{mlr_learners$get("surv.parametric")
lrn("surv.parametric")
}\if{html}{\out{</div>}}
}

\section{Meta Information}{

\itemize{
\item Task type: \dQuote{surv}
\item Predict Types: \dQuote{crank}, \dQuote{distr}, \dQuote{lp}
\item Feature Types: \dQuote{logical}, \dQuote{integer}, \dQuote{numeric}, \dQuote{factor}
\item Required Packages: \CRANpkg{mlr3}, \CRANpkg{mlr3proba}, \CRANpkg{mlr3extralearners}, \CRANpkg{survival}, \CRANpkg{pracma}
}
}

\section{Parameters}{
\tabular{lllll}{
   Id \tab Type \tab Default \tab Levels \tab Range \cr
   type \tab character \tab aft \tab aft, ph, po, tobit \tab - \cr
   na.action \tab untyped \tab - \tab  \tab - \cr
   dist \tab character \tab weibull \tab weibull, exponential, gaussian, lognormal, loglogistic \tab - \cr
   parms \tab untyped \tab - \tab  \tab - \cr
   init \tab untyped \tab - \tab  \tab - \cr
   scale \tab numeric \tab 0 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   maxiter \tab integer \tab 30 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   rel.tolerance \tab numeric \tab 1e-09 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   toler.chol \tab numeric \tab 1e-10 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   debug \tab integer \tab 0 \tab  \tab \eqn{[0, 1]}{[0, 1]} \cr
   outer.max \tab integer \tab 10 \tab  \tab \eqn{(-\infty, \infty)}{(-Inf, Inf)} \cr
   robust \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   score \tab logical \tab FALSE \tab TRUE, FALSE \tab - \cr
   cluster \tab untyped \tab - \tab  \tab - \cr
}
}

\examples{
learner = mlr3::lrn("surv.parametric")
print(learner)

# available parameters:
learner$param_set$ids()
}
\references{
Kalbfleisch, D J, Prentice, L R (2011).
\emph{The statistical analysis of failure time data}.
John Wiley & Sons.
}
\seealso{
\itemize{
\item \link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}.
\item \code{as.data.table(mlr_learners)} for a table of available \link[=Learner]{Learners} in the running session (depending on the loaded packages).
\item Chapter in the \href{https://mlr3book.mlr-org.com/}{mlr3book}: \url{https://mlr3book.mlr-org.com/basics.html#learners}
\item \CRANpkg{mlr3learners} for a selection of recommended learners.
\item \CRANpkg{mlr3cluster} for unsupervised clustering learners.
\item \CRANpkg{mlr3pipelines} to combine learners with pre- and postprocessing steps.
\item \CRANpkg{mlr3tuning} for tuning of hyperparameters, \CRANpkg{mlr3tuningspaces} for established default tuning spaces.
}
}
\author{
RaphaelS1
}
\section{Super classes}{
\code{\link[mlr3:Learner]{mlr3::Learner}} -> \code{\link[mlr3proba:LearnerSurv]{mlr3proba::LearnerSurv}} -> \code{LearnerSurvParametric}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LearnerSurvParametric-new}{\code{LearnerSurvParametric$new()}}
\item \href{#method-LearnerSurvParametric-clone}{\code{LearnerSurvParametric$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="base_learner"><a href='../../mlr3/html/Learner.html#method-Learner-base_learner'><code>mlr3::Learner$base_learner()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="format"><a href='../../mlr3/html/Learner.html#method-Learner-format'><code>mlr3::Learner$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="help"><a href='../../mlr3/html/Learner.html#method-Learner-help'><code>mlr3::Learner$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict"><a href='../../mlr3/html/Learner.html#method-Learner-predict'><code>mlr3::Learner$predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict_newdata"><a href='../../mlr3/html/Learner.html#method-Learner-predict_newdata'><code>mlr3::Learner$predict_newdata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="print"><a href='../../mlr3/html/Learner.html#method-Learner-print'><code>mlr3::Learner$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="reset"><a href='../../mlr3/html/Learner.html#method-Learner-reset'><code>mlr3::Learner$reset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="train"><a href='../../mlr3/html/Learner.html#method-Learner-train'><code>mlr3::Learner$train()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerSurvParametric-new"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerSurvParametric-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerSurvParametric$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerSurvParametric-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerSurvParametric-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerSurvParametric$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
