% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner_survivalsvm_surv_svm.R
\name{mlr_learners_surv.svm}
\alias{mlr_learners_surv.svm}
\alias{LearnerSurvSVM}
\title{Survival Support Vector Machine Learner}
\description{
Survival support vector machine.
Calls \code{\link[survivalsvm:survivalsvm]{survivalsvm::survivalsvm()}} from \CRANpkg{survivalsvm}.
}
\details{
Four possible SVMs can be implemented, dependent on the \code{type} parameter. These correspond
to predicting the survival time via regression (\code{regression}), predicting a continuous rank
(\code{vanbelle1}, \code{vanbelle2}), or a hybrid of the two (\code{hybrid}).
Whichever \code{type} is chosen determines how the \code{crank} predict type is calculated,
but in any case all can be considered a valid continuous ranking.

\code{makediff3} is recommended when using \code{type = "hybrid"}.
}
\section{Dictionary}{

This \link{Learner} can be instantiated via the \link[mlr3misc:Dictionary]{dictionary} \link{mlr_learners} or with the associated sugar function \code{\link[=lrn]{lrn()}}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{mlr_learners$get("surv.svm")
lrn("surv.svm")
}\if{html}{\out{</div>}}
}

\section{Meta Information}{

\itemize{
\item Task type: \dQuote{surv}
\item Predict Types: \dQuote{crank}, \dQuote{response}
\item Feature Types: \dQuote{logical}, \dQuote{integer}, \dQuote{numeric}, \dQuote{character}, \dQuote{factor}
\item Required Packages: \CRANpkg{mlr3}, \CRANpkg{mlr3proba}, \CRANpkg{mlr3extralearners}, \CRANpkg{survivalsvm}
}
}

\section{Parameters}{
\tabular{lllll}{
   Id \tab Type \tab Default \tab Levels \tab Range \cr
   type \tab character \tab regression \tab regression, vanbelle1, vanbelle2, hybrid \tab - \cr
   diff.meth \tab character \tab - \tab makediff1, makediff2, makediff3 \tab - \cr
   gamma.mu \tab untyped \tab - \tab  \tab - \cr
   opt.meth \tab character \tab quadprog \tab quadprog, ipop \tab - \cr
   kernel \tab character \tab lin_kernel \tab lin_kernel, add_kernel, rbf_kernel, poly_kernel \tab - \cr
   kernel.pars \tab untyped \tab - \tab  \tab - \cr
   sgf.sv \tab integer \tab 5 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   sigf \tab integer \tab 7 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   maxiter \tab integer \tab 20 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   margin \tab numeric \tab 0.05 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   bound \tab numeric \tab 10 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   eig.tol \tab numeric \tab 1e-06 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   conv.tol \tab numeric \tab 1e-07 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   posd.tol \tab numeric \tab 1e-08 \tab  \tab \eqn{[0, \infty)}{[0, Inf)} \cr
}
}

\examples{
learner = mlr3::lrn("surv.svm")
print(learner)

# available parameters:
learner$param_set$ids()
}
\references{
Van Belle, Vanya, Pelckmans, Kristiaan, Van Huffel, Sabine, Suykens, AK J (2011).
\dQuote{Improved performance on high-dimensional survival data by application of Survival-SVM.}
\emph{Bioinformatics}, \bold{27}(1), 87--94.

Van Belle, Vanya, Pelckmans, Kristiaan, Van Huffel, Sabine, Suykens, AK J (2011).
\dQuote{Support vector methods for survival analysis: a comparison between ranking and regression approaches.}
\emph{Artificial intelligence in medicine}, \bold{53}(2), 107--118.

Shivaswamy, K P, Chu, Wei, Jansche, Martin (2007).
\dQuote{A support vector approach to censored targets.}
In \emph{Seventh IEEE international conference on data mining (ICDM 2007)}, 655--660.
IEEE.
}
\seealso{
\itemize{
\item \link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}.
\item \code{as.data.table(mlr_learners)} for a table of available \link[=Learner]{Learners} in the running session (depending on the loaded packages).
\item Chapter in the \href{https://mlr3book.mlr-org.com/}{mlr3book}: \url{https://mlr3book.mlr-org.com/basics.html#learners}
\item \CRANpkg{mlr3learners} for a selection of recommended learners.
\item \CRANpkg{mlr3cluster} for unsupervised clustering learners.
\item \CRANpkg{mlr3pipelines} to combine learners with pre- and postprocessing steps.
\item \CRANpkg{mlr3tuning} for tuning of hyperparameters, \CRANpkg{mlr3tuningspaces} for established default tuning spaces.
}
}
\author{
RaphaelS1
}
\section{Super classes}{
\code{\link[mlr3:Learner]{mlr3::Learner}} -> \code{\link[mlr3proba:LearnerSurv]{mlr3proba::LearnerSurv}} -> \code{LearnerSurvSVM}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LearnerSurvSVM-new}{\code{LearnerSurvSVM$new()}}
\item \href{#method-LearnerSurvSVM-clone}{\code{LearnerSurvSVM$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="base_learner"><a href='../../mlr3/html/Learner.html#method-Learner-base_learner'><code>mlr3::Learner$base_learner()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="format"><a href='../../mlr3/html/Learner.html#method-Learner-format'><code>mlr3::Learner$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="help"><a href='../../mlr3/html/Learner.html#method-Learner-help'><code>mlr3::Learner$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict"><a href='../../mlr3/html/Learner.html#method-Learner-predict'><code>mlr3::Learner$predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict_newdata"><a href='../../mlr3/html/Learner.html#method-Learner-predict_newdata'><code>mlr3::Learner$predict_newdata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="print"><a href='../../mlr3/html/Learner.html#method-Learner-print'><code>mlr3::Learner$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="reset"><a href='../../mlr3/html/Learner.html#method-Learner-reset'><code>mlr3::Learner$reset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="train"><a href='../../mlr3/html/Learner.html#method-Learner-train'><code>mlr3::Learner$train()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerSurvSVM-new"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerSurvSVM-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerSurvSVM$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerSurvSVM-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerSurvSVM-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerSurvSVM$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
